### 2025-07-11 #8574 @editorialbot
[PRE REVIEW]: zea: A Toolbox for Cognitive Ultrasound Imaging
**Submitting author:** @tristan-deep (<a href="http://orcid.org/0000-0002-8563-5931">Tristan Stevens</a>)
**Repository:** https://github.com/tue-bmd/zea
**Branch with paper.md** (empty if default branch): feature/joss-paper
**Version:** v0.0.9
