### 2025-07-12 f363f35
bib updates
