+++ paper/paper.bib
@@ -1,136 +1,267 @@
 @misc{stevens2025zea,
-    author = {Stevens, Tristan S.W. and van Nierop, Wessel L. and Luijten, Ben and van de Schaft, Vincent and Nolan, Oisín and Federici, Beatrice and Penninga, Simon W. and Schueler, Noortje I.P. and van Sloun, Ruud J.G.},
-    title = {zea: A Toolbox for Cognitive Ultrasound Imaging},
-    url = {https://github.com/tue-bmd/zea},
-    year = {2025},
+  author       = {Stevens, Tristan S.W. and van Nierop, Wessel L. and Luijten, Ben and van de Schaft, Vincent and Nolan, Oisín and Federici, Beatrice and Penninga, Simon W. and Schueler, Noortje I.P. and van Sloun, Ruud J.G.},
+  title        = {zea: {A} {T}oolbox for {C}ognitive {U}ltrasound {I}maging},
+  url          = {https://github.com/tue-bmd/zea},
+  year         = {2025}
 }
 
 @article{van2024active,
-  author={van Sloun, Ruud J. G.},
-  journal={IEEE Transactions on Ultrasonics, Ferroelectrics, and Frequency Control},
-  title={Active Inference and Deep Generative Modeling for Cognitive Ultrasound},
-  year={2024},
-  volume={71},
-  number={11},
-  pages={1478-1490},
-  doi={10.1109/TUFFC.2024.3466290}
+  author       = {Ruud J. G. van Sloun},
+  bibsource    = {dblp computer science bibliography, https://dblp.org},
+  biburl       = {https://dblp.org/rec/journals/corr/abs-2410-13310.bib},
+  doi          = {10.48550/ARXIV.2410.13310},
+  eprint       = {2410.13310},
+  eprinttype   = {arXiv},
+  journal      = {CoRR},
+  timestamp    = {Fri, 29 Nov 2024 00:00:00 +0100},
+  title        = {{A}ctive inference and deep generative modeling for cognitive ultrasound},
+  url          = {https://doi.org/10.48550/arXiv.2410.13310},
+  volume       = {abs/2410.13310},
+  year         = {2024}
 }
 
-@misc{van2024off,
-  title={Off-Grid Ultrasound Imaging by Stochastic Optimization},
-  author={Vincent van de Schaft and Oisín Nolan and Ruud J. G. van Sloun},
-  year={2024},
-  eprint={2407.02285},
-  archivePrefix={arXiv},
-  primaryClass={eess.SP},
-  url={https://arxiv.org/abs/2407.02285},
+@article{van2024off,
+  author       = {Van de Schaft, Vincent and Nolan, Oisín and Van Sloun, Ruud JG},
+  doi          = {10.1109/TUFFC.2025.3586377},
+  journal      = {IEEE Transactions on Ultrasonics, Ferroelectrics, and Frequency Control},
+  keywords     = {Imaging;Ultrasonic imaging;Radio frequency;Array signal processing;Inverse problems;Stochastic processes;Reflectivity;IP networks;Data models;Computational modeling;ultrasound imaging;inverse problems;optimization;beamforming},
+  number       = {},
+  pages        = {1-1},
+  title        = {{O}ff-Grid {U}ltrasound {I}maging by {S}tochastic {O}ptimization},
+  volume       = {},
+  year         = {2025}
 }
 
 @article{stevens2024dehazing,
-  author={Stevens, Tristan S. W. and Meral, Faik C. and Yu, Jason and Apostolakis, Iason Z. and Robert, Jean-Luc and van Sloun, Ruud J. G.},
-  journal={IEEE Transactions on Medical Imaging},
[…]
