### 2025-07-12 8d8742b
fix affiliations
