### 2025-07-12 #8574 @tristan-deep
The suggested Keras DOI seems incorrect (see [here](https://keras.io/getting_started/faq/#how-should-i-cite-keras)). Also the invalid DOI is existing, see [here](https://royalsocietypublishing.org/doi/full/10.1098/rsta.2024.0327). It is quite a recently published paper, so that might be causing it not to find it.
