### 2025-07-31 c1d81ca
Refactor LogCompress class and add new operations: Translate, To8Bit, and Constant
