### 2025-08-04 331d0ea
allow for single valued dynamic range parameter
