### 2025-08-07 252b2d4
Allow more scan class types (to be discussed)
