### 2025-08-07 69a71ac
📝 Add docstrings to `feature/diff-backprop`
Docstrings generation was requested by @tristan-deep.

* https://github.com/tue-bmd/zea/pull/75#issuecomment-3164933724

The following files were modified:

* `zea/models/diffusion.py`
