+++ zea/models/diffusion.py
@@ -118,6 +118,11 @@ class DiffusionModel(DeepGenerativeModel):
         return config
 
     def _init_operator_and_guidance(self, operator, guidance):
+        """
+        Initializes the operator and guidance function for the diffusion model based on provided specifications.
+        
+        Accepts operator and guidance as string keys, configuration dictionaries, or instantiated objects. Instantiates and assigns the appropriate operator and guidance function, ensuring required parameters (such as `image_range`) are set when necessary. Raises a ValueError if invalid types are provided or if guidance is specified without an operator.
+        """
         if operator is not None:
             if isinstance(operator, str):
                 operator_class = operator_registry[operator]
@@ -162,10 +167,18 @@ class DiffusionModel(DeepGenerativeModel):
                 )
 
     def call(self, inputs, training=False, network=None, **kwargs):
-        """Calls the score network.
-
-        If network is not provided, will use the exponential moving
-        average network if training is False, otherwise the regular network.
+        """
+        Runs the score network on the given inputs using the specified or default network.
+        
+        If `network` is not provided, uses the main network during training and the exponential moving average (EMA) network during inference.
+        
+        Parameters:
+            inputs: Input tensor to the score network.
+            training (bool): Whether to run in training mode.
+            network: Optional network instance to use instead of the default.
+        
+        Returns:
+            The output of the selected score network applied to the inputs.
         """
         if network is None:
             network = self.network if training else self.ema_network
@@ -173,16 +186,16 @@ class DiffusionModel(DeepGenerativeModel):
         return network(inputs, training=training, **kwargs)
 
     def sample(self, n_samples=1, n_steps=20, seed=None, **kwargs):
-        """Sample from the model.
-
-        Args:
-            n_samples: Number of samples to generate.
-            n_steps: Number of diffusion steps.
-            seed: Random seed generator.
-            **kwargs: Additional arguments.
-
+        """
+        Generates samples from the diffusion model by running the reverse diffusion process starting from Gaussian noise.
+        
+        Parameters:
+        	n_samples (int): Number of samples to generate.
+        	n_steps (int): Number of diffusion steps to perform.
+        	seed: Optional random seed generator for reproducibility.
+        
         Returns:
-            Generated samples.
+        	Samples generated by the reverse diffusion process.
         """
         seed, seed1 = split_seed(seed, 2)
 
@@ -365,10 +378,27 @@ class DiffusionModel(DeepGenerativeModel):
         return noise_rates, signal_rates
 
     def linear_diffusion_schedule(self, diffusion_times):
-        """Create a linear diffusion schedule"""
+        """
+        Compute signal and noise rates for a linear diffusion schedule based on the provided diffusion times.
+        
+        Parameters:
+            diffusion_times: A tensor representing the diffusion time steps.
+        
+        Returns:
+            signal_rates: Tensor of signal rates for each diffusion step.
+            noise_rates: Tensor of noise rates for each diffusion step.
+        """
 
         def _compute_alpha_t(t):
-            """Compute alpha_t for linear diffusion schedule"""
+            """
+            Computes the cumulative product of (1 - diffusion rates) up to time step t for a linear diffusion schedule.
+            
+            Parameters:
+                t: The current diffusion time step.
+            
+            Returns:
+                The cumulative product of (1 - diffusion rates) up to step t.
+            """
             return ops.prod(1 - diffusion_times[:t], axis=diffusion_times.shape[1:])
 
         alphas = ops.vectorized_map(_compute_alpha_t, ops.arange(len(diffusion_times)))
@@ -384,7 +414,20 @@ class DiffusionModel(DeepGenerativeModel):
         training,
         network=None,
     ):
-        """Predict noise component and calculate the image component using it."""
+        """
[…]
