+++ zea/scan.py
@@ -83,15 +83,12 @@ from keras import ops
 
 from zea import log
 from zea.beamform.pfield import compute_pfield
-from zea.beamform.pixelgrid import (
-    cartesian_pixel_grid,
-    check_for_aliasing,
-    polar_pixel_grid,
-)
+from zea.beamform.pixelgrid import cartesian_pixel_grid, check_for_aliasing, polar_pixel_grid
 from zea.display import (
     compute_scan_convert_2d_coordinates,
     compute_scan_convert_3d_coordinates,
 )
+from zea.internal.core import DEFAULT_DYNAMIC_RANGE
 from zea.internal.parameters import Parameters, cache_with_dependencies
 
 
@@ -164,66 +161,21 @@ class Scan(Parameters):
             Defined in dB as (min_dB, max_dB). Defaults to (-60, 0).
     """
 
-    # VALID_PARAMS = {
-    #     # beamforming related parameters
-    #     "grid_size_x": {"type": int},
-    #     "grid_size_z": {"type": int},
-    #     "xlims": {"type": (tuple, list)},
-    #     "ylims": {"type": (tuple, list)},
-    #     "zlims": {"type": (tuple, list)},
-    #     "pixels_per_wavelength": {"type": int, "default": 4},
-    #     "pfield_kwargs": {"type": dict, "default": {}},
-    #     "apply_lens_correction": {"type": bool, "default": False},
-    #     "lens_sound_speed": {"type": (float, int)},
-    #     "lens_thickness": {"type": float},
-    #     "grid_type": {"type": str, "default": "cartesian"},
-    #     "polar_limits": {"type": (tuple, list)},
-    #     "dynamic_range": {"type": (tuple, list), "default": DEFAULT_DYNAMIC_RANGE},
-    #     # acquisition parameters
-    #     "sound_speed": {"type": (float, int), "default": 1540.0},
-    #     "sampling_frequency": {"type": float},
-    #     "center_frequency": {"type": float},
-    #     "n_el": {"type": int},
-    #     "n_tx": {"type": int},
-    #     "n_ax": {"type": int},
-    #     "n_ch": {"type": int},
-    #     "bandwidth_percent": {"type": float, "default": 200.0},
-    #     "demodulation_frequency": {"type": float},
-    #     "element_width": {"type": float},
-    #     "attenuation_coef": {"type": float, "default": 0.0},
-    #     "f_number": {"type": float, "default": 1.0},
-    #     # array parameters
-    #     "probe_geometry": {"type": np.ndarray},
-    #     "polar_angles": {"type": np.ndarray},
-    #     "azimuth_angles": {"type": np.ndarray},
-    #     "t0_delays": {"type": np.ndarray},
-    #     "tx_apodizations": {"type": np.ndarray},
-    #     "focus_distances": {"type": np.ndarray},
-    #     "initial_times": {"type": np.ndarray},
-    #     "time_to_next_transmit": {"type": np.ndarray},
-    #     # scan conversion parameters
-    #     "theta_range": {"type": (tuple, list)},
-    #     "phi_range": {"type": (tuple, list)},
-    #     "rho_range": {"type": (tuple, list)},
-    #     "fill_value": {"type": float, "default": 0.0},
-    #     "resolution": {"type": float, "default": None},
-    # }
-
     VALID_PARAMS = {
         # beamforming related parameters
-        "Nx": {"type": (int, np.integer, np.ndarray), "default": None},
-        "Nz": {"type": (int, np.integer, np.ndarray), "default": None},
-        "xlims": {"type": ((tuple, list, np.ndarray)), "default": None},
-        "ylims": {"type": ((tuple, list, np.ndarray)), "default": None},
-        "zlims": {"type": ((tuple, list, np.ndarray)), "default": None},
[…]
