### 2025-09-19 8214c9f
Add segmentation model
