### 2025-09-19 4bc18b2
Add image quality model
