### 2025-10-08 6e63a1c
Fix matlab.py read_raw_data() to handle single frame case
Co-authored-by: tristan-deep <58224357+tristan-deep@users.noreply.github.com>
