### 2025-10-08 cddb1a6
Add tests for MATLAB read_raw_data dimension handling fix
Co-authored-by: tristan-deep <58224357+tristan-deep@users.noreply.github.com>
