+++ zea/models/able.py
@@ -1,0 +1,195 @@
+"""Adaptive Beamforming by Deep LEarning (ABLE).
+
+Original implementation of paper:
+    - "Adaptive Ultrasound Beamforming Using Deep Learning"
+    - https://doi.org/10.1109/TMI.2020.3008537
+    - Author: Ben Luijten
+"""
+
+import keras
+from keras import ops
+from keras.layers import (
+    BatchNormalization,
+    Conv2D,
+    Dropout,
+    Input,
+    MaxPooling2D,
+    UpSampling2D,
+    concatenate,
+)
+
+from zea.internal.registry import model_registry
+from zea.models.base import BaseModel
+
+
+@model_registry(name="able")
+class ABLE(BaseModel):
+    """Adaptive Beamforming by Deep LEarning (ABLE) model.
+
+    This class implements a small configurable convolutional encoder/decoder
+    architecture used for adaptive ultrasound beamforming. The constructor
+    accepts either an explicit list of layer channel sizes via ``layers`` or
+    constructs a symmetric list using ``num_layers`` and ``latent_dim``. The
+    ``kernel_size`` argument supports multiple forms (int, tuple, or list) and
+    is parsed into a list of 2-tuples (height, width) matching the number
+    of layers. The default configuration matches the original paper.
+
+    Parameters
+    ----------
+    elements : int
+        Number of fast-time channels / transducer elements (used as first and
+        last layer channel size when ``layers`` is not provided).
+    latent_dim : int
+        Channel size for middle (latent) layers when ``layers`` is not provided.
+    kernel_size : int, tuple, or list
+        Kernel size specification. Accepted forms:
+          - int -> every conv uses (k, k)
+          - tuple (h, w) -> every conv uses (h, w)
+          - list of ints/tuples -> layer-wise kernels (length must match layers)
+    num_layers : int
+        Number of layers to construct when ``layers`` is not provided. Must be
+        >= 2. The constructed list will be: [elements, latent_dim, ..., elements].
+    layers : list or None
+        Optional explicit list of channel sizes per layer. When provided this
+        overrides ``num_layers`` and ``latent_dim``.
+    name : str
+        Model name forwarded to :class:`BaseModel`.
+
+
+    Example
+    -------
+    .. code-block:: python
+
+        # create a 4-layer ABLE model with default (1,1) kernels.
+        model = ABLE(elements=128, latent_dim=32, num_layers=4, kernel_size=(1,1))
+    """
+
+    def __init__(
+        self,
+        elements=128,  # Number of fast-time channels/transducer elements
+        latent_dim=32,  # Latent dimension size (used for middle layers)
+        kernel_size=1,  # See normalization rules below
+        num_layers=4,  # number of layers (default 4 -> [elements, latent_dim..., elements])
+        layers=None,  # explicit list of layer channel sizes, overrides num_layers
+        name="able",
+        **kwargs,
+    ):
+        """Initializes the ABLE model.
+
+        Kernel size handling (accepted forms):
+        - single int (e.g. 1) -> every conv uses (1,1)
+        - single tuple (h, w) (e.g. (1,3)) -> every conv uses (h,w)
+        - list of ints (e.g. [1,3,3,1]) -> converted to [(1,1),(3,3),...]
+        - list of tuples (e.g. [(1,1),(1,3),(1,3),(1,1)]) -> used as-is
+
+        Layer / dimensionality handling:
+        - If `layers` is provided and is a list, that list is used as the channel
+          sizes for each layer.
+        - Else, `num_layers` is used to create a list of length `num_layers`:
+            [elements, latent_dim, latent_dim, ..., elements]
+          (first and last are `elements`, middle layers are `latent_dim`).
+        """
+        super().__init__(name=name, **kwargs)
+
[…]
