### 2025-10-21 f5036b5
Implement get_network function
