### 2025-10-22 6b6c4ad
add bib
