+++ paper/paper.bib
@@ -413,3 +413,10 @@
   journal={arXiv preprint arXiv:2508.17326},
   year={2025}
 }
+
+@article{stevens2025nuclear,
+  title={Nuclear Diffusion Models for Low-Rank Background Suppression in Videos},
+  author={Stevens, Tristan SW and Nolan, Ois{\'\i}n and Robert, Jean-Luc and van Sloun, Ruud JG},
+  journal={arXiv preprint arXiv:2509.20886},
+  year={2025}
+}
