+++ zea/models/able.py
@@ -6,17 +6,8 @@ Original implementation of paper:
     - Author: Ben Luijten
 """
 
-import keras
 from keras import ops
-from keras.layers import (
-    BatchNormalization,
-    Conv2D,
-    Dropout,
-    Input,
-    MaxPooling2D,
-    UpSampling2D,
-    concatenate,
-)
+from keras.layers import Conv2D
 
 from zea.internal.registry import model_registry
 from zea.models.base import BaseModel
@@ -26,51 +17,44 @@ from zea.models.base import BaseModel
 class ABLE(BaseModel):
     """Adaptive Beamforming by Deep LEarning (ABLE) model.
 
-    This class implements a small configurable convolutional encoder/decoder
-    architecture used for adaptive ultrasound beamforming. The constructor
-    accepts either an explicit list of layer channel sizes via ``layers`` or
-    constructs a symmetric list using ``num_layers`` and ``latent_dim``. The
-    ``kernel_size`` argument supports multiple forms (int, tuple, or list) and
-    is parsed into a list of 2-tuples (height, width) matching the number
-    of layers. The default configuration matches the original paper.
+    This class implements a configurable convolutional encoder/decoder
+    architecture for adaptive ultrasound beamforming. The constructor
+    allows for flexible configuration of layer dimensions and kernel sizes.
+    The default configuration corresponds to the setup used in the original paper
+    "Adaptive Ultrasound Beamforming Using Deep Learning"
+    (DOI: https://doi.org/10.1109/TMI.2020.3008537).
 
-    Parameters
-    ----------
-    elements : int
-        Number of fast-time channels / transducer elements (used as first and
-        last layer channel size when ``layers`` is not provided).
-    latent_dim : int
-        Channel size for middle (latent) layers when ``layers`` is not provided.
-    kernel_size : int, tuple, or list
+    Args:
+    latent_dim : int, optional
+        Channel size for the middle (latent) layers when `latent_layers` is not provided.
+        Default is 32.
+    kernel_size : int, tuple, or list, optional
         Kernel size specification. Accepted forms:
-          - int -> every conv uses (k, k)
-          - tuple (h, w) -> every conv uses (h, w)
+          - int -> every convolution uses (k, k)
+          - tuple (h, w) -> every convolution uses (h, w)
           - list of ints/tuples -> layer-wise kernels (length must match layers)
-    num_layers : int
-        Number of layers to construct when ``layers`` is not provided. Must be
-        >= 2. The constructed list will be: [elements, latent_dim, ..., elements].
-    layers : list or None
-        Optional explicit list of channel sizes per layer. When provided this
-        overrides ``num_layers`` and ``latent_dim``.
-    name : str
-        Model name forwarded to :class:`BaseModel`.
-
-
-    Example
-    -------
-    .. code-block:: python
-
-        # create a 4-layer ABLE model with default (1,1) kernels.
-        model = ABLE(elements=128, latent_dim=32, num_layers=4, kernel_size=(1, 1))
+        Default is 1.
+    n_latent_layers : int, optional
+        Number of inner layers to construct when `latent_layers` is not provided.
+        Must be >= 2. Default is 2.
+    latent_layers : list or None, optional
+        Explicit list of channel sizes for inner layers. Overrides `n_latent_layers`
+        and `latent_dim` when provided. Default is None.
+    axis : int, optional
+        Axis that contains the transducer elements. Default is 3.
+    name : str, optional
+        Model name forwarded to `BaseModel`. Default is "able".
+    **kwargs : dict
+        Additional keyword arguments forwarded to `BaseModel`.
     """
 
     def __init__(
         self,
-        elements=128,  # Number of fast-time channels/transducer elements
-        latent_dim=32,  # Latent dimension size (used for middle layers)
[…]
