### 2025-11-04 bed3f0a
Add reshapetogrid operation to unflatten data
