### 2025-11-04 dafcee1
working version of able, pre-cleanup
