+++ zea/models/able.py
@@ -54,7 +54,7 @@ class ABLE(BaseModel):
         kernel_size=1,
         n_latent_layers=2,
         latent_layers=None,
-        axis=3,
+        axis=None,
         name="able",
         **kwargs,
     ):
@@ -186,9 +186,13 @@ class ABLE(BaseModel):
 
             # Infer input shape and dynamically initialize layers on first pass
             if not self._able_layers:
-                # Stack final channel dim into element axis when needed
-                x_reshaped, meta = self.stack_channels(x, axis)
-                stacked_input_dim = ops.shape(x_reshaped)[axis]
+                # Stack final channel dim into element axis when needed.
+                # We assume supported input formats:
+                #   - (batch, H, W, elements)
+                #   - (batch, H, W, elements, n_ch)
+                # After stack_channels the last axis is the channel axis.
+                x_reshaped, meta = self.stack_channels(x, None)
+                stacked_input_dim = ops.shape(x_reshaped)[-1]
 
                 # Dynamically initialize layer dimensions and kernel sizes
                 self.layer_dims = self._init_layers(
@@ -204,7 +208,7 @@ class ABLE(BaseModel):
                     )
 
             # Use the reshaped input for the forward pass
-            x_reshaped, meta = self.stack_channels(x, axis)
+            x_reshaped, meta = self.stack_channels(x, None)
 
             out = x_reshaped
             for conv in self._able_layers:
@@ -224,76 +228,68 @@ class ABLE(BaseModel):
     def stack_channels(self, x, axis):
         """Stack the final channel dimension into the element axis.
 
-        Args:
-            x (Tensor): Input tensor.
-            axis (int): Axis index corresponding to the transducer elements.
-
-        Returns:
-            tuple: Transformed tensor and metadata for reversing the operation.
+        SIMPLIFIED:
+        - Assumes element axis is the second-last axis and optional per-element channel
+          axis is the last axis.
+        - If input rank == 4: (batch, H, W, elements) -> no change.
+        - If input rank == 5: (batch, H, W, elements, n_ch) -> merge elements and n_ch
+          into a single channel axis (last axis) only when n_ch > 1.
+        - Caller should ensure input layout meets the assumption or transpose beforehand.
         """
         rank = len(x.shape)
-        elem_axis = axis if axis >= 0 else rank + axis
         meta = {"stacked": False}
 
-        if rank >= 5:
-            last_idx = rank - 1
-            perm = list(range(rank))
-            if last_idx != elem_axis + 1:
-                perm.pop(last_idx)
-                perm.insert(elem_axis + 1, last_idx)
-                x_perm = ops.permute_dimensions(x, perm)
-                meta["perm"] = perm
-            else:
-                x_perm = x
-                meta["perm"] = None
+        # supported shapes: rank == 4 or rank == 5
+        if rank == 4:
+            # (batch, H, W, elements) -> elements already the channel axis
+            return x, meta
 
-            shape = ops.shape(x_perm)
-            left = shape[:elem_axis]
-            elem = shape[elem_axis]
-            ch = shape[elem_axis + 1]
-            right = shape[elem_axis + 2 :]
-            new_elem = elem * ch
-            new_shape = left + (new_elem,) + right
-            x_reshaped = ops.reshape(x_perm, new_shape)
-            meta.update({"stacked": True, "elem_axis": elem_axis, "elem": elem, "ch": ch})
+        if rank == 5:
+            # (batch, H, W, elements, n_ch)
+            shape = ops.shape(x)
+            elem = shape[-2]
+            ch = shape[-1]
[…]
