### 2025-11-04 3d57fb0
Fix input and output data types in ReshapeToGrid operation and adjust axis calculation
