### 2025-11-24 caf1e09
📝 Add docstrings to `154-echonet-conversion-script-needs-avi-support`
Docstrings generation was requested by @swpenninga.

* https://github.com/tue-bmd/zea/pull/173#issuecomment-3542385161

The following files were modified:

* `zea/data/convert/__main__.py`
* `zea/data/convert/camus.py`
* `zea/data/convert/echonet.py`
* `zea/data/convert/echonetlvh/convert_raw_to_usbmd.py`
* `zea/data/convert/echonetlvh/precompute_crop.py`
* `zea/data/convert/matlab.py`
* `zea/data/convert/picmus.py`
* `zea/data/convert/utils.py`
* `zea/tools/fit_scan_cone.py`
