+++ zea/data/convert/echonet.py
@@ -21,14 +21,16 @@ from zea.data.convert.utils import unzip
 
 
 def segment(tensor, number_erasing=0, min_clip=0):
-    """Segments the background of the echonet images by setting it to 0 and creating a hard edge.
-
-    Args:
-        tensor (ndarray): Input image (sc) with 3 dimensions. (N, 112, 112)
-        number_erasing (float, optional): number to fill the background with.
+    """
+    Segment the background of EchoNet frames by masking regions outside the ultrasound cone and enforcing edge clipping.
+    
+    Parameters:
+        tensor (ndarray): Input array of shape (N, 112, 112); frames are modified in-place to mask background regions.
+        number_erasing (float, optional): Value used to fill masked background regions. Defaults to 0.
+        min_clip (float, optional): If > 0, values on the computed cone edge will be clipped to be at least this value. Defaults to 0.
+    
     Returns:
-        tensor (ndarray): Segmented matrix of same dimensions as input
-
+        ndarray: The same `tensor` array with background regions set to `number_erasing` and edge pixels clipped when `min_clip > 0`.
     """
     # Start with the upper part
 
@@ -238,7 +240,16 @@ def cartesian_to_polar_matrix(
 
 
 def find_split_for_file(file_dict, target_file):
-    """Function that finds the split for a given file in a dictionary."""
+    """
+    Locate which split contains a given filename.
+    
+    Parameters:
+        file_dict (dict): Mapping from split name (e.g., "train", "val", "test", "rejected") to an iterable of filenames.
+        target_file (str): Filename to search for within the split lists.
+    
+    Returns:
+        str: The split name that contains `target_file`, or `"rejected"` if the file is not found.
+    """
     for split, files in file_dict.items():
         if target_file in files:
             return split
@@ -247,6 +258,12 @@ def find_split_for_file(file_dict, target_file):
 
 
 def count_init(shared_counter):
+    """
+    Initialize the module-level shared counter used by worker processes.
+    
+    Parameters:
+        shared_counter (multiprocessing.Value): A process-shared integer Value that will be assigned to the module-global COUNTER for coordinated counting across processes.
+    """
     global COUNTER
     COUNTER = shared_counter
 
@@ -290,7 +307,18 @@ class H5Processor:
         return translate(data, self._process_range, self.range_to)
 
     def get_split(self, hdf5_file: str, sequence):
-        """Determine the split for a given file."""
+        """
+        Determine the dataset split label for a given file and its image sequence.
+        
+        This method checks acceptance based on the first frame of `sequence`. If explicit splits were provided to the processor, it returns the split found for `hdf5_file` (and asserts that the acceptance result matches the split). If no explicit splits are provided, rejected sequences are labeled `"rejected"`. Accepted sequences increment a shared counter and are assigned `"val"`, `"test"`, or `"train"` according to the processor's `num_val` and `num_test` quotas.
+        
+        Parameters:
+            hdf5_file (str): Filename or identifier used to look up an existing split when splits are provided.
+            sequence (array-like): Time-ordered sequence of images; the first frame is used for acceptance checking.
+        
+        Returns:
+            str: One of `"train"`, `"val"`, `"test"`, or `"rejected"` indicating the assigned split.
+        """
         # Always check acceptance
         accepted = accept_shape(sequence[0])
 
@@ -320,6 +348,14 @@ class H5Processor:
             return "train"
 
     def validate_split_copy(self, split_file):
+        """
+        Validate that a generated split YAML matches the original splits provided to the processor.
+        
[…]
