+++ zea/data/convert/camus.py
@@ -24,22 +24,15 @@ from zea.data.convert.utils import unzip
 
 def transform_sc_image_to_polar(image_sc, output_size=None, fit_outline=True):
     """
-    Transform a scan converted input image (cone) into square
-        using radial stretching and downsampling. Note that it assumes the background to be zero!
-        Please verify if your results make sense, especially if the image contains black parts
-        at the edges. This function is not perfect by any means, but it works for most cases.
-
-    Args:
-        image (numpy.ndarray): Input image as a 2D numpy array (height, width).
-        output_size (tuple, optional): Output size of the image as a tuple.
-            Defaults to image_sc.shape.
-        fit_outline (bool, optional): Whether to fit a polynomial the outline of the image.
-            Defaults to True. If this is set to False, and the ultrasound image contains
-            some black parts at the edges, weird artifacts can occur, because the jagged outline
-            is stretched to the desired width.
-
+    Convert a scan-converted 2D ultrasound (cone) image into a square (polar-like) representation.
+    
+    Parameters:
+        image_sc (numpy.ndarray): Input 2D image array (height, width) with background assumed to be zero.
+        output_size (tuple, optional): Desired output size as (height, width). Defaults to the input shape.
+        fit_outline (bool, optional): If True, fit smooth polynomial outlines to image borders before remapping to reduce edge artifacts; if False, use the raw detected outline.
+    
     Returns:
-        numpy.ndarray: Squared image as a 2D numpy array (height, width).
+        numpy.ndarray: Squared 2D image resized to `output_size`.
     """
     assert len(image_sc.shape) == 2, "function only allows for 2D data"
 
@@ -116,14 +109,20 @@ def transform_sc_image_to_polar(image_sc, output_size=None, fit_outline=True):
 
 
 def sitk_load(filepath: str | Path) -> Tuple[np.ndarray, Dict[str, Any]]:
-    """Loads an image using SimpleITK and returns the image and its metadata.
-
-    Args:
-        filepath: Path to the image.
-
+    """
+    Load an image file using SimpleITK and return its pixel array along with extracted metadata.
+    
+    Parameters:
+        filepath (str | Path): Path to the image file to load.
+    
     Returns:
-        - ([N], H, W), Image array.
-        - Collection of metadata.
+        im_array (np.ndarray): NumPy array produced by SimpleITK's GetArrayFromImage. For 3D images the shape is (Z, H, W); for 2D images the shape is (H, W). Pixel dtype is preserved.
+        metadata (dict): Dictionary with keys:
+            - "origin": image origin tuple,
+            - "ElementSpacing": spacing tuple,
+            - "direction": direction cosines tuple,
+            - "NDims": number of image dimensions,
+            - "metadata": mapping of SimpleITK metadata keys to their values.
     """
     # Load image and save info
     import SimpleITK as sitk
@@ -149,14 +148,16 @@ def sitk_load(filepath: str | Path) -> Tuple[np.ndarray, Dict[str, Any]]:
 
 
 def process_camus(source_path, output_path, output_path_npz=None, overwrite=False):
-    """Converts the camus database to the zea format.
-
-    Args:
-        source_path (str, pathlike): The path to the original camus file.
-        output_path (str, pathlike): The path to the output file.
-        output_path_npz (str, pathlike, optional): The path to the numpy output if desired.
-        overwrite (bool, optional): Set to True to overwrite existing file.
-            Defaults to False.
+    """
+    Convert a CAMUS image or sequence to ZE A format and write output files.
+    
[…]
