+++ zea/data/convert/__main__.py
@@ -8,6 +8,25 @@ from .echonetlvh.convert_raw_to_usbmd import convert_echonetlvh
 
 
 def get_args():
+    """
+    Build and parse command-line arguments for converting raw datasets to a USBMD dataset.
+    
+    Returns:
+        argparse.Namespace: Parsed arguments with the following attributes:
+            dataset (str): One of "echonet", "echonetlvh", "camus", "picmus", "matlab".
+            src (str): Source folder path.
+            dst (str): Destination folder path.
+            dst_npz (str|None): Optional alternate destination path for numpy output.
+            split_path (str|None): Optional path to a split.yaml to copy dataset splits.
+            no_hyperthreading (bool): Disable hyperthreading for multiprocessing.
+            frames (list[str]): MATLAB frames spec (e.g., ["all"], integers, or ranges like "4-8").
+            no_rejection (bool): EchonetLVH flag to skip manual_rejections.txt filtering.
+            batch (str|None): EchonetLVH Batch directory to process (e.g., "Batch2").
+            convert_measurements (bool): EchonetLVH flag to convert only measurements CSV.
+            convert_images (bool): EchonetLVH flag to convert only image files.
+            max_files (int|None): EchonetLVH maximum number of files to process.
+            force (bool): EchonetLVH flag to force recomputation even if parameters exist.
+    """
     parser = argparse.ArgumentParser(description="Convert raw data to a USBMD dataset.")
     parser.add_argument(
         "dataset",
@@ -79,6 +98,11 @@ def get_args():
 
 
 def main():
+    """
+    Parse command-line arguments and dispatch to the selected dataset conversion routine.
+    
+    This function obtains CLI arguments via get_args() and calls the corresponding converter (convert_echonet, convert_echonetlvh, convert_camus, convert_picmus, or convert_matlab) based on args.dataset. Raises a ValueError if args.dataset is not one of the supported choices.
+    """
     args = get_args()
     if args.dataset == "echonet":
         convert_echonet(args)

[…]
