### 2025-11-26 c608962
segment clarius
