+++ segment_carotid_clarius.py
@@ -1,0 +1,251 @@
+import argparse
+import os
+from pathlib import Path
+
+os.environ["KERAS_BACKEND"] = "tensorflow"
+os.environ["TF_CPP_MIN_LOG_LEVEL"] = "3"
+
+import cv2
+import numpy as np
+import matplotlib.pyplot as plt
+from keras import ops
+from tqdm import tqdm
+
+from zea import init_device, log
+from zea.models.carotid_segmenter import CarotidSegmenter
+from zea.visualize import plot_shape_from_mask, set_mpl_style
+from zea.tensor_ops import vmap
+
+
+def load_video_frames(video_path, max_seconds=9):
+    """Load frames from video file for the first max_seconds.
+    
+    Args:
+        video_path: Path to the video file.
+        max_seconds: Maximum number of seconds to load.
+        
+    Returns:
+        frames: List of frames as numpy arrays (H, W, 3).
+        fps: Frames per second of the video.
+    """
+    cap = cv2.VideoCapture(str(video_path))
+    
+    if not cap.isOpened():
+        raise ValueError(f"Could not open video file: {video_path}")
+    
+    fps = cap.get(cv2.CAP_PROP_FPS)
+    max_frames = int(fps * max_seconds)
+    
+    frames = []
+    frame_count = 0
+    
+    log.info(f"Loading first {max_seconds} seconds ({max_frames} frames) at {fps:.2f} fps")
+    
+    while frame_count < max_frames:
+        ret, frame = cap.read()
+        if not ret:
+            log.warning(f"Video ended at frame {frame_count}")
+            break
+        
+        # Convert BGR to RGB
+        frame_rgb = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
+        frames.append(frame_rgb)
+        frame_count += 1
+    
+    cap.release()
+    log.info(f"Loaded {len(frames)} frames")
+    
+    return frames, fps
+
+
+def preprocess_frames(frames):
+    """Preprocess frames for the carotid segmenter.
+    
+    Args:
+        frames: List of RGB frames (H, W, 3).
+        
+    Returns:
+        Batch tensor of shape (N, H, W, 1) normalized to [0, 1].
+    """
+    # Convert to grayscale and normalize
+    gray_frames = []
+    for frame in frames:
+        # Convert RGB to grayscale
+        gray = cv2.cvtColor(frame, cv2.COLOR_RGB2GRAY)
+        # Normalize to [0, 1]
+        gray_normalized = gray.astype(np.float32) / 255.0
+        gray_frames.append(gray_normalized)
+    
+    # Stack and add channel dimension
+    batch = np.stack(gray_frames, axis=0)  # (N, H, W)
+    batch = batch[..., None]  # (N, H, W, 1)
+    batch = batch[:,26:,182:438,:] # crop to center square
+    
+    return ops.convert_to_tensor(batch)
+
+
+def segment_frames(model, frames, batch_size=8):
+    """Segment frames using the carotid segmenter.
+    
+    Args:
+        model: CarotidSegmenter model.
+        frames_batch: Tensor of shape (N, H, W, 1).
+        
+    Returns:
+        masks: Binary masks of shape (N, H, W).
+    """
+    log.info(f"Segmenting {frames.shape[0]} frames...")
+    
+    masks = vmap(model, batch_size=batch_size)(frames[:, None, ...])[:, 0, ...]
+    
+    # Remove channel dimension and apply threshold
+    masks = ops.squeeze(masks, axis=-1)
+    masks_binary = ops.where(masks > 0.5, 1, 0)
+    masks_binary = ops.convert_to_numpy(masks_binary)
+    
+    return masks_binary
+
+
[…]
