### 2025-12-15 c399a52
fix snippets paper and test
