+++ paper/paper.bib
@@ -3,6 +3,7 @@
   title        = {zea: {A} {T}oolbox for {C}ognitive {U}ltrasound {I}maging},
   url          = {https://github.com/tue-bmd/zea},
   eprint       = {2512.01433},
+  doi          = {10.48550/ARXIV.2512.01433},
   year         = {2025}
 }
 
@@ -377,12 +378,16 @@
   year         = {2016}
 }
 
-@inproceedings{ouyang2019echonet,
-  author       = {Ouyang, David and He, Bryan and Ghorbani, Amirata and Lungren, Matt P and Ashley, Euan A and Liang, David H and Zou, James Y},
-  booktitle    = {NeurIPS ML4H Workshop: Vancouver, BC, Canada},
-  title        = {{E}chonet-dynamic: a large new cardiac motion video data resource for medical machine learning},
-  volume       = {5},
-  year         = {2019}
+@article{ouyang2020video,
+  title        = {Video-based AI for beat-to-beat assessment of cardiac function},
+  author       = {Ouyang, David and He, Bryan and Ghorbani, Amirata and Yuan, Neal and Ebinger, Joseph and Langlotz, Curtis P and Heidenreich, Paul A and Harrington, Robert A and Liang, David H and Ashley, Euan A and others},
+  journal      = {Nature},
+  volume       = {580},
+  doi          = {10.1038/s41586-020-2145-8},
+  number       = {7802},
+  pages        = {252--256},
+  year         = {2020},
+  publisher    = {Nature Publishing Group}
 }
 
 @article{van2025patient,
