### 2025-12-17 #8574 @tomelse
I won't be available to review until mid January, sorry. Happy to do it then if still needed though.

> Hi @xurui1, @ebezzam, @suyashkumar, @tomelse, @rafaelorozco! 👋 Would you be able to review this submission to JOSS?
>
> If you haven't reviewed for JOSS before, we carry out our checklist-driven reviews here in GitHub issues following [these guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html). If you have any questions, let me know.
