### 2026-01-05 209c05c
early return + explain margin in docstrings
