### 2026-01-05 6e0c793
include test coverage of scan.grid & dependencies
