### 2026-01-21 #8574 @tristan-deep
@adamltyson from [the list of past reviewers](https://reviewers.joss.theoj.org/lookup), I see [Scivision](https://github.com/Scivision) also has ultrasound listed under their review topics. Alternatively we could wait for @tomelse, but don't know if that would interfere with the other 2 reviewer's timeline.
