### 2026-01-22 #9881 @adamltyson
@tomelse, @benfrey, @rafaelorozco this is the review thread for the paper. All of our communications will happen here from now on.

As a reviewer, the first step is to create a checklist for your review by entering:
