### 2026-01-22 #9881 @tristan-deep
@adamltyson looks like version in this review thread is still v0.0.1, while in the [pre-review](https://github.com/openjournals/joss-reviews/issues/8574) it was bumped to v0.0.9
