### 2026-01-22 #9881 @adamltyson
I don't think it's an issue. One of the very last things we do after review, and before publishing is to set the version to the latest one (and this is the one that gets archived).
