### 2026-01-26 1fd54ae
Implement rf2iq cli operation
