### 2026-02-02 83c9b67
deeplab mini architecture
