+++ zea/models/deeplabv3.py
@@ -60,11 +60,11 @@ def DilatedSpatialPyramidPooling(dspp_input):
     Returns:
         Tensor: Multi-scale feature representation
     """
-    dims = dspp_input.shape
+    dims = ops.shape(dspp_input)
     x = layers.AveragePooling2D(pool_size=(dims[-3], dims[-2]))(dspp_input)
     x = convolution_block(x, kernel_size=1, use_bias=True)
     out_pool = layers.UpSampling2D(
-        size=(dims[-3] // x.shape[1], dims[-2] // x.shape[2]),
+        size=(dims[-3] // ops.shape(x)[1], dims[-2] // ops.shape(x)[2]),
         interpolation="bilinear",
     )(x)
 
@@ -114,7 +114,7 @@ def DeeplabV3Plus(image_shape, num_classes, pretrained_weights=None):
     x = DilatedSpatialPyramidPooling(x)
 
     input_a = layers.UpSampling2D(
-        size=(image_shape[0] // 4 // x.shape[1], image_shape[1] // 4 // x.shape[2]),
+        size=(image_shape[0] // 4 // ops.shape(x)[1], image_shape[1] // 4 // ops.shape(x)[2]),
         interpolation="bilinear",
     )(x)
     input_b = resnet50.get_layer("conv2_block3_2_relu").output
@@ -124,8 +124,153 @@ def DeeplabV3Plus(image_shape, num_classes, pretrained_weights=None):
     x = convolution_block(x)
     x = convolution_block(x)
     x = layers.UpSampling2D(
-        size=(image_shape[0] // x.shape[1], image_shape[1] // x.shape[2]),
+        size=(image_shape[0] // ops.shape(x)[1], image_shape[1] // ops.shape(x)[2]),
         interpolation="bilinear",
     )(x)
     model_output = layers.Conv2D(num_classes, kernel_size=(1, 1), padding="same")(x)
     return keras.Model(inputs=model_input, outputs=model_output)
+
+
+def conv_bn_relu(x, filters, k=3, d=1):
+    """
+    Simple convolution -> batch normalization -> ReLU block.
+
+    This helper implements a common pattern used throughout the DeepLabV3+
+    decoder and ASPP: a 2D convolution followed by BatchNormalization and
+    a ReLU activation.
+
+    Args:
+        x (Tensor): Input tensor.
+        filters (int): Number of output channels for the convolution.
+        k (int): Kernel size for the convolution. Defaults to 3.
+        d (int): Dilation rate for the convolution (atrous conv). Defaults to 1.
+
+    Returns:
+        Tensor: Activated output tensor after Conv2D -> BatchNormalization -> ReLU.
+    """
+    x = layers.Conv2D(
+        filters, k, padding="same", dilation_rate=d, use_bias=False, kernel_initializer="he_normal"
+    )(x)
+    x = layers.BatchNormalization()(x)
+    return layers.ReLU()(x)
+
+
+def ASPP(x, filters):
+    """
+    Atrous Spatial Pyramid Pooling (ASPP) module.
+
+    ASPP captures multi-scale context by computing several parallel branches:
+    - Image-level pooling branch (global context)
+    - 1x1 convolution branch
+    - 3x3 dilated convolutions with dilation rates 6, 12, 18
+
+    The outputs of all branches are concatenated and followed by a 1x1
+    convolution to fuse the multi-scale features. This is the ASPP used in
+    DeepLabV3/V3+ for robust multi-scale context aggregation.
+
+    Args:
+        x (Tensor): Input feature map (typically the backbone features at OS=16).
+        filters (int): Number of filters to use inside each ASPP branch and for
+                       the final projection.
+
+    Returns:
[…]
