### 2026-02-10 317e153
allow specifying initial_noise
