### 2026-02-10 4c14817
initial noise
