### 2026-02-10 tue-bmd/zea#258 @benfrey
Review 1/n:  Documentation (seeking support) #9881
Hi Tristan,

Wow, this is an incredible toolbox that you and your co-authors have developed! I really enjoyed reviewing some of your examples and paper - especially the agentic work. I believe the toolbox will have a large impact on the community moving forward.

I did have a few comments before completing the review, which I broke up into a few repo issues/requests.

The first one involves the documentation - overall, your documentation was very well-explained and developed! One of the reviewer checkboxes involved a section for "parties wishing to (3) seek support". Would it be possible to add a section or a note to your documentation, either pointing users to personal contact information or a section labeled "Receiving Support" or something similar? Maybe the best way to receive support is through submitting an issue, and writing this in your documentation may help users who are lost but haven't had much experience with git mechanics.

Related to JOSS review #9881
