### 2026-02-10 tue-bmd/zea#259 @benfrey
JOSS review 2/3: Documentation (statement of need)
Hi again Tristan,

Your paper was written very well and thoroughly. Thank you for making my job as a reviewer so easy!

One suggestion I had would be to add a sentence or two on the [main page of your documentation](https://zea.readthedocs.io/en/latest/index.html) that helps motivate a) what problems in ultrasound that zea is designed to solve, and b) who your target market is. For example, I think that zea can help significantly decrease the time from idea conceptualization to implementation via bypassing the time to develop dataloaders, debug autodiff frameworks, connect beamformers, etc..

For your target market, is it mainly graduate/postdoc ultrasound engineers, researchers/engineers in the ultrasound industry, researchers outside of ultrasound, or all of the above?

Thanks in advance for considering!

Related to JOSS review #9881
