### 2026-02-10 tue-bmd/zea#260 @benfrey
JOSS review 3/3: Software paper (references)
The software paper was excellent! I just had a question on the reference citation formatting.

After reading the citation syntax, I am a little confused about how the citation style impacts the rendering (in reference to the [citation style](https://pandoc.org/MANUAL.html#extension-citations) guideline). For example, in the second paragraph of the summary, there are many references in successive order with the style “author et al. (year)...”. Is this consistent with the required formatting?

Thanks!

Related to JOSS review #9881
