+++ zea/beamform/beamformergrid.py
@@ -1,0 +1,284 @@
+"""Main sos-grid based beamforming functions for ultrasound imaging."""
+
+import keras
+import numpy as np
+from keras import ops
+
+from zea.func.tensor import vmap
+
+def tof_correction_grid(
+    data,
+    sos_grid,
+    x_sos_grid,
+    z_sos_grid,
+    flatgrid,
+    t0_delays,
+    probe_geometry,
+    initial_times,
+    sampling_frequency,
+    demodulation_frequency,
+    f_number,
+    t_peak,
+    tx_waveform_indices,
+):
+    """Time-of-flight correction for a flat grid.
+
+    Args:
+        data (ops.Tensor): Input RF/IQ data of shape `(n_tx, n_ax, n_el, n_ch)`.
+        sos_grid (ops.Tensor): sound speed grid
+        x_sos_grid (ops.Tensor): x coordinates of the sound speed grid
+        z_sos_grid (ops.Tensor): z coordinates of the sound speed grid
+        flatgrid (ops.Tensor): Pixel locations x, y, z of shape `(n_pix, 3)`
+        t0_delays (ops.Tensor): Times at which the elements fire shifted such
+            that the first element fires at t=0 of shape `(n_tx, n_el)`
+        probe_geometry (ops.Tensor): Element positions x, y, z of shape (n_el, 3)
+        initial_times (Tensor): The probe transmit time offsets of shape `(n_tx,)`.
+        sampling_frequency (float): Sampling frequency.
+        demodulation_frequency (float): Demodulation frequency.
+        f_number (float): Focus number (ratio of focal depth to aperture size).
+        t_peak (ops.Tensor): Time of the peak of the pulse in seconds.
+            Shape `(n_waveforms,)`.
+        tx_waveform_indices (ops.Tensor): The indices of the waveform used for each
+            transmit of shape `(n_tx,)`.
+    Returns:
+        (ops.Tensor): time-of-flight corrected data
+        with shape: `(n_tx, n_pix, n_el, n_ch)`.
+    """
+
+    assert len(data.shape) == 4, (
+        "The input data should have 4 dimensions, "
+        f"namely n_tx, n_ax, n_el, n_ch, got {len(data.shape)} dimensions: {data.shape}"
+    )
+    assert data.shape[0]==data.shape[2], (
+        "The dataset should be multistatic"
+    )
+
+    def calculate_delays_grid(
+        data,
+        grid,
+        sos_grid,
+        x_sos_grid,
+        z_sos_grid,
+        t0_delays,
+        probe_geometry,
+        initial_times,
+        sampling_frequency,
+        f_number,
+        t_peak=t_peak,
+        tx_waveform_indices=tx_waveform_indices,
+    ):
+        """
+        Uses supplied sos_grid to interpolate the delays in samples to every pixel in the grid
+        , only valid for multistatic data.
+
+            
+        The delay consists of two components: The transmit delay and the
+        receive delay.
+
+        The transmit delay is the delay between transmission and the
+        wavefront reaching the pixel.
+
+        The receive delay is the delay between the
+        wavefront reaching a pixel and the reflections returning to a specific
+        element.
+
+        Args:
+            data (ops.Tensor): Input RF/IQ data of shape `(n_tx, n_ax, n_el, n_ch)`.
+            grid (Tensor): The pixel coordinates to beamform to of shape `(n_pix, 3)`.
[…]
