+++ docs/source/index.rst
@@ -10,8 +10,6 @@ Welcome to the documentation for the ``zea`` package: *A Toolbox for Cognitive U
 - A collection of pretrained :doc:`models` for ultrasound image and signal processing.
 - Multi-Backend Support via `Keras3 <https://keras.io/keras_3/>`_: You can use `PyTorch <https://github.com/pytorch/pytorch>`_, `TensorFlow <https://github.com/tensorflow/tensorflow>`_, or `JAX <https://github.com/google/jax>`_.
 
-Check out the :doc:`about` page for more information and the motivation behind ``zea``.
-
 .. admonition:: Beta!
    :class: warning
 
