### 2026-02-11 #9881 @tristan-deep
> Hi [@adamltyson](https://github.com/adamltyson) and [@tristan-deep](https://github.com/tristan-deep) , thanks for the review reminder! I believe I have almost completed my review. I added a few minor suggestions as "Issues" on the tue-bmd/zea repo (while referencing [#9881](https://github.com/openjournals/joss-reviews/issues/9881)), but I'm not sure if the hyperlinks were created in this thread. Here are the links to the three small suggestions:
>
> 1. [Review 1/n: Documentation (seeking support) #9881](https://github.com/tue-bmd/zea/issues/258)
> 2. [Review 2/n: Documentation (statement of need) #9881](https://github.com/tue-bmd/zea/issues/259)
> 3. [Review 3/n: Software paper (references) #9881](https://github.com/tue-bmd/zea/issues/260)
>
> zea will make a great contribution to the ultrasound community! Thanks again for reaching out about reviewing it.

Hi @benfrey, thanks a lot for the review and the positive remarks! I have addressed your feedback in https://github.com/tue-bmd/zea/pull/261, and also individually answered your questions in the issues you linked.
