+++ zea/data/data_format.py
@@ -2,6 +2,8 @@
 Functions to write and validate datasets in the zea format.
 """
 
+from __future__ import annotations
+
 import inspect
 from dataclasses import dataclass
 from pathlib import Path
@@ -14,6 +16,299 @@ from zea import log
 from zea.data.file import File, validate_file
 from zea.internal.checks import _DATA_TYPES
 from zea.internal.utils import first_not_none_item
+from zea.probes import Probe
+from zea.scan import Scan
+
+
+def generate_zea_dataset_from_scan_probe(
+    path: str | Path,
+    scan: Scan,
+    probe: Probe | None = None,
+    raw_data=None,
+    aligned_data=None,
+    envelope_data=None,
+    beamformed_data=None,
+    image=None,
+    image_sc=None,
+    description="No description was supplied",
+):
+    generate_zea_dataset(
+        path=path,
+        raw_data=raw_data,
+        aligned_data=aligned_data,
+        envelope_data=envelope_data,
+        beamformed_data=beamformed_data,
+        image=image,
+        image_sc=image_sc,
+        probe_geometry=scan.probe_geometry if probe is None else probe.probe_geometry,
+        sampling_frequency=scan.sampling_frequency,
+        center_frequency=scan.center_frequency,
+        demodulation_frequency=scan.demodulation_frequency,
+        initial_times=scan.initial_times,
+        t0_delays=scan.t0_delays,
+        sound_speed=scan.sound_speed,
+        probe_name="generic",
+        description=description,
+        focus_distances=scan.focus_distances,
+        transmit_origins=scan.transmit_origins,
+        polar_angles=scan.polar_angles,
+        azimuth_angles=scan.azimuth_angles,
+        tx_apodizations=scan.tx_apodizations,
+        bandwidth_percent=scan.bandwidth_percent,
+        time_to_next_transmit=scan.time_to_next_transmit,
+        tgc_gain_curve=scan.tgc_gain_curve,
+        element_width=scan.element_width,
+        tx_waveform_indices=scan.tx_waveform_indices,
+        waveforms_one_way=scan.waveforms_one_way,
+        waveforms_two_way=scan.waveforms_two_way,
+        additional_elements=None,
+        event_structure=False,
+        cast_to_float=True,
+        overwrite=False,
+    )
+
+
+def filter_dict_for_function_arguments(func, dictionary):
+    """Filters a dictionary to only include keys that are arguments of a given function.
+
+    Args:
+        func (function): The function to filter the dictionary for.
+        dictionary (dict): The dictionary to filter.
+
+    Returns:
+        dict: A filtered dictionary containing only keys that are arguments of the function.
+    """
+    func_args = set(inspect.signature(func).parameters.keys())
+    filtered_dict = {k: v for k, v in dictionary.items() if k in func_args}
+    return filtered_dict
+
+
+def generate_zea_dataset(
+    path: str | Path,
+    raw_data=None,
+    aligned_data=None,
+    envelope_data=None,
+    beamformed_data=None,
+    image=None,
+    image_sc=None,
+    probe_geometry=None,
+    sampling_frequency=None,
+    center_frequency=None,
+    demodulation_frequency=None,
+    initial_times=None,
+    t0_delays=None,
+    sound_speed=None,
+    probe_name=None,
+    description="No description was supplied",
+    focus_distances=None,
+    transmit_origins=None,
+    polar_angles=None,
+    azimuth_angles=None,
+    tx_apodizations=None,
+    bandwidth_percent=None,
+    time_to_next_transmit=None,
+    tgc_gain_curve=None,
+    element_width=None,
+    tx_waveform_indices=None,
+    waveforms_one_way=None,
+    waveforms_two_way=None,
[…]
