+++ zea/models/unet.py
@@ -18,6 +18,11 @@ import keras
 from keras import layers
 
 from zea import log
+
+# ---------------------------------------------------------------------------
+# Registry for network architecture builders (not full Model classes).
+# This maps string names → callable(image_shape, **kwargs) → keras.Model
+# ---------------------------------------------------------------------------
 from zea.internal.registry import model_registry
 from zea.models.base import BaseModel
 from zea.models.layers import DownBlock, ResidualBlock, UpBlock, sinusoidal_embedding
@@ -217,5 +222,345 @@ def get_time_conditional_unetwork(
     return keras.Model([noisy_images, noise_variances], x, name="residual_unet")
 
 
+# ===========================================================================
+# Temporal layers and blocks
+# ===========================================================================
+
+
+@keras.saving.register_keras_serializable()
+class TemporalConv(layers.Layer):
+    """Lightweight temporal convolution operating on frames packed as channels.
+
+    Expects input of shape ``(B, H, W, T*C)`` where ``T`` frames of ``C``
+    channels are interleaved along the last axis.  Internally reshapes to
+    5-D ``(B, H, W, T, C)``, applies a ``(1, 1, k)`` Conv3D, and reshapes
+    back.
+
+    Args:
+        n_frames: Number of temporal frames ``T``.
+        filters: Number of output filters (per frame).  If ``None``, keeps
+            the same number as input channels per frame.
+        temporal_kernel_size: Kernel size along the temporal axis.
+        depthwise: If ``True``, use grouped (depthwise) convolution along
+            the temporal axis for minimal parameter cost.
+    """
+
+    def __init__(
+        self,
+        n_frames,
+        filters=None,
+        temporal_kernel_size=3,
+        depthwise=False,
+        **kwargs,
+    ):
+        super().__init__(**kwargs)
+        self.n_frames = n_frames
+        self.filters = filters
+        self.temporal_kernel_size = temporal_kernel_size
+        self.depthwise = depthwise
+        self._temporal_conv = None
+
+    def build(self, input_shape):
+        total_channels = input_shape[-1]
+        assert total_channels % self.n_frames == 0, (
+            f"Last dim ({total_channels}) must be divisible by n_frames ({self.n_frames})"
+        )
+        c = total_channels // self.n_frames
+        out_filters = self.filters if self.filters is not None else c
+
+        groups = out_filters if self.depthwise else 1
+        self._temporal_conv = layers.Conv3D(
+            filters=out_filters,
+            kernel_size=(1, 1, self.temporal_kernel_size),
+            padding="same",
+            groups=groups,
+        )
+        self._temporal_conv.build((None, None, None, self.n_frames, c))
+        super().build(input_shape)
+
+    def call(self, x):
+        shape = keras.ops.shape(x)
+        b, h, w = shape[0], shape[1], shape[2]
+        c = shape[3] // self.n_frames
+
+        x = keras.ops.reshape(x, (b, h, w, self.n_frames, c))
+        x = self._temporal_conv(x)
+        out_c = keras.ops.shape(x)[-1]
+        return keras.ops.reshape(x, (b, h, w, self.n_frames * out_c))
+
+    def get_config(self):
+        config = super().get_config()
+        config.update(
+            {
+                "n_frames": self.n_frames,
+                "filters": self.filters,
+                "temporal_kernel_size": self.temporal_kernel_size,
+                "depthwise": self.depthwise,
+            }
+        )
+        return config
+
+
+@keras.saving.register_keras_serializable()
+class TemporalAttention(layers.Layer):
[…]
