+++ zea/models/diffusion.py
@@ -30,14 +30,18 @@ from zea.internal.core import Object
 from zea.internal.operators import Operator
 from zea.internal.registry import diffusion_guidance_registry, model_registry, operator_registry
 from zea.internal.utils import fn_requires_argument
-from zea.models.dense import get_time_conditional_dense_network
 from zea.models.generative import DeepGenerativeModel
 from zea.models.preset_utils import register_presets
 from zea.models.presets import diffusion_model_presets
-from zea.models.unet import get_time_conditional_unetwork
 from zea.models.utils import LossTrackerWrapper
 
 tf = _import_tf()
+
+VALID_NETWORK_MODELS = [
+    "unet_time_conditional",
+    "dense_time_conditional",
+    "unet_temporal_time_conditional",
+]
 
 
 @model_registry(name="diffusion")
@@ -97,19 +101,26 @@ class DiffusionModel(DeepGenerativeModel):
         self.min_t = min_t
         self.max_t = max_t
 
-        if network_name == "unet_time_conditional":
-            self.network = get_time_conditional_unetwork(
-                image_shape=self.input_shape,
-                **self.network_kwargs,
-            )
-        elif network_name == "dense_time_conditional":
-            assert len(input_shape) == 1, "Dense network only supports 1D input"
-            self.network = get_time_conditional_dense_network(
-                input_dim=self.input_shape[0],
-                **self.network_kwargs,
-            )
-        else:
-            raise ValueError("Invalid network name provided.")
+        assert network_name in VALID_NETWORK_MODELS, (
+            f"Invalid network name: {network_name}. Valid options are: {VALID_NETWORK_MODELS}"
+        )
+        self.network = model_registry[network_name](
+            input_shape=self.input_shape,
+            **self.network_kwargs,
+        )
+        # if network_name == "unet_time_conditional":
+        #     self.network = get_time_conditional_unetwork(
+        #         image_shape=self.input_shape,
+        #         **self.network_kwargs,
+        #     )
+        # elif network_name == "dense_time_conditional":
+        #     assert len(input_shape) == 1, "Dense network only supports 1D input"
+        #     self.network = get_time_conditional_dense_network(
+        #         input_dim=self.input_shape[0],
+        #         **self.network_kwargs,
+        #     )
+        # else:
+        #     raise ValueError("Invalid network name provided.")
 
         # Also initialize the exponential moving average network
         self.ema_network = keras.models.clone_model(self.network)
