+++ zea/beamform/pfield.py
@@ -92,6 +92,13 @@ def compute_pfield(
     alpha_db = ops.cast(alpha_db, "float32")
     db_thresh = ops.cast(db_thresh, "float32")
 
+    # to tensor
+    probe_geometry = ops.convert_to_tensor(probe_geometry, dtype="float32")
+    grid_x = ops.convert_to_tensor(grid[:, :, 0], dtype="float32")
+    grid_z = ops.convert_to_tensor(grid[:, :, 2], dtype="float32")
+    t0_delays = ops.convert_to_tensor(t0_delays, dtype="float32")
+    tx_apodizations = ops.convert_to_tensor(tx_apodizations, dtype="float32")
+
     # probe params
     center_frequency = center_frequency / downmix  # downmixing the frequency
 
@@ -99,8 +106,6 @@ def compute_pfield(
     num_waveforms = 1  # number of waveforms in the pulse
 
     # array params
-    probe_geometry = ops.convert_to_tensor(probe_geometry, dtype="float32")
-
     pitch = ops.abs(probe_geometry[1, 0] - probe_geometry[0, 0])  # element pitch
 
     kerf = 0.1 * pitch  # for now this is hardcoded
@@ -116,70 +121,60 @@ def compute_pfield(
     lambda_min = sound_speed / (center_frequency * (1 + bandwidth_percent / 200))
     num_sub_elements = ops.ceil(element_width / lambda_min)
 
-    x_orig = ops.convert_to_tensor(grid[:, :, 0], dtype="float32")
-    z_orig = ops.convert_to_tensor(grid[:, :, 2], dtype="float32")
-
-    size_orig = ops.shape(x_orig)
+    size_orig = ops.shape(grid_x)
 
     # Nearest-neighbor downsampling the grid
-    x = x_orig[::downsample, ::downsample]
-    z = z_orig[::downsample, ::downsample]
-    size_downsampled = ops.shape(x)
+    grid_x = grid_x[::downsample, ::downsample]
+    grid_z = grid_z[::downsample, ::downsample]
+    size_downsampled = ops.shape(grid_x)
 
     # Coordinates of the points where pressure is needed
-    x = ops.reshape(x, (-1,))
-    z = ops.reshape(z, (-1,))
+    grid_x = ops.reshape(grid_x, (-1,))
+    grid_z = ops.reshape(grid_z, (-1,))
 
     # Centers of the transducer elements (x- and z-coordinates)
-    xe = (ops.arange(0.0, n_el) - (n_el - 1) / 2) * pitch
-    ze = ops.zeros(n_el)
-    the = ops.zeros(n_el)
+    element_x = (ops.arange(0.0, n_el) - (n_el - 1) / 2) * pitch
+    element_z = ops.zeros(n_el)
+    element_theta = ops.zeros(n_el)
 
     # Centroids of the sub-elements
     seg_length = element_width / num_sub_elements
-    tmp = (
+    sub_element_x = (
         -element_width / 2
         + seg_length / 2
         + ops.arange(0, num_sub_elements, dtype=seg_length.dtype) * seg_length
     )
-    xi = tmp
-    zi = ops.zeros((int(num_sub_elements),))
+    sub_element_z = ops.zeros((int(num_sub_elements),))
 
     # Distances between the points and the transducer elements
-    x_expanded = x[:, None, None]
-    xi_expanded = xi[None, :, None]
-    xe_expanded = xe[None, None, :]
+    delta_x = grid_x[:, None, None] - sub_element_x[None, :, None] - element_x[None, None, :]
+    delta_z = grid_z[:, None, None] - sub_element_z[None, :, None] - element_z[None, None, :]
 
[…]
