### 2026-02-16 672a24a
Avoid unwanted nans
