### 2026-02-16 64e7246
Big speedup through parallism
