+++ zea/beamform/pfield.py
@@ -23,8 +23,7 @@ import keras
 import numpy as np
 from keras import ops
 
-from zea import log
-from zea.func.tensor import sinc
+from zea.func.tensor import sinc, vmap
 from zea.internal.cache import cache_output
 
 
@@ -49,7 +48,6 @@ def compute_pfield(
     alpha=1,
     percentile=10,
     norm=True,
-    verbose=True,
 ):
     """Compute the pressure field for ultrasound imaging.
 
@@ -117,8 +115,6 @@ def compute_pfield(
 
     kerf = 0.1 * pitch  # for now this is hardcoded
     element_width = pitch - kerf
-
-    n_tx = len(tx_apodizations)
 
     # %------------------------------------%
     # % POINT LOCATIONS, DISTANCES & GRIDS %
@@ -188,75 +184,62 @@ def compute_pfield(
         # Apply the negative sign and exponential
         return ops.exp(-exponent)
 
-    p_list = []
+    # The frequency response is a pulse-echo (transmit + receive) response.
+    # The spectrum of the pulse (pulse_spectrum) will be then multiplied
+    # by the frequency-domain tapering window of the transducer (probe_spectrum)
+    # The frequency step df is chosen to avoid interferences due to
+    # inadequate discretization.
+    # df = frequency step (must be sufficiently small):
+    # One has exp[-i(k r + w delay)] = exp[-2i pi(f r/c + f delay)] in the Eq.
+    # One wants: the phase increment 2pi(df r/c + df delay) be < 2pi.
+    # Therefore: df < 1/(r/c + delay).
 
-    if verbose:
-        log.info("Computing pressure field for all transmits")
-        progbar = keras.utils.Progbar(n_tx, unit_name="transmits")
-    for j in range(0, n_tx):
-        # delays and apodization of transmit event
-        delays_tx = t0_delays[j]
-        tx_apodization = tx_apodizations[j]
+    freq_step = 1 / (ops.max(distance / sound_speed) + ops.max(t0_delays))
+    freq_step = frequency_step * freq_step
 
-        # The frequency response is a pulse-echo (transmit + receive) response.
-        # The spectrum of the pulse (pulse_spectrum) will be then multiplied
-        # by the frequency-domain tapering window of the transducer (probe_spectrum)
-        # The frequency step df is chosen to avoid interferences due to
-        # inadequate discretization.
-        # df = frequency step (must be sufficiently small):
-        # One has exp[-i(k r + w delay)] = exp[-2i pi(f r/c + f delay)] in the Eq.
-        # One wants: the phase increment 2pi(df r/c + df delay) be < 2pi.
-        # Therefore: df < 1/(r/c + delay).
+    # FREQUENCY SAMPLES
+    num_freq = 2 * ops.cast(ops.ceil(center_frequency / freq_step), "int32") + 1
+    freq = ops.arange(0, num_freq) * freq_step
 
-        freq_step = 1 / (ops.max(distance / sound_speed) + ops.max(delays_tx))
-        freq_step = frequency_step * freq_step
+    # keep the significant components only by using db_thresh
+    spectrum = ops.abs(
+        pulse_spectrum(2 * np.pi * freq) * ops.cast(probe_spectrum(2 * np.pi * freq), "complex64")
+    )
+    gain_db = 20 * ops.log10(keras.config.epsilon() + spectrum / (ops.max(spectrum)))
+    idx = gain_db > db_thresh
 
-        # FREQUENCY SAMPLES
-        num_freq = 2 * ops.cast(ops.ceil(center_frequency / freq_step), "int32") + 1
[…]
