### 2026-02-16 31df860
Use ops.scan
