### 2026-02-16 891ffdc
noise correlation
