+++ zea/models/unet.py
@@ -357,14 +357,67 @@ class TemporalAttention(layers.Layer):
         return config
 
 
+@keras.saving.register_keras_serializable()
+class TemporalResidualBlock(layers.Layer):
+    """A residual block that applies spatial convolutions followed by a
+    lightweight temporal convolution.
+
+    Wraps a standard ``ResidualBlock`` and appends a ``TemporalConv``.
+
+    Args:
+        width: Number of spatial filters.
+        n_frames: Number of temporal frames.
+        temporal_kernel_size: Temporal conv kernel size.
+        depthwise: Use depthwise temporal conv.
+    """
+
+    def __init__(
+        self,
+        width,
+        n_frames,
+        temporal_kernel_size=3,
+        depthwise=False,
+        **kwargs,
+    ):
+        super().__init__(**kwargs)
+        self.width = width
+        self.n_frames = n_frames
+        self.temporal_kernel_size = temporal_kernel_size
+        self.depthwise = depthwise
+
+        self.spatial_block = ResidualBlock(width)
+        self.temporal_conv = TemporalConv(
+            n_frames=n_frames,
+            temporal_kernel_size=temporal_kernel_size,
+            depthwise=depthwise,
+        )
+
+    def call(self, x):
+        x = self.spatial_block(x)
+        x = x + self.temporal_conv(x)  # residual around temporal conv
+        return x
+
+    def get_config(self):
+        config = super().get_config()
+        config.update(
+            {
+                "width": self.width,
+                "n_frames": self.n_frames,
+                "temporal_kernel_size": self.temporal_kernel_size,
+                "depthwise": self.depthwise,
+            }
+        )
+        return config
+
+
 # ===========================================================================
 # Temporal UNet model and network builder
 # ===========================================================================
 
 
-@model_registry(name="unet_temporal_time_conditional")
-class UNetTemporalTimeConditional(BaseModel):
-    """Temporal UNet with (2+1)D factorized convolutions at the bottleneck
+@model_registry(name="unet_light_temporal_time_conditional")
+class UNetLightTemporalTimeConditional(BaseModel):
+    """Light temporal UNet with (2+1)D factorized convolutions at the bottleneck
     and time-conditional sinusoidal embeddings for video diffusion.
 
     Keeps the encoder and decoder **identical** to :class:`UNetTimeConditional`
@@ -407,7 +460,7 @@ class UNetTemporalTimeConditional(BaseModel):
         embedding_min_frequency=1.0,
         embedding_max_frequency=1000.0,
         embedding_dims=32,
-        name="unet_temporal_time_conditional",
+        name="unet_light_temporal_time_conditional",
         **kwargs,
     ):
         super().__init__(name=name, **kwargs)
@@ -424,7 +477,7 @@ class UNetTemporalTimeConditional(BaseModel):
         self.embedding_max_frequency = embedding_max_frequency
         self.embedding_dims = embedding_dims
 
-        self.network = get_temporal_time_conditional_unetwork(
+        self.network = get_light_temporal_time_conditional_unetwork(
             image_shape=self.image_shape,
             n_frames=self.n_frames,
             widths=self.widths,
@@ -462,7 +515,7 @@ class UNetTemporalTimeConditional(BaseModel):
         return self.network(*args, **kwargs)
 
 
-def get_temporal_time_conditional_unetwork(
+def get_light_temporal_time_conditional_unetwork(
     image_shape,
     n_frames,
     widths=None,
@@ -561,6 +614,217 @@ def get_temporal_time_conditional_unetwork(
     return keras.Model([noisy_images, noise_variances], x, name="temporal_residual_unet")
 
 
[…]
