+++ zea/scan.py
@@ -90,10 +90,7 @@ from zea.beamform.pixelgrid import (
     check_for_aliasing,
     polar_pixel_grid,
 )
-from zea.display import (
-    compute_scan_convert_2d_coordinates,
-    compute_scan_convert_3d_coordinates,
-)
+from zea.display import compute_scan_convert_2d_coordinates
 from zea.internal.parameters import Parameters, cache_with_dependencies
 
 
@@ -699,14 +696,14 @@ class Scan(Parameters):
         if value is None:
             return None
 
-        selected = self.selected_transmits
-        return value[:, selected]
+        return value[:, self.selected_transmits]
 
     @cache_with_dependencies("n_ax")
     def tgc_gain_curve(self):
         """Time gain compensation (TGC) curve of shape (n_ax,)."""
         value = self._params.get("tgc_gain_curve")
         if value is None:
+            log.warning("No TGC gain curve provided, using ones")
             return np.ones(self.n_ax)
         return value[: self.n_ax]
 
@@ -715,6 +712,7 @@ class Scan(Parameters):
         """Indices of the waveform used for each transmit event of shape (n_tx,)."""
         value = self._params.get("tx_waveform_indices")
         if value is None:
+            log.warning("No transmit waveform indices provided, using zeros")
             return np.zeros(self.n_tx, dtype=int)
 
         return value[self.selected_transmits]
@@ -733,7 +731,9 @@ class Scan(Parameters):
     def pfield(self) -> np.ndarray:
         """Compute or return the pressure field (pfield) for weighting
         of shape (n_tx, grid_size_z, grid_size_x)."""
-        assert not self.is_3d, "Pfield computation only supported for 2D scans"
+        if self.is_3d:
+            raise NotImplementedError("Pfield computation is not yet implemented for 3D scans.")
+
         pfield = compute_pfield(
             sound_speed=self.sound_speed,
             center_frequency=self.center_frequency,
@@ -794,28 +794,21 @@ class Scan(Parameters):
     )
     def coordinates_3d(self):
         """The coordinates for scan conversion."""
-        # TODO: no grid_size_y... this is broken
-        coords, _ = compute_scan_convert_3d_coordinates(
-            (self.grid_size_z, self.grid_size_x),
-            self.rho_range,
-            self.theta_range,
-            self.phi_range,
-            self.resolution,
-        )
-        return coords
+        raise NotImplementedError
 
     @cache_with_dependencies("phi_range", "coordinates_2d", "coordinates_3d")
     def coordinates(self):
         """Get the coordinates for scan conversion, will be 3D if phi_range is set,
         otherwise 2D."""
-        return self.coordinates_3d if getattr(self, "phi_range", None) else self.coordinates_2d
+        return self.coordinates_3d if self.is_3d else self.coordinates_2d
 
     @cache_with_dependencies("time_to_next_transmit")
     def pulse_repetition_frequency(self):
         """The pulse repetition frequency (PRF) [Hz]. Assumes a constant PRF."""
         if self.time_to_next_transmit is None:
-            log.warning("Time to next transmit is not set, cannot compute PRF")
-            return None
+            raise ValueError(
+                "Time to next transmit must be set to compute pulse repetition frequency"
+            )
 
         pulse_repetition_interval = np.mean(self.time_to_next_transmit)
 
@@ -833,8 +826,7 @@ class Scan(Parameters):
         """
         time_to_next_transmit = self._params.get("time_to_next_transmit")
         if time_to_next_transmit is None:
-            log.warning("Time to next transmit is not set, cannot compute fps")
-            return None
+            raise ValueError("Time to next transmit must be set to compute frames per second")
 
         # Check if fps is constant
         uniq = np.unique(time_to_next_transmit, axis=0)  # frame axis
