### 2026-02-17 722c55c
Omit time_to_next_transmit
