### 2026-02-17 25c42dc
improve diffusion interpretability
