+++ zea/models/unet.py
@@ -223,8 +223,10 @@ def get_time_conditional_unetwork(
 
 
 # ===========================================================================
-# Temporal layers and blocks
+# Temporal layers
 # ===========================================================================
+
+VALID_TEMPORAL_SCOPES = ("none", "bottleneck", "all")
 
 
 @keras.saving.register_keras_serializable()
@@ -275,7 +277,6 @@ class TemporalConv(layers.Layer):
             padding="same",
             groups=groups,
         )
-        self._temporal_conv.build((None, None, None, self.n_frames, c))
         super().build(input_shape)
 
     def call(self, x):
@@ -330,10 +331,6 @@ class TemporalAttention(layers.Layer):
             num_heads=self.num_heads,
             key_dim=key_dim,
         )
-        self._attn.build(
-            query_shape=(None, self.n_frames, c),
-            value_shape=(None, self.n_frames, c),
-        )
         super().build(input_shape)
 
     def call(self, x):
@@ -357,276 +354,61 @@ class TemporalAttention(layers.Layer):
         return config
 
 
-@keras.saving.register_keras_serializable()
-class TemporalResidualBlock(layers.Layer):
-    """A residual block that applies spatial convolutions followed by a
-    lightweight temporal convolution.
-
-    Wraps a standard ``ResidualBlock`` and appends a ``TemporalConv``.
-
-    Args:
-        width: Number of spatial filters.
-        n_frames: Number of temporal frames.
-        temporal_kernel_size: Temporal conv kernel size.
-        depthwise: Use depthwise temporal conv.
-    """
-
-    def __init__(
-        self,
-        width,
-        n_frames,
-        temporal_kernel_size=3,
-        depthwise=False,
-        **kwargs,
-    ):
-        super().__init__(**kwargs)
-        self.width = width
-        self.n_frames = n_frames
-        self.temporal_kernel_size = temporal_kernel_size
-        self.depthwise = depthwise
-
-        self.spatial_block = ResidualBlock(width)
-        self.temporal_conv = TemporalConv(
-            n_frames=n_frames,
-            temporal_kernel_size=temporal_kernel_size,
-            depthwise=depthwise,
-        )
-
-    def call(self, x):
-        x = self.spatial_block(x)
-        x = x + self.temporal_conv(x)  # residual around temporal conv
-        return x
-
-    def get_config(self):
-        config = super().get_config()
-        config.update(
-            {
-                "width": self.width,
-                "n_frames": self.n_frames,
-                "temporal_kernel_size": self.temporal_kernel_size,
-                "depthwise": self.depthwise,
-            }
-        )
-        return config
-
-
 # ===========================================================================
-# Temporal UNet model and network builder
+# Temporal UNet — unified model and network builder
 # ===========================================================================
 
 
-@model_registry(name="unet_light_temporal_time_conditional")
-class UNetLightTemporalTimeConditional(BaseModel):
-    """Light temporal UNet with (2+1)D factorized convolutions at the bottleneck
-    and time-conditional sinusoidal embeddings for video diffusion.
-
-    Keeps the encoder and decoder **identical** to :class:`UNetTimeConditional`
-    (standard spatial ``ResidualBlock`` / ``DownBlock`` / ``UpBlock``).  Only
-    the bottleneck is augmented with lightweight temporal convolutions and
-    optional temporal self-attention.  This follows the "95% identical"
-    design principle: temporal inductive bias is injected only where spatial
-    resolution is smallest and the cost is negligible.
-
[…]
