+++ zea/models/layers.py
@@ -23,47 +23,107 @@ def sinusoidal_embedding(x, embedding_min_frequency, embedding_max_frequency, em
     return embeddings
 
 
-def ResidualBlock(width):
-    """Residual block with swish activation."""
+@keras.saving.register_keras_serializable()
+class ResidualBlock(layers.Layer):
+    """Residual block with swish activation.
 
-    def apply(x):
-        input_width = ops.shape(x)[3]
-        if input_width == width:
-            residual = x
-        else:
-            residual = layers.Conv2D(width, kernel_size=1)(x)
-        x = layers.BatchNormalization(center=False, scale=False)(x)
-        x = layers.Conv2D(width, kernel_size=3, padding="same", activation="swish")(x)
-        x = layers.Conv2D(width, kernel_size=3, padding="same")(x)
-        x = layers.Add()([x, residual])
-        return x
+    If the input channel dimension differs from ``width``, a 1×1 convolution
+    is used to project the residual connection.
+    """
 
-    return apply
+    def __init__(self, width, **kwargs):
+        super().__init__(**kwargs)
+        self.width = width
+
+    def build(self, input_shape):
+        input_width = input_shape[-1]
+        self.needs_projection = input_width != self.width
+        if self.needs_projection:
+            self.proj = layers.Conv2D(self.width, kernel_size=1, name="residual_proj")
+        self.norm = layers.BatchNormalization(center=False, scale=False)
+        self.conv1 = layers.Conv2D(self.width, kernel_size=3, padding="same", activation="swish")
+        self.conv2 = layers.Conv2D(self.width, kernel_size=3, padding="same")
+        self.add = layers.Add()
+        super().build(input_shape)
+
+    def call(self, x):
+        residual = self.proj(x) if self.needs_projection else x
+        x = self.norm(x)
+        x = self.conv1(x)
+        x = self.conv2(x)
+        return self.add([x, residual])
+
+    def get_config(self):
+        config = super().get_config()
+        config.update({"width": self.width})
+        return config
 
 
-def DownBlock(width, block_depth):
-    """Downsampling block with residual connections."""
+@keras.saving.register_keras_serializable()
+class DownBlock(layers.Layer):
+    """Downsampling block with residual connections.
 
-    def apply(x):
-        x, skips = x
-        for _ in range(block_depth):
-            x = ResidualBlock(width)(x)
+    Expects a tuple ``(x, skips)`` where ``skips`` is a **mutable** list.
+    Appends intermediate activations to ``skips`` for the decoder.
+    """
+
+    def __init__(self, width, block_depth, **kwargs):
+        super().__init__(**kwargs)
+        self.width = width
+        self.block_depth = block_depth
+
+    def build(self, input_shape):
+        # input_shape is a list of shapes: [x_shape, skips_shape (ignored)]
+        self.res_blocks = [
+            ResidualBlock(self.width, name=f"res_{i}") for i in range(self.block_depth)
+        ]
+        self.pool = layers.AveragePooling2D(pool_size=2)
+        super().build(input_shape)
+
+    def call(self, inputs):
+        x, skips = inputs
+        for block in self.res_blocks:
+            x = block(x)
             skips.append(x)
-        x = layers.AveragePooling2D(pool_size=2)(x)
-        return x
+        x = self.pool(x)
+        return x, skips
 
-    return apply
+    def get_config(self):
+        config = super().get_config()
+        config.update({"width": self.width, "block_depth": self.block_depth})
+        return config
 
 
-def UpBlock(width, block_depth):
-    """Upsampling block with residual connections."""
+@keras.saving.register_keras_serializable()
+class UpBlock(layers.Layer):
[…]
