+++ zea/models/unet.py
@@ -233,70 +233,53 @@ VALID_TEMPORAL_SCOPES = ("none", "bottleneck", "all")
 class TemporalConv(layers.Layer):
     """Lightweight temporal convolution operating on frames packed as channels.
 
-    Expects input of shape ``(B, H, W, T*C)`` where ``T`` frames of ``C``
-    channels are interleaved along the last axis.  Internally reshapes to
-    5-D ``(B, H, W, T, C)``, applies a ``(1, 1, k)`` Conv3D, and reshapes
-    back.
+    For grayscale video data, the input shape is ``(B, H, W, T)`` where ``T``
+    is the number of frames. Internally reshapes to 5-D ``(B, H, W, T, 1)``,
+    applies a ``(1, 1, k)`` Conv3D, and reshapes back. Output shape matches
+    input (shape-preserving).
 
     Args:
-        n_frames: Number of temporal frames ``T``.
-        filters: Number of output filters (per frame).  If ``None``, keeps
-            the same number as input channels per frame.
+        n_frames: Number of temporal frames ``T`` (equals the channel dimension).
         temporal_kernel_size: Kernel size along the temporal axis.
-        depthwise: If ``True``, use grouped (depthwise) convolution along
-            the temporal axis for minimal parameter cost.
+        depthwise: If ``True``, use depthwise convolution (minimal params).
     """
 
     def __init__(
         self,
-        n_frames,
-        filters=None,
         temporal_kernel_size=3,
-        depthwise=False,
+        depthwise=True,
         **kwargs,
     ):
         super().__init__(**kwargs)
-        self.n_frames = n_frames
-        self.filters = filters
         self.temporal_kernel_size = temporal_kernel_size
         self.depthwise = depthwise
         self._temporal_conv = None
 
     def build(self, input_shape):
-        total_channels = input_shape[-1]
-        assert total_channels % self.n_frames == 0, (
-            f"Last dim ({total_channels}) must be divisible by n_frames ({self.n_frames})"
-        )
-        c = total_channels // self.n_frames
-        out_filters = self.filters if self.filters is not None else c
-
-        groups = out_filters if self.depthwise else 1
+        # Since C=1, filters=1 and groups=1 (depthwise with 1 channel is just regular conv)
         self._temporal_conv = layers.Conv3D(
-            filters=out_filters,
+            filters=1,
             kernel_size=(1, 1, self.temporal_kernel_size),
             padding="same",
-            groups=groups,
         )
         super().build(input_shape)
 
     def call(self, x):
         shape = keras.ops.shape(x)
-        b, h, w = shape[0], shape[1], shape[2]
-        c = shape[3] // self.n_frames
+        b, h, w, t = shape[0], shape[1], shape[2], shape[3]
 
-        x = keras.ops.reshape(x, (b, h, w, self.n_frames, c))
+        # Reshape (B, H, W, T) -> (B, H, W, T, 1)
+        x = keras.ops.reshape(x, (b, h, w, t, 1))
         x = self._temporal_conv(x)
-        out_c = keras.ops.shape(x)[-1]
-        return keras.ops.reshape(x, (b, h, w, self.n_frames * out_c))
+        # Reshape back (B, H, W, T, 1) -> (B, H, W, T)
+        return keras.ops.reshape(x, (b, h, w, t))
 
     def get_config(self):
         config = super().get_config()
         config.update(
             {
-                "n_frames": self.n_frames,
-                "filters": self.filters,
                 "temporal_kernel_size": self.temporal_kernel_size,
-                "depthwise": self.depthwise,
[…]
