### 2026-02-18 476fb36
make attention actually be temporal
