### 2026-02-19 tue-bmd/zea#260 @tristan-deep
Review finished in https://github.com/openjournals/joss-reviews/issues/9881#issuecomment-3920901039

Thanks @benfrey !!!
