### 2026-02-19 337e966
DO NOT MERGE THIS COMMIT; sloppy parallelization of Lucas Kanade. Slightly better than the original, but should be rewritten before merging.
