+++ zea/tracking/lucas_kanade.py
@@ -107,17 +107,7 @@ class LucasKanadeTracker(BaseTracker):
             prev_img = prev_pyr[level]
             next_img = next_pyr[level]
 
-            # Track each point
-            new_flows = []
-
-            for i in range(n_points):
-                pt = curr_points[i]
-                flow_guess = flows[i]
-
-                flow = self._track_point(prev_img, next_img, pt, flow_guess)
-                new_flows.append(flow)
-
-            flows = ops.stack(new_flows)
+            flows = self._track_points_batched(prev_img, next_img, curr_points, flows)
 
             # Scale for next level (if not at finest)
             if level < n_levels - 1:
@@ -172,122 +162,146 @@ class LucasKanadeTracker(BaseTracker):
             pyramid.append(downsampled)
         return pyramid[::-1]
 
-    def _track_point(
+    def _track_points_batched(
         self,
         prev_img,
         next_img,
-        point,
-        flow_guess,
+        points,
+        flow_guesses,
     ):
-        """Track a single point using iterative Lucas-Kanade."""
-        # Extract template window
-        template, valid_template = self._extract_window(prev_img, point)
-        if not valid_template:
-            return flow_guess
+        """Track all points using batched iterative Lucas-Kanade.
 
-        # Compute template gradients (Sobel) - returns tensors
-        gradients = self._sobel_gradients(template)
+        Args:
+            prev_img: Previous image/volume
+            next_img: Next image/volume
+            points: All points to track, shape (N, ndim)
+            flow_guesses: Initial flow estimates, shape (N, ndim)
 
-        # Flatten gradients for 2D or 3D
+        Returns:
+            flows: Refined flows for all points, shape (N, ndim)
+        """
+        n_points = int(points.shape[0])
+
+        templates, valid_templates = self._extract_windows_batched(prev_img, points)
+        gradients = self._sobel_gradients_batched(templates)
+
+        # Pre-compute structure tensors for all points
         if self.ndim == 2:
-            Iy, Ix = gradients
-            Ix_flat = ops.reshape(Ix, [-1])
-            Iy_flat = ops.reshape(Iy, [-1])
+            Iy, Ix = gradients  # (N, H, W)
+            Ix_flat = ops.reshape(Ix, (n_points, -1))  # (N, H*W)
+            Iy_flat = ops.reshape(Iy, (n_points, -1))
 
-            # Structure tensor 2D components
-            IxIx = ops.sum(Ix_flat * Ix_flat)
-            IxIy = ops.sum(Ix_flat * Iy_flat)
-            IyIy = ops.sum(Iy_flat * Iy_flat)
+            IxIx = ops.sum(Ix_flat * Ix_flat, axis=1)  # (N,)
+            IxIy = ops.sum(Ix_flat * Iy_flat, axis=1)
+            IyIy = ops.sum(Iy_flat * Iy_flat, axis=1)
+
+            structure = ops.stack(
+                [
+                    ops.stack([IxIx, IxIy], axis=1),
+                    ops.stack([IxIy, IyIy], axis=1),
+                ],
+                axis=1,
+            )
+            eye = ops.eye(2, dtype=structure.dtype)
+            structure = structure + ops.expand_dims(eye, 0) * 1e-5
 
         else:  # 3D
-            Iz, Iy, Ix = gradients
-            Ix_flat = ops.reshape(Ix, [-1])
-            Iy_flat = ops.reshape(Iy, [-1])
[…]
