### 2026-02-20 69039d3
frame time embeddings
