+++ debug_mach_beamform.py
@@ -1,0 +1,351 @@
+"""Standalone MachBeamform debug script."""
+
+import time
+
+import matplotlib.pyplot as plt
+import numpy as np
+
+
+def _block_until_ready(value):
+    if hasattr(value, "block_until_ready"):
+        value.block_until_ready()
+
+
+def _time_call(label, fn, iterations=10, warmup=2):
+    result = None
+    for _ in range(warmup):
+        result = fn()
+    _block_until_ready(result)
+
+    start = time.perf_counter()
+    for _ in range(iterations):
+        result = fn()
+    _block_until_ready(result)
+    elapsed = time.perf_counter() - start
+    print(f"{label} total time ({iterations} iters): {elapsed:.3f} s")
+    print(f"{label} per iter: {elapsed / iterations * 1e3:.3f} ms")
+    return result
+
+
+def _nan_stats(label, array):
+    arr = np.asarray(array)
+    nan_count = np.isnan(arr).sum()
+    total = arr.size
+    finite_min = np.nanmin(arr) if nan_count < total else np.nan
+    finite_max = np.nanmax(arr) if nan_count < total else np.nan
+    print(f"{label} nan count: {nan_count}/{total}, min: {finite_min:.3e}, max: {finite_max:.3e}")
+
+
+def _save_image(path, image, scan, title, dynamic_range=(-60, 0)):
+    from zea.display import to_8bit
+
+    image_8bit = to_8bit(image, dynamic_range=dynamic_range, pillow=False)
+    plt.figure(figsize=(6, 5))
+    plt.imshow(image_8bit, cmap="gray", extent=scan.extent)
+    plt.title(title)
+    plt.xlabel("X (m)")
+    plt.ylabel("Z (m)")
+    plt.tight_layout()
+    plt.savefig(path, dpi=200)
+    plt.close()
+
+
+def _zea_beamform_reference(
+    data_frames,
+    flatgrid,
+    probe_geometry,
+    sampling_frequency,
+    sound_speed,
+    f_number,
+    demodulation_frequency,
+):
+    import keras
+
+    from zea.beamform.beamformer import tof_correction
+    from zea.ops import DelayAndSum
+
+    n_tx, n_el = data_frames.shape[1], data_frames.shape[3]
+
+    t0_delays = np.zeros((n_tx, n_el), dtype="float32")
+    tx_apodizations = np.ones((n_tx, n_el), dtype="float32")
+    polar_angles = np.zeros((n_tx,), dtype="float32")
+    focus_distances = np.ones((n_tx,), dtype="float32")
+    t_peak = np.zeros((1,), dtype="float32")
+    tx_waveform_indices = np.zeros((n_tx,), dtype="int32")
+    transmit_origins = np.zeros((n_tx, 3), dtype="float32")
+    initial_times = np.zeros((n_tx,), dtype="float32")
+
+    flatgrid_t = keras.ops.convert_to_tensor(flatgrid)
+    probe_geometry_t = keras.ops.convert_to_tensor(probe_geometry)
+    t0_delays_t = keras.ops.convert_to_tensor(t0_delays)
+    tx_apodizations_t = keras.ops.convert_to_tensor(tx_apodizations)
+    polar_angles_t = keras.ops.convert_to_tensor(polar_angles)
+    focus_distances_t = keras.ops.convert_to_tensor(focus_distances)
+    t_peak_t = keras.ops.convert_to_tensor(t_peak)
+    tx_waveform_indices_t = keras.ops.convert_to_tensor(tx_waveform_indices)
+    transmit_origins_t = keras.ops.convert_to_tensor(transmit_origins)
+    initial_times_t = keras.ops.convert_to_tensor(initial_times)
+
[…]
