+++ mach/plane_wave_compound.py
@@ -1,0 +1,267 @@
+"""
+Plane Wave Compounding
+===================================================
+
+This example demonstrates plane wave compounding of data from the
+PICMUS challenge (Plane-wave Imaging Challenge in Medical Ultrasound). PICMUS
+provides standardized datasets for evaluating plane wave imaging algorithms.
+
+**Example overview:**
+
+- Load ultrasound data from UFF format files
+- Beamform data from multiple plane-wave transmits
+- Coherently compound the results
+- Visualize the results
+
+Attribution:
+
+- Example inspired by vbeam examples (https://github.com/magnusdk/vbeam/)
+- Dataset from PICMUS challenge (https://www.creatis.insa-lyon.fr/Challenge/IEEE_IUS_2016/)
+"""
+
+# %%
+# .. note::
+#    To run this example, we recommend installing mach with all optional dependencies:
+#
+#    .. code-block:: bash
+#
+#       pip install mach-beamform[all]
+#
+#    This ensures all dependencies for the examples are available.
+
+# %%
+# Import Required Libraries
+# -------------------------
+# Let's start by importing the necessary libraries.
+
+import hashlib
+import time
+
+import matplotlib.pyplot as plt
+import numpy as np
+
+# Import mach modules
+from mach import experimental
+from mach._vis import db_zero
+from mach.io.uff import create_beamforming_setup
+from mach.io.utils import cached_download
+
+try:
+    from pyuff_ustb import Uff
+except ImportError as err:
+    raise ImportError(
+        "⚠️  pyuff_ustb is required for UFF data loading. Install with: pip install pyuff-ustb"
+    ) from err
+
+# Convenience constants
+MM_PER_METER = 1000
+
+# %%
+# Download PICMUS Challenge Dataset
+# ---------------------------------
+#
+# The PICMUS challenge contains multiple datasets with multi-angle plane-wave
+# transmits. This example uses the resolution dataset, which features:
+#
+# - **128-element linear array** at 5.2 MHz center frequency
+# - **75 plane wave transmits** at angles from -16° to +16°
+# - **Point targets and cysts** for resolution and contrast assessment
+
+print("📂 Downloading PICMUS challenge dataset...")
+
+# Download the UFF data file (cached locally after first download)
+url = "http://www.ustb.no/datasets/PICMUS_experiment_resolution_distortion.uff"
+uff_path = cached_download(
+    url,
+    expected_size=145_518_524,
+    expected_hash="c93af0781daeebf771e53a42a629d4f311407410166ef2f1d227e9d2a1b8c641",
+    digest=hashlib.sha256,
+    filename="PICMUS_experiment_resolution_distortion.uff",
+)
+
+print(f"✓ Dataset downloaded to: {uff_path}")
+print(f"  File size: {uff_path.stat().st_size / 1e6:.1f} MB")
+
+# %%
+# Load and Inspect PICMUS Data
+# ----------------------------
+#
+# UFF files contain structured ultrasound data including channel data (RF signals),
+# probe geometry, and scan parameters. Let's examine the PICMUS dataset structure.
+
+print("\n📋 Loading PICMUS data structure...")
+
+# Open UFF file and extract components
+uff_file = Uff(str(uff_path))
+channel_data = uff_file.read("/channel_data")
+scan = uff_file.read("/scan")
+
+# Display challenge dataset information
+print("\n📊 PICMUS Challenge Dataset:")
[…]
