+++ debug_mach_beamform.py
@@ -1,5 +1,6 @@
 """Standalone MachBeamform debug script."""
 
+import subprocess
 import time
 
 import matplotlib.pyplot as plt
@@ -11,7 +12,7 @@ def _block_until_ready(value):
         value.block_until_ready()
 
 
-def _time_call(label, fn, iterations=10, warmup=2):
+def _time_call(label, fn, iterations=100, warmup=2):
     result = None
     for _ in range(warmup):
         result = fn()
@@ -22,9 +23,10 @@ def _time_call(label, fn, iterations=10, warmup=2):
         result = fn()
     _block_until_ready(result)
     elapsed = time.perf_counter() - start
+    per_iter_s = elapsed / iterations
     print(f"{label} total time ({iterations} iters): {elapsed:.3f} s")
-    print(f"{label} per iter: {elapsed / iterations * 1e3:.3f} ms")
-    return result
+    print(f"{label} per iter: {per_iter_s * 1e3:.3f} ms")
+    return result, per_iter_s
 
 
 def _nan_stats(label, array):
@@ -50,67 +52,78 @@ def _save_image(path, image, scan, title, dynamic_range=(-60, 0)):
     plt.close()
 
 
-def _zea_beamform_reference(
-    data_frames,
-    flatgrid,
-    probe_geometry,
-    sampling_frequency,
-    sound_speed,
-    f_number,
-    demodulation_frequency,
-):
-    import keras
-
-    from zea.beamform.beamformer import tof_correction
-    from zea.ops import DelayAndSum
-
-    n_tx, n_el = data_frames.shape[1], data_frames.shape[3]
-
-    t0_delays = np.zeros((n_tx, n_el), dtype="float32")
-    tx_apodizations = np.ones((n_tx, n_el), dtype="float32")
-    polar_angles = np.zeros((n_tx,), dtype="float32")
-    focus_distances = np.ones((n_tx,), dtype="float32")
-    t_peak = np.zeros((1,), dtype="float32")
-    tx_waveform_indices = np.zeros((n_tx,), dtype="int32")
-    transmit_origins = np.zeros((n_tx, 3), dtype="float32")
-    initial_times = np.zeros((n_tx,), dtype="float32")
-
-    flatgrid_t = keras.ops.convert_to_tensor(flatgrid)
-    probe_geometry_t = keras.ops.convert_to_tensor(probe_geometry)
-    t0_delays_t = keras.ops.convert_to_tensor(t0_delays)
-    tx_apodizations_t = keras.ops.convert_to_tensor(tx_apodizations)
-    polar_angles_t = keras.ops.convert_to_tensor(polar_angles)
-    focus_distances_t = keras.ops.convert_to_tensor(focus_distances)
-    t_peak_t = keras.ops.convert_to_tensor(t_peak)
-    tx_waveform_indices_t = keras.ops.convert_to_tensor(tx_waveform_indices)
-    transmit_origins_t = keras.ops.convert_to_tensor(transmit_origins)
-    initial_times_t = keras.ops.convert_to_tensor(initial_times)
-
-    das = DelayAndSum(with_batch_dim=False)
-    results = []
-    for frame in range(data_frames.shape[0]):
-        frame_data = keras.ops.convert_to_tensor(data_frames[frame])
-        aligned = tof_correction(
-            frame_data,
-            flatgrid=flatgrid_t,
-            t0_delays=t0_delays_t,
-            tx_apodizations=tx_apodizations_t,
-            sound_speed=sound_speed,
-            probe_geometry=probe_geometry_t,
-            initial_times=initial_times_t,
-            sampling_frequency=sampling_frequency,
-            demodulation_frequency=demodulation_frequency,
-            f_number=f_number,
-            polar_angles=polar_angles_t,
[…]
