+++ zea/ops/__init__.py
@@ -133,6 +133,7 @@ from .ultrasound import (
     LeeFilter,
     LogCompress,
     LowPassFilterIQ,
+    MachBeamform,
     PfieldWeighting,
     ReshapeGrid,
     ScanConvert,
@@ -179,6 +180,7 @@ __all__ = [
     "LeeFilter",
     "LogCompress",
     "LowPassFilterIQ",
+    "MachBeamform",
     "PfieldWeighting",
     "ReshapeGrid",
     "ScanConvert",

[…]
