### 2026-02-23 58b88c8
Enhance Mach beamforming integration: add Mach API example and update comparison image functionality
