+++ debug_mach_beamform.py
@@ -96,18 +96,22 @@ def _save_comparison_image(
     path,
     zea_image,
     mach_image,
+    mach_api_image,
     scan,
     zea_title,
     mach_title,
+    mach_api_title,
     full_title,
+    stats_text,
     dynamic_range=(-60, 0),
 ):
     from zea.display import to_8bit
 
     zea_8bit = to_8bit(zea_image, dynamic_range=dynamic_range, pillow=False)
     mach_8bit = to_8bit(mach_image, dynamic_range=dynamic_range, pillow=False)
+    mach_api_8bit = to_8bit(mach_api_image, dynamic_range=dynamic_range, pillow=False)
 
-    fig, axes = plt.subplots(1, 2, figsize=(12, 5), dpi=200)
+    fig, axes = plt.subplots(1, 3, figsize=(18, 5), dpi=200)
     extent = scan.extent
 
     axes[0].imshow(zea_8bit, cmap="gray", extent=extent)
@@ -120,7 +124,13 @@ def _save_comparison_image(
     axes[1].set_xlabel("X (m)")
     axes[1].set_ylabel("Z (m)")
 
+    axes[2].imshow(mach_api_8bit, cmap="gray", extent=extent)
+    axes[2].set_title(mach_api_title)
+    axes[2].set_xlabel("X (m)")
+    axes[2].set_ylabel("Z (m)")
+
     fig.suptitle(full_title)
+    fig.text(0.5, 0.01, stats_text, ha="center", va="bottom")
     fig.tight_layout()
     fig.savefig(path, dpi=200)
     plt.close(fig)
@@ -171,7 +181,7 @@ def run_pipeline_zea_example(data_frame, scan, probe):
     result, per_iter_s = _time_call(
         "Zea pipeline",
         lambda: pipeline(**inputs)["data"],
-        iterations=10,
+        iterations=100,
     )
     result = keras.ops.convert_to_numpy(result)
     _nan_stats("Zea pipeline output", result)
@@ -251,20 +261,20 @@ def run_pipeline_mach_example(data_frame, scan, probe):
     _nan_stats("Mach beamformed (flat)", beamformed_np)
 
     result, per_iter_s = _time_call(
-        "Mach pipeline",
+        "mach+zea",
         lambda: pipeline(**inputs)["data"],
-        iterations=10,
+        iterations=100,
     )
     result = keras.ops.convert_to_numpy(result)
-    _nan_stats("Mach pipeline output", result)
-    print("Mach pipeline input shape:", data_frame.shape)
-    print("Mach pipeline tx arrivals shape:", tx_wave_arrivals_s.shape)
-    print("Mach pipeline output shape:", result.shape)
-    print("Mach pipeline output dtype:", result.dtype)
-    _save_image("mach_pipeline_output.png", result, scan, "Mach pipeline output")
+    _nan_stats("mach+zea output", result)
+    print("mach+zea input shape:", data_frame.shape)
+    print("mach+zea tx arrivals shape:", tx_wave_arrivals_s.shape)
+    print("mach+zea output shape:", result.shape)
+    print("mach+zea output dtype:", result.dtype)
+    _save_image("mach_pipeline_output.png", result, scan, "mach+zea output")
     points_per_second = _calculate_points_per_second(data_frame, result, per_iter_s)
     title = _format_pipeline_title(
-        "Mach pipeline",
+        "mach+zea",
         data_frame.shape,
         result.shape,
         per_iter_s,
@@ -273,21 +283,125 @@ def run_pipeline_mach_example(data_frame, scan, probe):
     return result, title
 
 
+def run_mach_api_example(data_frame, scan, tx_wave_arrivals_s):
+    import mach
+
+    try:
[…]
