### 2026-02-23 c722bbc
debug script refactor
