+++ debug_mach_beamform.py
@@ -6,6 +6,10 @@ import time
 import matplotlib.pyplot as plt
 import numpy as np
 
+# =========================
+# Generic helpers
+# =========================
+
 
 def _block_until_ready(value):
     if hasattr(value, "block_until_ready"):
@@ -136,6 +140,30 @@ def _save_comparison_image(
     plt.close(fig)
 
 
+def _build_tx_wave_arrivals(scan):
+    import mach
+
+    origin = np.array([0.0, 0.0, 0.0], dtype="float32")
+    if scan.polar_angles is None:
+        directions = np.array([[0.0, 0.0, 1.0]], dtype="float32")
+    else:
+        angles = np.asarray(scan.polar_angles, dtype="float32")
+        directions = np.stack(
+            [np.array([np.sin(angle), 0.0, np.cos(angle)], dtype="float32") for angle in angles],
+            axis=0,
+        )
+
+    tx_wave_arrivals_s = np.stack(
+        [
+            mach.wavefront.plane(origin_m=origin, points_m=scan.flatgrid, direction=direction)
+            / scan.sound_speed
+            for direction in directions
+        ],
+        axis=0,
+    )
+    return tx_wave_arrivals_s
+
+
 def _load_picmus_sample():
     from zea.data import load_file
 
@@ -156,6 +184,11 @@ def _load_picmus_sample():
     scan.set_transmits(3)
     data_frame = data[0][scan.selected_transmits]
     return data_frame, scan, probe
+
+
+# =========================
+# Zea pipeline
+# =========================
 
 
 def run_pipeline_zea_example(data_frame, scan, probe):
@@ -200,30 +233,17 @@ def run_pipeline_zea_example(data_frame, scan, probe):
     return result, title
 
 
+# =========================
+# Zea + Mach pipeline
+# =========================
+
+
 def run_pipeline_mach_example(data_frame, scan, probe):
     import keras
 
-    import mach
     from zea.ops import EnvelopeDetect, LogCompress, MachBeamform, Normalize, Pipeline, ReshapeGrid
 
-    origin = np.array([0.0, 0.0, 0.0], dtype="float32")
-    if scan.polar_angles is None:
-        directions = np.array([[0.0, 0.0, 1.0]], dtype="float32")
-    else:
-        angles = np.asarray(scan.polar_angles, dtype="float32")
-        directions = np.stack(
-            [np.array([np.sin(angle), 0.0, np.cos(angle)], dtype="float32") for angle in angles],
-            axis=0,
-        )
-
-    tx_wave_arrivals_s = np.stack(
-        [
-            mach.wavefront.plane(origin_m=origin, points_m=scan.flatgrid, direction=direction)
-            / scan.sound_speed
-            for direction in directions
-        ],
-        axis=0,
-    )
+    tx_wave_arrivals_s = _build_tx_wave_arrivals(scan)
 
     pipeline = Pipeline(
         [
@@ -281,6 +301,11 @@ def run_pipeline_mach_example(data_frame, scan, probe):
         points_per_second,
     )
     return result, title
+
+
+# =========================
+# Mach API only
+# =========================
 
 
 def run_mach_api_example(data_frame, scan, tx_wave_arrivals_s):
@@ -353,42 +378,39 @@ def run_mach_api_example(data_frame, scan, tx_wave_arrivals_s):
     return mach_api_image, title
 
 
-if __name__ == "__main__":
-    data_frame, scan, probe = _load_picmus_sample()
[…]
