+++ debug_mach_beamform.py
@@ -2,6 +2,7 @@
 
 import subprocess
 import time
+import gc
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -68,6 +69,64 @@ def _get_gpu_info():
         return lines[0] if lines else "Unknown GPU"
     except Exception:
         return "Unknown GPU"
+
+
+def _query_vram():
+    try:
+        result = subprocess.run(
+            [
+                "nvidia-smi",
+                "--query-gpu=memory.used,memory.free,memory.total",
+                "--format=csv,noheader,nounits",
+            ],
+            capture_output=True,
+            text=True,
+            check=True,
+        )
+        lines = [line.strip() for line in result.stdout.splitlines() if line.strip()]
+        stats = []
+        for line in lines:
+            used, free, total = [value.strip() for value in line.split(",")]
+            stats.append(
+                {
+                    "used": int(used),
+                    "free": int(free),
+                    "total": int(total),
+                }
+            )
+        return stats
+    except Exception:
+        return None
+
+
+def _vram_checkpoint(label):
+    stats = _query_vram()
+    if not stats:
+        print(f"VRAM {label}: unavailable")
+        return
+    parts = []
+    for idx, gpu in enumerate(stats):
+        parts.append(
+            f"GPU{idx} {gpu['used']}MB/{gpu['total']}MB (free {gpu['free']}MB)"
+        )
+    print(f"VRAM {label}: " + " | ".join(parts))
+
+
+def _flush_vram():
+    try:
+        import cupy as cp
+
+        cp.get_default_memory_pool().free_all_blocks()
+        cp.get_default_pinned_memory_pool().free_all_blocks()
+    except Exception:
+        pass
+    try:
+        import jax
+
+        jax.clear_caches()
+    except Exception:
+        pass
+    gc.collect()
 
 
 def _format_pipeline_title(label, input_shape, output_shape, per_iter_s, points_per_second):
@@ -161,7 +220,7 @@ def _build_tx_wave_arrivals(scan):
         ],
         axis=0,
     )
-    return tx_wave_arrivals_s
+    return tx_wave_arrivals_s.T
 
 
 def _load_picmus_sample():
@@ -181,7 +240,7 @@ def _load_picmus_sample():
     scan.xlims = probe.xlims
     scan.zlims = (0.0, 0.06)
 
-    scan.set_transmits(3)
+    scan.set_transmits(2)
     data_frame = data[0][scan.selected_transmits]
     return data_frame, scan, probe
 
@@ -210,6 +269,12 @@ def run_pipeline_zea_example(data_frame, scan, probe):
     inputs = pipeline.prepare_parameters(probe=probe, scan=scan)
     inputs["data"] = data_frame
     inputs["demodulation_frequency"] = scan.demodulation_frequency
+
+    _flush_vram()
+    _vram_checkpoint("Zea pipeline (before)")
+    one_pass = pipeline(**inputs)["data"]
+    _block_until_ready(one_pass)
+    _vram_checkpoint("Zea pipeline (after)")
 
     result, per_iter_s = _time_call(
         "Zea pipeline",
@@ -243,8 +308,6 @@ def run_pipeline_mach_example(data_frame, scan, probe):
 
     from zea.ops import EnvelopeDetect, LogCompress, MachBeamform, Normalize, Pipeline, ReshapeGrid
 
-    tx_wave_arrivals_s = _build_tx_wave_arrivals(scan)
-
     pipeline = Pipeline(
         [
             MachBeamform(),
[…]
