+++ zea/ops/ultrasound.py
@@ -1,3 +1,4 @@
+import inspect
 import uuid
 from typing import Tuple
 
@@ -6,7 +7,8 @@ import numpy as np
 from keras import ops
 
 from zea import log
-from zea.beamform.beamformer import tof_correction
+from zea.backend import jit as backend_jit
+from zea.beamform.beamformer import calculate_delays, tof_correction
 from zea.display import scan_convert
 from zea.func.tensor import (
     apply_along_axis,
@@ -33,6 +35,23 @@ from zea.internal.registry import ops_registry
 from zea.ops.base import Filter, Operation
 from zea.simulator import simulate_rf
 from zea.utils import canonicalize_axis
+
+try:
+    from mach import experimental as mach_experimental
+except ImportError:
+    mach_experimental = None
+
+try:
+    import cupy as cp
+except ImportError:
+    cp = None
+
+try:
+    import jax
+    import jax.dlpack as jdlpack
+except ImportError:
+    jax = None
+    jdlpack = None
 
 
 @ops_registry("simulate_rf")
@@ -218,17 +237,30 @@ class MachBeamform(Operation):
         )
         self.interp_type = interp_type
         self.tukey_alpha = float(tukey_alpha)
+        if keras.backend.backend() == "jax":
+            self._calculate_delays = backend_jit(calculate_delays, static_argnums=(7, 8))
+        else:
+            self._calculate_delays = backend_jit(calculate_delays)
 
     def call(
         self,
         flatgrid,
         probe_geometry,
-        tx_wave_arrivals_s,
         sampling_frequency,
         sound_speed,
         f_number,
         demodulation_frequency=None,
         initial_times=None,
+        t0_delays=None,
+        tx_apodizations=None,
+        focus_distances=None,
+        polar_angles=None,
+        t_peak=None,
+        tx_waveform_indices=None,
+        transmit_origins=None,
+        apply_lens_correction=False,
+        lens_thickness=None,
+        lens_sound_speed=None,
         rx_start_s=None,
         **kwargs,
     ):
@@ -237,20 +269,31 @@ class MachBeamform(Operation):
         Args:
             flatgrid: Flattened grid points of shape (n_pix, 3).
             probe_geometry: Receive element positions of shape (n_el, 3).
-            tx_wave_arrivals_s: Transmit arrivals of shape (n_tx, n_pix) or (n_pix,).
             sampling_frequency: Sampling frequency in Hz.
             sound_speed: Speed of sound in m/s.
             f_number: F-number for dynamic aperture.
             demodulation_frequency: Center frequency in Hz for IQ data.
             initial_times: Optional per-transmit receive start times (n_tx,).
+            t0_delays: Transmit delays in seconds of shape (n_tx, n_el).
+            tx_apodizations: Transmit apodizations of shape (n_tx, n_el).
+            focus_distances: Focus distances of shape (n_tx,).
+            polar_angles: Polar angles of shape (n_tx,).
+            t_peak: Time of the transmit peak of shape (n_waveforms,).
+            tx_waveform_indices: Waveform indices of shape (n_tx,).
+            transmit_origins: Transmit origins of shape (n_tx, 3).
+            apply_lens_correction: Whether to apply lens correction to delays.
+            lens_thickness: Lens thickness in meters.
+            lens_sound_speed: Lens sound speed in m/s.
             rx_start_s: Optional scalar receive start time in seconds.
         """
-        try:
-            from mach import experimental
-        except ImportError as err:
+        if mach_experimental is None:
             raise ImportError(
                 "mach is required for MachBeamform. Install with: pip install mach-beamform"
[…]
