### 2026-02-23 4219334
new denoiser layer type
