### 2026-02-24 261882c
Move ops.where outside loop
