### 2026-02-24 36984ea
Added additional attributes
