### 2026-03-09 tue-bmd/zea#278 @tomelse
JOSS Review: issues with the first example script
https://github.com/openjournals/joss-reviews/issues/9881

Hello, I'm making a start on the JOSS review. I've installed zea, no problem, and started running the first example: https://zea.readthedocs.io/en/latest/notebooks/data/zea_data_example.html.

I run into an error where the first example requires tensorflow. I'm more than happy to install this, but I would just flag that in your installation instructions, it suggests that you can use any of the three major ML backends (https://zea.readthedocs.io/en/latest/installation.html).

I'd recommend either updating your example notebook to stipulate that tensorflow must also be installed, or indicating that this is required in the installation instructions.

Thanks :)
