### 2026-03-09 tue-bmd/zea#278 @tristan-deep
> Further issue with the first example script. The zea.Dataset class takes in a key as well as the file name.
>
> <img alt="Image" width="573" height="402" src="https://private-user-images.githubusercontent.com/56722586/560301265-871e80d1-af00-4b11-a6ba-b2c6feec14b7.png?jwt=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NzMwNjIyMjUsIm5iZiI6MTc3MzA2MTkyNSwicGF0aCI6Ii81NjcyMjU4Ni81NjAzMDEyNjUtODcxZTgwZDEtYWYwMC00YjExLWE2YmEtYjJjNmZlZWMxNGI3LnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNjAzMDklMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjYwMzA5VDEzMTIwNVomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPWJiODg1N2FhYjc3OTEzZGIwM2JlMGY4Y2JjZmE3NWZhOWYyMmQ2ZTBlNDcxODVmNDFiM2YzNDhjZmNlMTU3MzEmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0.q66w3ynAwHWXATExOmFuY6lL4V9YFJMidslrjDgDq1E">

This issue was resolved in #273, but the fix has not yet been included in the latest release. The example notebooks are currently aligned with the main branch, so installing from main should resolve the problem.

```bash
pip install git+https://github.com/tue-bmd/zea.git@main
```

More generally, I’m not entirely sure what the best practice is here. One option is to keep the notebooks aligned with the latest release, which avoids compatibility issues but may leave them slightly outdated. The alternative is to keep them in sync with main, which ensures they reflect the latest changes but can occasionally lead to incompatibilities with the most recent release.

EDIT:
Actually I think we can change the install line at the top of the notebook

```
%pip install zea
```

Into
```
%pip install git+https://github.com/tue-bmd/zea.git@main
```
