### 2026-03-09 tue-bmd/zea#278 @tomelse
Good point. I guess I am looking at the "latest" branch of read the docs but using the "stable" branch from PyPI. I think you're doing the right thing in that case. I can't remember if there is a way to make read the docs use the "stable" branch by default?

I think what you're doing at the moment is fine for now though. I've installed from git and that's working nwo
