### 2026-03-20 0ba3b8f
Clip after translate
