### 2026-03-20 cd3fd89
Support loading operations from module string
