### 2026-03-20 d1e2ef8
Fix * in docstring
