### 2026-03-20 0de4fd6
Fix append_to_yaml with multiple funcs
