+++ zea/data/process.py
@@ -1,0 +1,202 @@
+import argparse
+from pathlib import Path
+
+import keras
+import numpy as np
+from keras import ops
+
+import zea
+from zea.utils import FunctionTimer
+
+
+def get_parser() -> argparse.ArgumentParser:
+    parser = argparse.ArgumentParser(
+        description="Convert a zea dataset to processed videos using a zea pipeline "
+        + "defined in a config.yaml file.",
+    )
+    parser.add_argument(
+        "dataset",
+        type=str,
+        help="Path/URI to the zea dataset.",
+    )
+    parser.add_argument(
+        "save_dir",
+        type=Path,
+        help="Directory where output videos are written.",
+    )
+    parser.add_argument(
+        "--config",
+        type=str,
+        default=None,
+        help="Path to config.yaml. Defaults to <dataset>/config.yaml.",
+    )
+    parser.add_argument(
+        "--data_type",
+        type=str,
+        default="raw_data",
+        help="Data type to load from each file.",
+    )
+    parser.add_argument(
+        "--n_frames",
+        type=int,
+        default=None,
+        help="Number of frames to load (all frames when omitted).",
+    )
+    parser.add_argument(
+        "--save_as",
+        type=str,
+        default="gif",
+        help="Format to save output videos (e.g. 'gif' or 'mp4').",
+    )
+    parser.add_argument(
+        "--keep_keys",
+        nargs="+",
+        default=["maxval"],
+        help="Keys from the pipeline output to keep for the next iteration.",
+    )
+    parser.add_argument(
+        "--timings",
+        action="store_true",
+        help="Whether to time dataloader and pipeline operations and save timings to a yaml file.",
+    )
+    return parser
+
+
+def run_processing(
+    dataset_path: str,
+    config_path: str,
+    data_type: str,
+    n_frames: int | None,
+    save_dir: Path,
+    save_as: str = "gif",
+    keep_keys=("maxval",),
+    timings=False,
+) -> None:
+    def parse_metadata(metadata: list[str]) -> dict:
+        return zea.data.utils.json_loads(metadata[0])
+
+    def to_8bit(data, scan):
+        data = ops.convert_to_numpy(data)
+        data = np.nan_to_num(data, copy=False, nan=scan.dynamic_range[0])
+        data = zea.display.to_8bit(data, scan.dynamic_range)
+        return data
+
+    def save_video_frames(filestem: str, frames: list[np.ndarray], fps: float) -> None:
+        video = np.stack(frames, axis=0)
+        save_path = save_dir / f"{filestem}.{save_as}"
+        print("\n")
+        zea.io_lib.save_video(video, save_path, fps=fps)
+        print("\n")
+
+    # Load config and pipeline
+    config = zea.Config.from_path(config_path)
+    pipeline = zea.Pipeline.from_path(config_path, with_batch_dim=False)
+
+    # Create save directory if it doesn't exist
+    save_dir.mkdir(parents=True, exist_ok=True)
+
+    # Set up dataloader
+    dataloader = zea.Dataloader(
+        dataset_path,
+        key=data_type,
+        batch_size=1,
+        shuffle=False,
+        return_filename=True,
+        limit_n_frames=n_frames,
+        n_frames=1,
+        num_threads=4,
+        insert_frame_axis=False,
+        sort_files=True,
+    )
+    iterator = iter(dataloader)
+    total_batches = len(dataloader)
+
+    # Define functions to time if timings flag is set
+    get_data = lambda: next(iterator)
+    prepare_parameters = pipeline.prepare_parameters
+    pipeline_call = pipeline.__call__
+
+    if timings:
+        timer = FunctionTimer()
+        get_data = timer(get_data, name="dataloader")
[…]
