### 2026-03-23 c7963bf
() and (1,) are both okay
