### 2026-03-23 4a7e60d
Configure num_threads
