### 2026-03-23 f4d8920
Fix when no fps can be calculaed from file
