### 2026-03-26 60ac934
Assert shapes in metrics module
