### 2026-04-01 f9e85c3
Convert spec to python (just a start)
