+++ zea/data/data_format.yaml
@@ -1,119 +1,302 @@
+data:
+  raw_data:
+    description: The raw channel data of shape {shape}.
+    type: float32
+    required: true  # for zea this is not required.
+    shape: [n_frames, n_tx, n_el, n_ax, n_ch]
+    unit: "-"
+  image:
+    description: Reconstructed (log-compressed) image data and spatial extent metadata.
+    pixels:
+      description: The image pixels of shape {shape}.
+      type: uint8
+      shape: [n_frames, h, w, d]
+    extent:
+      description: The image extent in {unit} of shape {shape} ordered as (xmin, xmax, ymin, ymax, zmax, zmin).
+      type: float32
+      unit: m
+      shape: [n_frames, 6]
+  segmentation:
+    description: Segmentation data and spatial extent metadata.
+    pixels:
+      description: The segmentation pixels of shape {shape}.
+      type: uint8
+      shape: [n_frames, h, w, d]
+    extent:
+      description: The segmentation extent in {unit} of shape {shape} ordered as (xmin, xmax, ymin, ymax, zmax, zmin).
+      type: float32
+      unit: m
+      shape: [n_frames, 6]
+    labels:
+      description: The labels corresponding to the segmentation pixels, where each unique value in the pixels corresponds to a label in this list of shape {shape}.
+      type: string
+      shape: [n_labels]
+  sos_map:
+    description: Speed-of-sound map data and spatial extent metadata.
+    pixels:
+      description: The speed-of-sound map pixels in {unit} of shape {shape}.
+      type: float32
+      unit: m/s
+      shape: [n_frames, h, w, d]
+    extent:
+      description: The speed-of-sound map extent in {unit} of shape {shape} ordered as (xmin, xmax, ymin, ymax, zmax, zmin).
+      type: float32
+      unit: m
+      shape: [n_frames, 6]
+  strain:
+    description: Strain map data and spatial extent metadata.
+    pixels:
+      description: The strain pixels in {unit} of shape {shape}.
+      type: float32
+      unit: "%"
+      shape: [n_frames, h, w, d]
+    extent:
+      description: The strain extent in {unit} of shape {shape} ordered as (xmin, xmax, ymin, ymax, zmax, zmin).
+      type: float32
+      unit: m
+      shape: [n_frames, 6]
+  swe:
+    description: Shear-wave elastography data and spatial extent metadata.
+    pixels:
+      description: The shear-wave elastography pixels in {unit} of shape {shape}.
+      type: float32
+      unit: m/s
+      shape: [n_frames, h, w, d]
+    extent:
+      description: The shear-wave elastography extent in {unit} of shape {shape} ordered as (xmin, xmax, ymin, ymax, zmax, zmin).
+      type: float32
+      unit: m
+      shape: [n_frames, 6]
+  tissue_doppler:
+    description: Tissue Doppler data and spatial extent metadata.
+    pixels:
+      description: The tissue Doppler pixels in {unit} of shape {shape}.
+      type: float32
+      unit: m/s
+      shape: [n_frames, h, w, d]
+    extent:
+      description: The tissue Doppler extent in {unit} of shape {shape} ordered as (xmin, xmax, ymin, ymax, zmax, zmin).
+      type: float32
+      unit: m
+      shape: [n_frames, 6]
+
 scan:
   us_machine:
     description: The ultrasound machine used to acquire the data.
     type: string
-    unit: "-"
-    required: true
   probe_name:
     description: The name of the probe used to acquire the data.
     type: string
-    unit: "-"
-    required: true
   n_ax:
     description: The number of axial samples.
     type: int
-    unit: "-"
+    unit: "#"
     required: true
   n_el:
     description: The number of elements in the probe.
     type: int
[…]
