### 2026-04-02 4503c9f
Refactor into Map
