+++ zea/data/spec.py
@@ -7,7 +7,14 @@ from zea import log
 
 
 class Spec:
-    SCHEMA = {}
+    """Base class for data specifications with schema validation.
+
+    Subclasses should define a SCHEMA class variable that specifies the expected dtype and shape
+    for each field. The __post_init__ method will validate that the actual fields match the schema,
+    including checking that dimensions with the same name have consistent sizes across fields.
+    """
+
+    SCHEMA: dict
 
     def __post_init__(self):
         dim_to_fields = defaultdict(set)
@@ -16,13 +23,12 @@ class Spec:
         for field_name, field_info in self.SCHEMA.items():
             field_value = getattr(self, field_name)
             expected_dtype = field_info["dtype"]
-            expected_shape = field_info["shape"]
+            shape_spec = field_info["shape"]
 
-            # Track dimension names and sizes for consistency checks
-            for i, dim_name in enumerate(expected_shape):
-                if isinstance(dim_name, str):
-                    dim_to_fields[dim_name].add(field_name)
-                    dim_to_sizes[dim_name].add(field_value.shape[i])
+            if shape_spec and isinstance(shape_spec[0], tuple):
+                expected_shapes = shape_spec
+            else:
+                expected_shapes = (shape_spec,)
 
             # Check dtype
             try:
@@ -43,22 +49,31 @@ class Spec:
                         f"{field_name} must be of type {expected_dtype}, got {field_value.dtype}"
                     )
 
-            # Check ndims
-            if len(field_value.shape) != len(expected_shape):
+            matched_shape = None
+            for expected_shape in expected_shapes:
+                if len(field_value.shape) != len(expected_shape):
+                    continue
+
+                for dim_size, expected_dim in zip(field_value.shape, expected_shape):
+                    if isinstance(expected_dim, str):
+                        continue
+                    if dim_size != expected_dim:
+                        break
+                else:
+                    matched_shape = expected_shape
+                    break
+
+            if matched_shape is None:
+                allowed_shapes = ", ".join(str(shape) for shape in expected_shapes)
                 raise ValueError(
-                    f"{field_name} must have {len(expected_shape)} dimensions, "
-                    f"got {len(field_value.shape)} dimensions with shape {field_value.shape}"
+                    f"{field_name} has shape {field_value.shape}, expected one of: {allowed_shapes}"
                 )
 
-            # Check static shape dimensions
-            for dim_size, expected_dim in zip(field_value.shape, expected_shape):
-                if isinstance(expected_dim, str):
-                    continue  # skip literal dimensions
-                if dim_size != expected_dim:
-                    raise ValueError(
-                        f"{field_name} dimension size mismatch: expected {expected_dim}, "
-                        f"got {dim_size} with shape {field_value.shape}"
-                    )
+            # Track dimension names and sizes for consistency checks
+            for i, dim_name in enumerate(matched_shape):
+                if isinstance(dim_name, str):
+                    dim_to_fields[dim_name].add(field_name)
+                    dim_to_sizes[dim_name].add(field_value.shape[i])
 
         # Check that dimensions with the same name have consistent sizes across fields
         for dim_name, sizes in dim_to_sizes.items():
@@ -71,26 +86,22 @@ class Spec:
 
 
 @dataclass
-class Segmentation(Spec):
-    """Segmentation data and spatial extent metadata.
+class Map(Spec):
+    """Map data and spatial extent metadata.
 
     Args:
[…]
