### 2026-04-02 08bc02c
Allow dict initialization of dataset
