### 2026-04-02 2b0918c
Add full spec
