+++ zea/data/spec.py
@@ -1,37 +1,50 @@
 from collections import defaultdict
-from dataclasses import dataclass
-from typing import List
+from dataclasses import dataclass, field
+from typing import Any, List
 
 import numpy as np
 
 from zea import log
 
 
-def check_dtype(value: np.ndarray, expected_dtype: type) -> None:
-    """Check if the dtype of a numpy array matches the expected dtype,
-    allowing for compatible types."""
+def check_dtype(value: Any, expected_dtype: type) -> None:
+    """Check if the dtype of a value matches the expected dtype,
+    allowing for compatible types.
+
+    Works for numpy arrays and scalar values.
+    """
     try:
         expected_np_dtype = np.dtype(expected_dtype)
         is_numpy_dtype = True
     except TypeError:
         is_numpy_dtype = False
 
+    value_dtype = value.dtype if isinstance(value, np.ndarray) else np.asarray(value).dtype
+
     if is_numpy_dtype:
-        if not np.issubdtype(value.dtype, expected_np_dtype):
+        if not np.issubdtype(value_dtype, expected_np_dtype):
             raise TypeError(
-                f"Expected dtype compatible with {expected_np_dtype}, got {value.dtype}"
+                f"Expected dtype compatible with {expected_np_dtype}, got {value_dtype}"
             )
     else:
-        if value.dtype != expected_dtype:
-            raise TypeError(f"Expected type {expected_dtype}, got {value.dtype}")
+        if value_dtype != expected_dtype:
+            raise TypeError(f"Expected type {expected_dtype}, got {value_dtype}")
 
 
-def match_shape(value: np.ndarray, expected_shape: tuple) -> bool:
-    """Check if the shape of a numpy array matches the expected shape specification."""
-    if len(value.shape) != len(expected_shape):
+def value_shape(value: Any) -> tuple:
+    """Return the shape tuple for numpy arrays and scalar values."""
+    if isinstance(value, np.ndarray):
+        return value.shape
+    return ()
+
+
+def match_shape(value: Any, expected_shape: tuple) -> bool:
+    """Check if the shape of a value matches the expected shape specification."""
+    shape = value_shape(value)
+    if len(shape) != len(expected_shape):
         return False
 
-    for dim_size, expected_dim in zip(value.shape, expected_shape):
+    for dim_size, expected_dim in zip(shape, expected_shape):
         if isinstance(expected_dim, str):
             continue
         if dim_size != expected_dim:
@@ -40,7 +53,7 @@ def match_shape(value: np.ndarray, expected_shape: tuple) -> bool:
     return True
 
 
-def find_matched_shape(value: np.ndarray, expected_shapes: List[tuple]) -> tuple | None:
+def find_matched_shape(value: Any, expected_shapes: List[tuple]) -> tuple | None:
     """Find the first expected shape specification that matches the shape of the value."""
     for expected_shape in expected_shapes:
         if match_shape(value, expected_shape):
@@ -61,9 +74,29 @@ class Spec:
     def __post_init__(self):
         dim_to_fields = defaultdict(set)
         dim_to_sizes = defaultdict(set)
+        dim_to_count_fields = {}
 
         for field_name, field_info in self.SCHEMA.items():
             field_value = getattr(self, field_name)
+            required = field_info.get("required", True)
+
+            if field_value is None:
+                if required:
+                    raise ValueError(f"Missing required field '{field_name}'")
+                continue
+
+            nested_spec = field_info.get("spec")
+            if nested_spec is not None:
[…]
