+++ tests/data/test_spec.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from zea.data.spec import Segmentation
+from zea.data.spec import Dataset, Scan, Segmentation
 
 
 def test_segmentation_spec():
@@ -20,3 +20,126 @@ def test_segmentation_spec():
         ValueError, match="Segmentation pixels contain values that do not correspond to any label"
     ):
         Segmentation(pixels=pixels_invalid, labels=labels, extent=extent)
+
+
+def _scan_minimal(n_frames: int = 3, n_tx: int = 2, n_el: int = 4, n_ax: int = 8, n_ch: int = 1):
+    return {
+        "n_ax": n_ax,
+        "n_el": n_el,
+        "n_tx": n_tx,
+        "n_ch": n_ch,
+        "n_frames": n_frames,
+        "probe_geometry": np.zeros((n_el, 3), dtype=np.float32),
+        "sampling_frequency": np.float32(30e6),
+        "center_frequency": np.linspace(5e6, 6e6, n_tx, dtype=np.float32),
+        "demodulation_frequency": np.linspace(5e6, 6e6, n_tx, dtype=np.float32),
+        "initial_times": np.zeros((n_tx,), dtype=np.float32),
+        "t0_delays": np.zeros((n_tx, n_el), dtype=np.float32),
+        "tx_apodizations": np.ones((n_tx, n_el), dtype=np.float32),
+        "focus_distances": np.zeros((n_tx,), dtype=np.float32),
+        "transmit_origins": np.zeros((n_tx, 3), dtype=np.float32),
+        "polar_angles": np.zeros((n_tx,), dtype=np.float32),
+        "azimuth_angles": np.zeros((n_tx,), dtype=np.float32),
+        "time_to_next_transmit": np.ones((n_frames, n_tx), dtype=np.float32),
+    }
+
+
+def test_dataset_full_new_spec_from_dicts():
+    n_frames, n_tx, n_el, n_ax, n_ch = 3, 2, 4, 8, 1
+
+    dataset = Dataset(
+        data={
+            "raw_data": np.zeros((n_frames, n_tx, n_el, n_ax, n_ch), dtype=np.float32),
+            "image": {
+                "pixels": np.zeros((n_frames, 16, 12, 1), dtype=np.uint8),
+                "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
+            },
+            "segmentation": {
+                "pixels": np.zeros((n_frames, 16, 12, 1), dtype=np.uint8),
+                "labels": np.array(["background", "tissue"], dtype=np.str_),
+                "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
+            },
+            "sos_map": {
+                "pixels": np.full((n_frames, 16, 12, 1), 1540.0, dtype=np.float32),
+                "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
+            },
+            "strain": {
+                "pixels": np.zeros((n_frames, 16, 12, 1), dtype=np.float32),
+                "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
+            },
+            "swe": {
+                "pixels": np.zeros((n_frames, 16, 12, 1), dtype=np.float32),
+                "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
+            },
+            "tissue_doppler": {
+                "pixels": np.zeros((n_frames, 16, 12, 1), dtype=np.float32),
+                "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
+            },
+        },
[…]
