+++ tests/data/test_spec.py
@@ -119,7 +119,7 @@ def test_scan_requires_required_fields():
     scan = _scan_minimal()
     scan.pop("n_tx")
 
-    with pytest.raises(ValueError, match="Missing required field 'n_tx'"):
+    with pytest.raises(TypeError, match="missing 1 required positional argument: 'n_tx'"):
         Scan(**scan)
 
 
