+++ zea/data/spec.py
@@ -1,5 +1,5 @@
 from collections import defaultdict
-from dataclasses import dataclass, field
+from dataclasses import MISSING, dataclass, field, fields
 from typing import Any, List
 
 import numpy as np
@@ -75,13 +75,19 @@ class Spec:
         dim_to_fields = defaultdict(set)
         dim_to_sizes = defaultdict(set)
         dim_to_count_fields = {}
+        dataclass_fields = {f.name: f for f in fields(self)}
 
         for field_name, field_info in self.SCHEMA.items():
             field_value = getattr(self, field_name)
-            required = field_info.get("required", True)
+            field_def = dataclass_fields.get(field_name)
+            is_optional = False
+            if field_def is not None:
+                is_optional = (
+                    field_def.default is not MISSING or field_def.default_factory is not MISSING
+                )
 
             if field_value is None:
-                if required:
+                if not is_optional:
                     raise ValueError(f"Missing required field '{field_name}'")
                 continue
 
@@ -327,14 +333,13 @@ class Data(Spec):
         "raw_data": {
             "dtype": np.float32,
             "shape": ("n_frames", "n_tx", "n_el", "n_ax", "n_ch"),
-            "required": False,
         },
-        "image": {"spec": Image, "required": False},
-        "segmentation": {"spec": Segmentation, "required": False},
-        "sos_map": {"spec": SosMap, "required": False},
-        "strain": {"spec": StrainMap, "required": False},
-        "swe": {"spec": SweMap, "required": False},
-        "tissue_doppler": {"spec": TissueDopplerMap, "required": False},
+        "image": {"spec": Image},
+        "segmentation": {"spec": Segmentation},
+        "sos_map": {"spec": SosMap},
+        "strain": {"spec": StrainMap},
+        "swe": {"spec": SweMap},
+        "tissue_doppler": {"spec": TissueDopplerMap},
     }
 
 
@@ -345,40 +350,37 @@ class Scan(Spec):
     All fields are aligned with the data format specification.
     """
 
+    n_ax: np.ndarray | int
+    n_el: np.ndarray | int
+    n_tx: np.ndarray | int
+    n_ch: np.ndarray | int
+    n_frames: np.ndarray | int
+    probe_geometry: np.ndarray
+    sampling_frequency: np.ndarray | float
+    center_frequency: np.ndarray | float
+    demodulation_frequency: np.ndarray | float
+    initial_times: np.ndarray
+    t0_delays: np.ndarray
+    tx_apodizations: np.ndarray
+    focus_distances: np.ndarray
+    transmit_origins: np.ndarray
+    polar_angles: np.ndarray
+    azimuth_angles: np.ndarray
+    time_to_next_transmit: np.ndarray
     us_machine: np.ndarray | str | None = None
     probe_name: np.ndarray | str | None = None
-    n_ax: np.ndarray | int | None = None
-    n_el: np.ndarray | int | None = None
-    n_tx: np.ndarray | int | None = None
-    n_ch: np.ndarray | int | None = None
-    n_frames: np.ndarray | int | None = None
     sound_speed: np.ndarray | float | None = None
-    probe_geometry: np.ndarray | None = None
-    sampling_frequency: np.ndarray | float | None = None
-    center_frequency: np.ndarray | float | None = None
-    demodulation_frequency: np.ndarray | float | None = None
-    initial_times: np.ndarray | None = None
-    t0_delays: np.ndarray | None = None
[…]
