### 2026-04-02 207d77e
Fix saving for string and scalar types
