### 2026-04-02 b99e0da
Simplify
