### 2026-04-02 71828f9
Remove shape params in scan
