### 2026-04-02 c12e33f
Fix test and refactor
