+++ tests/data/test_spec.py
@@ -23,13 +23,8 @@ def test_segmentation_spec():
         Segmentation(pixels=pixels_invalid, labels=labels, extent=extent)
 
 
-def _scan_minimal(n_frames: int = 3, n_tx: int = 2, n_el: int = 4, n_ax: int = 8, n_ch: int = 1):
+def _scan_minimal(n_frames: int = 3, n_tx: int = 2, n_el: int = 4):
     return {
-        "n_ax": n_ax,
-        "n_el": n_el,
-        "n_tx": n_tx,
-        "n_ch": n_ch,
-        "n_frames": n_frames,
         "probe_geometry": np.zeros((n_el, 3), dtype=np.float32),
         "sampling_frequency": np.float32(30e6),
         "center_frequency": np.linspace(5e6, 6e6, n_tx, dtype=np.float32),
@@ -78,7 +73,7 @@ def dataset_spec():
                 "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
             },
         },
-        scan=_scan_minimal(n_frames=n_frames, n_tx=n_tx, n_el=n_el, n_ax=n_ax, n_ch=n_ch),
+        scan=_scan_minimal(n_frames=n_frames, n_tx=n_tx, n_el=n_el),
         metadata={
             "subject": {"type": "human", "age": np.uint8(42), "sex": "f", "fat": np.float32(17.5)},
             "credit": "example-lab",
@@ -160,10 +155,38 @@ def test_optional_fields_can_be_omitted():
 
     dataset = DatasetBuilder(
         data={"raw_data": np.zeros((n_frames, n_tx, n_el, n_ax, n_ch), dtype=np.float32)},
-        scan=_scan_minimal(n_frames=n_frames, n_tx=n_tx, n_el=n_el, n_ax=n_ax, n_ch=n_ch),
+        scan=_scan_minimal(n_frames=n_frames, n_tx=n_tx, n_el=n_el),
         metadata={},
         metrics={},
     )
 
     assert dataset.metadata.subject is None
     assert dataset.metrics.common_midpoint_phase_error is None
+
+
+def test_dataset_builder_dimension_consistency_across_nested_specs():
+    n_frames_data, n_frames_scan = 3, 4
+    n_tx, n_el, n_ax, n_ch = 2, 4, 8, 1
+
+    scan = {
+        "probe_geometry": np.zeros((n_el, 3), dtype=np.float32),
+        "sampling_frequency": np.float32(30e6),
+        "center_frequency": np.linspace(5e6, 6e6, n_tx, dtype=np.float32),
+        "demodulation_frequency": np.linspace(5e6, 6e6, n_tx, dtype=np.float32),
+        "initial_times": np.zeros((n_tx,), dtype=np.float32),
+        "t0_delays": np.zeros((n_tx, n_el), dtype=np.float32),
+        "tx_apodizations": np.ones((n_tx, n_el), dtype=np.float32),
+        "focus_distances": np.zeros((n_tx,), dtype=np.float32),
+        "transmit_origins": np.zeros((n_tx, 3), dtype=np.float32),
+        "polar_angles": np.zeros((n_tx,), dtype=np.float32),
+        "azimuth_angles": np.zeros((n_tx,), dtype=np.float32),
+        "time_to_next_transmit": np.ones((n_frames_scan, n_tx), dtype=np.float32),
+    }
+
+    with pytest.raises(ValueError, match="Dimension 'n_frames' has inconsistent sizes"):
+        DatasetBuilder(
+            data={"raw_data": np.zeros((n_frames_data, n_tx, n_el, n_ax, n_ch), dtype=np.float32)},
+            scan=scan,
+            metadata={},
+            metrics={},
+        )
