+++ zea/data/spec.py
@@ -72,6 +72,147 @@ class Spec:
 
     SCHEMA: dict
 
+    @staticmethod
+    def _is_optional_dataclass_field(field_def: Any) -> bool:
+        if field_def is None:
+            return False
+        return field_def.default is not MISSING or field_def.default_factory is not MISSING
+
+    @staticmethod
+    def _expected_shapes(shape_spec: Any) -> tuple[tuple, ...]:
+        if shape_spec and isinstance(shape_spec[0], tuple):
+            return tuple(shape_spec)
+        return (shape_spec,)
+
+    @staticmethod
+    def _merge_dimension_info(
+        dim_to_fields: defaultdict[str, set[str]],
+        dim_to_sizes: defaultdict[str, set[int]],
+        nested_dim_to_fields: defaultdict[str, set[str]],
+        nested_dim_to_sizes: defaultdict[str, set[int]],
+    ) -> None:
+        for dim_name, nested_fields in nested_dim_to_fields.items():
+            dim_to_fields[dim_name].update(nested_fields)
+        for dim_name, nested_sizes in nested_dim_to_sizes.items():
+            dim_to_sizes[dim_name].update(nested_sizes)
+
+    @staticmethod
+    def _track_named_dimensions(
+        dim_to_fields: defaultdict[str, set[str]],
+        dim_to_sizes: defaultdict[str, set[int]],
+        field_path: str,
+        matched_shape: tuple,
+        shape: tuple,
+    ) -> None:
+        for i, dim_name in enumerate(matched_shape):
+            if isinstance(dim_name, str):
+                dim_to_fields[dim_name].add(field_path)
+                dim_to_sizes[dim_name].add(shape[i])
+
+    @staticmethod
+    def _raise_if_shape_mismatch(
+        field_name: str, value: Any, expected_shapes: tuple[tuple, ...]
+    ) -> None:
+        allowed_shapes = ", ".join(str(shape) for shape in expected_shapes)
+        raise ValueError(
+            f"{field_name} has shape {value_shape(value)}, expected one of: {allowed_shapes}"
+        )
+
+    def _validate_nested_field(
+        self, field_name: str, nested_spec: type, field_value: Any
+    ) -> "Spec":
+        if isinstance(field_value, dict):
+            field_value = nested_spec(**field_value)
+            setattr(self, field_name, field_value)
+
+        if not isinstance(field_value, nested_spec):
+            raise TypeError(
+                f"Expected field '{field_name}' to be {nested_spec.__name__}, "
+                f"got {type(field_value).__name__}"
+            )
+
+        return field_value
+
+    def _validate_and_track_primitive_field(
+        self,
+        field_name: str,
+        field_info: dict,
+        field_value: Any,
+        dim_to_fields: defaultdict[str, set[str]],
+        dim_to_sizes: defaultdict[str, set[int]],
+    ) -> None:
+        expected_dtype = field_info["dtype"]
+        expected_shapes = self._expected_shapes(field_info["shape"])
+
+        check_dtype(field_value, expected_dtype)
+
+        matched_shape = find_matched_shape(field_value, expected_shapes)
+        if matched_shape is None:
+            self._raise_if_shape_mismatch(field_name, field_value, expected_shapes)
+
+        self._track_named_dimensions(
+            dim_to_fields=dim_to_fields,
+            dim_to_sizes=dim_to_sizes,
+            field_path=field_name,
+            matched_shape=matched_shape,
+            shape=value_shape(field_value),
+        )
+
+    @staticmethod
+    def _raise_if_inconsistent_dimensions(
+        dim_to_fields: defaultdict[str, set[str]],
+        dim_to_sizes: defaultdict[str, set[int]],
+    ) -> None:
[…]
