### 2026-04-02 bc916e2
doctest
