### 2026-04-02 f50bedb
Fix tests, clean up and add validation
