### 2026-04-02 633f57d
Add docstring and validation for Scan
