### 2026-04-07 71de42d
Rename fat -> fat_percentage
