### 2026-04-07 d3bc0eb
update fat_percentage in spec yaml
