### 2026-04-07 7b7c25d
Fix tests and add test that checks if SCHEMA en dataclass are aligned
