### 2026-04-07 4808da5
Refactor to read_scan in verasonics conversion file
