### 2026-04-07 e55c632
fix typos, tests now passing
