### 2026-04-07 ef07dbe
Rename bandwidth_percent
