### 2026-04-07 5fb26f7
Allow multiple dtypes in spec
