### 2026-04-07 f3dcf83
Store waveforms per n_tx
