### 2026-04-07 4a1a14a
add ColorDopplerMap + add implicit casting of primitive types to numpy equivalent
