### 2026-04-08 f66c4c7
Allow all float precision, but cast to SCHEMA
