+++ zea/data/spec.py
@@ -32,9 +32,7 @@ def check_dtype(value: Any, expected_dtype: List[type]) -> None:
                 return
 
     actual_type = (
-        f"dtype {value.dtype}"
-        if hasattr(value, "dtype")
-        else f"Python {type(value).__name__}"
+        f"dtype {value.dtype}" if hasattr(value, "dtype") else f"Python {type(value).__name__}"
     )
     expected_dtypes_str = ", ".join(str(dt) for dt in expected_dtype)
     raise TypeError(
@@ -88,9 +86,7 @@ class Spec:
     def _is_optional_dataclass_field(field_def: Any) -> bool:
         if field_def is None:
             return False
-        return (
-            field_def.default is not MISSING or field_def.default_factory is not MISSING
-        )
+        return field_def.default is not MISSING or field_def.default_factory is not MISSING
 
     @staticmethod
     def _expected_shapes(shape_spec: Any) -> tuple[tuple, ...]:
@@ -149,16 +145,35 @@ class Spec:
 
     @staticmethod
     def _cast_native_to_numpy(value: Any, expected_dtype: list) -> Any:
-        """Cast Python native types (str, int, float) to numpy equivalents.
+        """Cast values to expected numpy dtypes when possible.
 
-        If the value already has a ``.dtype`` attribute (numpy array or scalar),
-        it is returned unchanged.  Otherwise, each dtype in *expected_dtype* is
-        tried as a cast target; the first successful conversion wins.  If none
-        succeed the original value is returned so that downstream validation
-        can produce a clear error.
+        For fields that expect a floating dtype, all floating-point inputs are
+        accepted and normalized to the first floating dtype in ``expected_dtype``
+        (typically ``np.float32``).
         """
+        expected_np_dtypes = []
+        for dt in expected_dtype:
+            try:
+                expected_np_dtypes.append(np.dtype(dt))
+            except TypeError:
+                continue
+
+        expected_float_dtype = next(
+            (dt for dt in expected_np_dtypes if np.issubdtype(dt, np.floating)),
+            None,
+        )
+
         if hasattr(value, "dtype"):
-            return value  # already a numpy type
+            value_dtype = np.dtype(value.dtype)
+
+            if (
+                expected_float_dtype is not None
+                and np.issubdtype(value_dtype, np.floating)
+                and value_dtype != expected_float_dtype
+            ):
+                return value.astype(expected_float_dtype, copy=False)
+
+            return value
 
         for dt in expected_dtype:
             try:
@@ -230,8 +245,8 @@ class Spec:
 
             nested_spec = field_info.get("spec")
             if nested_spec is not None:
-                nested_dim_to_fields, nested_dim_to_sizes = (
-                    field_value._collect_dimension_info(prefix=f"{prefix}{field_name}.")
+                nested_dim_to_fields, nested_dim_to_sizes = field_value._collect_dimension_info(
+                    prefix=f"{prefix}{field_name}."
                 )
                 self._merge_dimension_info(
                     dim_to_fields,
@@ -275,12 +290,10 @@ class Spec:
 
             nested_spec = field_info.get("spec")
             if nested_spec is not None:
-                field_value = self._validate_nested_field(
-                    field_name, nested_spec, field_value
-                )
+                field_value = self._validate_nested_field(field_name, nested_spec, field_value)
 
-                nested_dim_to_fields, nested_dim_to_sizes = (
-                    field_value._collect_dimension_info(prefix=f"{field_name}.")
+                nested_dim_to_fields, nested_dim_to_sizes = field_value._collect_dimension_info(
+                    prefix=f"{field_name}."
                 )
[…]
