### 2026-04-08 9c1fa50
Improve file validation error
