### 2026-04-08 be3f7a7
Squeeze if all the same values
