### 2026-04-08 076aaa3
paper update (incorporating new zea.Dataloader class from #256
