+++ paper/paper.md
@@ -60,7 +60,7 @@ The ultrasound research community has advanced significantly due to publically a
 # Overview of functionality
 `zea` is an open-source Python package, available at [http://github.com/tue-bmd/zea](http://github.com/tue-bmd/zea), that consists of the following core components:
 
-- **Data**: A set of data handling classes such as `zea.data.File`, `zea.data.Dataset` and `make_dataloader()`, suited for machine learning workflows. `zea` works with HDF5 files, storing data and acquisition parameters together in a single file, which can be easily loaded and saved through the `zea.data` API. For more demanding workflows, such as training deep learning models, `zea` offers robust data loading utilities such as batching, shuffling, caching, and preprocessing. Additionally, we provide examples and conversion scripts for popular ultrasound datasets, such as CAMUS [@leclerc2019deep], PICMUS [@liebgott2016plane], and EchoNet [@ouyang2020video].
+- **Data**: A set of data handling classes such as `zea.File`, `zea.Dataset` and `zea.Dataloader`, suited for machine learning workflows. `zea` works with HDF5 files, storing data and acquisition parameters together in a single file, which can be easily loaded and saved through the `zea` API. For more demanding workflows, such as training deep learning models, `zea` offers robust data loading utilities such as batching, shuffling, caching, and preprocessing. Additionally, we provide examples and conversion scripts for popular ultrasound datasets, such as CAMUS [@leclerc2019deep], PICMUS [@liebgott2016plane], and EchoNet [@ouyang2020video].
 - **Pipeline**: A modular and differentiable pipeline class that allows users to define a sequence of operations (`zea.Operation`) to process ultrasound data. The pipeline is stateless and supports *Just in Time* (JIT) compilation. Ultimately, this allows for dynamic parameter adjustment, as well as real-time integration of deep learning models inside the ultrasound reconstruction pipeline.
 - **Models**: A collection of pre-defined models for ultrasound image and signal processing. Similar to the data, these models can be loaded locally or from the [Hugging Face Hub](https://huggingface.co/zeahub). Besides supervised models, `zea` also provides a set of (deep) generative models, with an interface to solve inverse problems in ultrasound imaging within a probabilistic machine learning framework.
 - **Agents**: A set of tools to interact with the pipeline and models. These agents can be used to alter the pipeline parameters, or select a subset of acquired data. The agent module closes the action-perception loop [@van2024active], tying together acquisition and reconstruction of ultrasound data.
