### 2026-04-09 c4c498a
Allow zeros in extent
