### 2026-04-09 55a8876
Allow arbitrary spatial maps in file["data"]
