+++ zea/data/spec.py
@@ -131,13 +131,16 @@ class Spec:
         )
 
     def _validate_nested_field(
-        self, field_name: str, nested_spec: type, field_value: Any
+        self, field_name: str, nested_spec: "Spec", field_value: Any
     ) -> "Spec":
+        """Validate a nested spec field, recursively validating its contents."""
         if isinstance(field_value, dict):
             field_value = nested_spec(**field_value)
             setattr(self, field_name, field_value)
 
-        if not isinstance(field_value, nested_spec):
+        # Check that the nested spec field is now an instance of the expected Spec subclass
+        # E.g. Segmentation if nested_spec is Map
+        if not issubclass(type(field_value), nested_spec):
             raise TypeError(
                 f"Expected field '{field_name}' to be {nested_spec}, got {type(field_value)}"
             )
@@ -397,7 +400,7 @@ class Map(Spec):
     """Map data and spatial extent metadata.
 
     Args:
-        pixels: The map pixels of shape (n_frames, h, w, d) of type uint8.
+        pixels: The map pixels of shape (n_frames, h, w, d) of type uint8 or float32
         extent: The map extent in meters of shape (n_frames, 6) or (6,).
             A shape of (6,) is broadcast to all frames. Values are ordered as
             (xmin, xmax, ymin, ymax, zmax, zmin) and stored as float32.
@@ -407,7 +410,7 @@ class Map(Spec):
     extent: np.ndarray
 
     SCHEMA = {
-        "pixels": {"dtype": np.uint8, "shape": ("n_frames", "h", "w", "d")},
+        "pixels": {"dtype": (np.uint8, np.float32), "shape": ("n_frames", "h", "w", "d")},
         "extent": {"dtype": np.float32, "shape": (("n_frames", 6), (6,))},
     }
 
@@ -431,7 +434,47 @@ class Map(Spec):
 
 
 @dataclass
-class Segmentation(Map):
+class FloatMap(Map):
+    """Map data with float32 pixel values and spatial extent metadata.
+
+    Args:
+        pixels: The map pixels of shape (n_frames, h, w, d) and type float32.
+        extent: The map extent in meters of shape (n_frames, 6) or (6,).
+            A shape of (6,) is broadcast to all frames. Values are ordered as
+            (xmin, xmax, ymin, ymax, zmax, zmin) and stored as float32.
+    """
+
+    pixels: np.ndarray
+    extent: np.ndarray
+
+    SCHEMA = {
+        "pixels": {"dtype": np.float32, "shape": ("n_frames", "h", "w", "d")},
+        "extent": {"dtype": np.float32, "shape": (("n_frames", 6), (6,))},
+    }
+
+
+@dataclass
+class UnsignedIntMap(Map):
+    """Map data with uint8 pixel values and spatial extent metadata.
+
+    Args:
+        pixels: The map pixels of shape (n_frames, h, w, d) and type uint8.
+        extent: The map extent in meters of shape (n_frames, 6) or (6,).
+            A shape of (6,) is broadcast to all frames. Values are ordered as
+            (xmin, xmax, ymin, ymax, zmax, zmin) and stored as float32.
+    """
+
+    pixels: np.ndarray
+    extent: np.ndarray
+
+    SCHEMA = {
+        "pixels": {"dtype": np.uint8, "shape": ("n_frames", "h", "w", "d")},
+        "extent": {"dtype": np.float32, "shape": (("n_frames", 6), (6,))},
+    }
+
+
+@dataclass
+class Segmentation(UnsignedIntMap):
     """Segmentation data and spatial extent metadata.
 
     Args:
@@ -463,27 +506,7 @@ class Segmentation(Map):
 
 
 @dataclass
-class FloatMap(Map):
-    """Map data with float32 pixel values and spatial extent metadata.
-
-    Args:
[…]
