### 2026-04-09 dd621b4
Allow more generic maps and signals (WIP)
