+++ zea/data/spec.py
@@ -9,6 +9,18 @@ from zea import log
 
 CONSISTENCY_DIMENSIONS = {"n_frames", "n_tx", "n_ax", "n_el", "n_ch"}
 
+UNITS = {
+    "m/s": "meters per second",
+    "m": "meters",
+    "Hz": "Hertz",
+    "s": "seconds",
+    "-": "unitless",
+    "rad": "radians",
+    "dB": "decibels",
+    "#": "count",
+    "%": "percent",
+}
+
 
 def check_dtype(value: Any, expected_dtype: List[type]) -> None:
     """Check if the dtype of a value matches the expected dtype,
@@ -399,23 +411,49 @@ class Spec:
 class Map(Spec):
     """Map data and spatial extent metadata.
 
+    The most flexible map spec, which can be used for any spatially aligned data product.
+
     Args:
-        pixels: The map pixels of shape (n_frames, x, z, y) of type uint8 or float32
+        pixels: The map pixels of shape (n_frames, x, z, y, n_ch) or (n_frames, x, z, y)
+            and type uint8 or float32 or int16.
         extent: The map extent in meters of shape (n_frames, 6) or (6,).
             A shape of (6,) is broadcast to all frames. Values are ordered as
             (xmin, xmax, ymin, ymax, zmax, zmin) and stored as float32.
+        channel_labels: The labels corresponding to the `n_ch` channels in the pixels.
+            This is required when pixels have an n_ch dimension, and should be None otherwise.
+            For IQ data, this would typically be ["I", "Q"].
+        description: An optional free-text description of the map.
+        unit: An optional string specifying the physical unit of the map values,
+            e.g. "m/s", "%", etc.
     """
 
     pixels: np.ndarray
     extent: np.ndarray
+    channel_labels: np.ndarray | None = None
+    description: str | None = None
+    unit: str | None = None
 
     SCHEMA = {
-        "pixels": {"dtype": (np.uint8, np.float32), "shape": ("n_frames", "x", "z", "y")},
+        "pixels": {
+            "dtype": (np.uint8, np.float32, np.int16, np.complex64),
+            "shape": (
+                ("n_frames", "x", "z", "y", "n_ch"),
+                ("n_frames", "x", "z", "y"),
+            ),
+        },
         "extent": {"dtype": np.float32, "shape": (("n_frames", 6), (6,))},
+        "channel_labels": {"dtype": np.str_, "shape": ("n_ch",)},
+        "description": {"dtype": str, "shape": ()},
+        "unit": {"dtype": str, "shape": ()},
     }
 
     def __post_init__(self):
         super().__post_init__()
+
+        if self.pixels.ndim == 5:
+            assert self.channel_labels is not None, (
+                "channel_labels must be provided when pixels have n_ch dimension"
+            )
 
         # Check sensible values
         if np.any(self.extent[..., 0] > self.extent[..., 1]):
@@ -435,63 +473,43 @@ class Map(Spec):
 
 @dataclass
 class FloatMap(Map):
-    """Map data with float32 pixel values and spatial extent metadata.
+    """Map data with float32 pixel values and spatial extent metadata."""
 
-    Args:
-        pixels: The map pixels of shape (n_frames, x, z, y) and type float32.
-        extent: The map extent in meters of shape (n_frames, 6) or (6,).
-            A shape of (6,) is broadcast to all frames. Values are ordered as
-            (xmin, xmax, ymin, ymax, zmax, zmin) and stored as float32.
-    """
+    def __post_init__(self):
+        self.SCHEMA["pixels"]["dtype"] = np.float32
+        return super().__post_init__()
 
[…]
