### 2026-04-09 e10260e
Allow custom signals
