+++ zea/data/spec.py
@@ -1,3 +1,4 @@
+import warnings
 from collections import defaultdict
 from dataclasses import MISSING, dataclass, field, fields
 from typing import Any, List
@@ -7,7 +8,7 @@ import numpy as np
 
 from zea import log
 
-CONSISTENCY_DIMENSIONS = {"n_frames", "n_tx", "n_ax", "n_el", "n_ch"}
+CONSISTENCY_DIMENSIONS = {"n_frames", "n_tx", "n_ax", "n_el", "n_ch", "n_spatial_ch"}
 
 UNITS = {
     "m/s": "meters per second",
@@ -333,7 +334,10 @@ class Spec:
 
             nested_spec = field_info.get("spec")
             if nested_spec is not None:
-                field_value = self._validate_nested_field(field_name, nested_spec, field_value)
+                try:
+                    field_value = self._validate_nested_field(field_name, nested_spec, field_value)
+                except (TypeError, ValueError) as e:
+                    raise type(e)(f"In field '{field_name}': {e}") from e
 
                 nested_dim_to_fields, nested_dim_to_sizes = field_value._collect_dimension_info(
                     prefix=f"{field_name}."
@@ -442,7 +446,7 @@ class Map(Spec):
         extent: The map extent in meters of shape (n_frames, 6) or (6,).
             A shape of (6,) is broadcast to all frames. Values are ordered as
             (xmin, xmax, ymin, ymax, zmax, zmin) and stored as float32.
-        channel_labels: The labels corresponding to the `n_ch` channels in the pixels.
+        labels: The labels corresponding to the `n_ch` channels in the pixels.
             This is required when pixels have an n_ch dimension, and should be None otherwise.
             For IQ data, this would typically be ["I", "Q"].
         description: An optional free-text description of the map.
@@ -452,7 +456,7 @@ class Map(Spec):
 
     pixels: np.ndarray
     extent: np.ndarray
-    channel_labels: np.ndarray | None = None
+    labels: np.ndarray | None = None
     description: str | None = None
     unit: str | None = None
 
@@ -460,12 +464,12 @@ class Map(Spec):
         "pixels": {
             "dtype": (np.uint8, np.float32, np.int16, np.complex64),
             "shape": (
-                ("n_frames", "x", "z", "y", "n_ch"),
+                ("n_frames", "x", "z", "y", "n_spatial_ch"),
                 ("n_frames", "x", "z", "y"),
             ),
         },
         "extent": {"dtype": np.float32, "shape": (("n_frames", 6), (6,))},
-        "channel_labels": {"dtype": np.str_, "shape": ("n_ch",)},
+        "labels": {"dtype": np.str_, "shape": ("n_spatial_ch",)},
         "description": {"dtype": str, "shape": ()},
         "unit": {"dtype": str, "shape": ()},
     }
@@ -474,8 +478,8 @@ class Map(Spec):
         super().__post_init__()
 
         if self.pixels.ndim == 5:
-            assert self.channel_labels is not None, (
-                "channel_labels must be provided when pixels have n_ch dimension"
+            assert self.labels is not None, (
+                "labels must be provided when pixels have n_ch dimension"
             )
 
         # Check sensible values
@@ -498,27 +502,39 @@ class Map(Spec):
 class FloatMap(Map):
     """Map data with float32 pixel values and spatial extent metadata."""
 
-    def __post_init__(self):
-        self.SCHEMA["pixels"]["dtype"] = np.float32
-        return super().__post_init__()
+    SCHEMA = {
+        **Map.SCHEMA,
+        "pixels": {
+            **Map.SCHEMA["pixels"],
+            "dtype": np.float32,
+        },
+    }
 
 
 @dataclass
[…]
