### 2026-04-09 862ce48
Add subject id, and warn if not provided
