+++ tests/data/test_spec.py
@@ -42,71 +42,79 @@ def _scan_minimal(n_frames: int = 3, n_tx: int = 2, n_el: int = 4):
     }
 
 
+def _example_metadata():
+    return {
+        "subject": {
+            "type": "human",
+            "age": np.uint8(42),
+            "sex": "f",
+            "fat_percentage": np.float32(17.5),
+        },
+        "credit": "example-lab",
+        "probe_orientation": {
+            "pose": np.zeros((25, 6), dtype=np.float32),
+            "offset": np.float32(0.0),
+            "sampling_frequency": np.float32(50.0),
+        },
+        "voice_narration": {
+            "samples": np.zeros((100), dtype=np.uint8),
+            "offset": np.float32(0.0),
+            "sampling_frequency": np.float32(8000.0),
+        },
+        "ecg": {
+            "samples": np.zeros((100), dtype=np.uint8),
+            "offset": np.float32(0.0),
+            "sampling_frequency": np.float32(250.0),
+        },
+        "text_report": "normal acquisition",
+        "annotations": {
+            "anatomy": "heart",
+            "view": np.array(["plax", "plax", "psax"], dtype=np.str_),
+            "label": np.array(["normal", "normal", "normal"], dtype=np.str_),
+            "image_quality": "high",
+        },
+    }
+
+
+def _example_data(n_frames, n_tx, n_el, n_ax, n_ch):
+    return {
+        "raw_data": np.zeros((n_frames, n_tx, n_ax, n_el, n_ch), dtype=np.float32),
+        "image": {
+            "pixels": np.zeros((n_frames, 16, 12, 1), dtype=np.uint8),
+            "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
+        },
+        "segmentation": {
+            "pixels": np.zeros((n_frames, 16, 12, 1, 2), dtype=np.bool_),
+            "labels": np.array(["background", "tissue"], dtype=np.str_),
+            "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
+        },
+        "sos_map": {
+            "pixels": np.full((n_frames, 16, 12, 1), 1540.0, dtype=np.float32),
+            "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
+        },
+        "strain": {
+            "pixels": np.zeros((n_frames, 16, 12, 1), dtype=np.float32),
+            "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
+        },
+        "swe": {
+            "pixels": np.zeros((n_frames, 16, 12, 1), dtype=np.float32),
+            "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
+        },
+        "tissue_doppler": {
+            "pixels": np.zeros((n_frames, 16, 12, 1), dtype=np.float32),
+            "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
+        },
+    }
+
+
 @pytest.fixture
 def dataset_spec():
     n_frames, n_tx, n_el, n_ax, n_ch = 3, 2, 4, 8, 1
 
     return FileSpec(
-        data={
-            "raw_data": np.zeros((n_frames, n_tx, n_ax, n_el, n_ch), dtype=np.float32),
-            "image": {
-                "pixels": np.zeros((n_frames, 16, 12, 1), dtype=np.uint8),
-                "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
[…]
