### 2026-04-09 515d0cd
Update setuptools requirement in the python-packages group
Updates the requirements on [setuptools](https://github.com/pypa/setuptools) to permit the latest version.

Updates `setuptools` to 82.0.1
- [Release notes](https://github.com/pypa/setuptools/releases)
- [Changelog](https://github.com/pypa/setuptools/blob/main/NEWS.rst)
- [Commits](https://github.com/pypa/setuptools/compare/0.6...v82.0.1)

---
updated-dependencies:
- dependency-name: setuptools
  dependency-version: 82.0.1
  dependency-type: direct:development
  dependency-group: python-packages
...

Signed-off-by: dependabot[bot] <sup
[…]
