+++ pyproject.toml
@@ -74,7 +74,7 @@ dev = [
     "sphinxcontrib-autoprogram",
     "sphinx-reredirects",
     "sphinxcontrib-bibtex",
-    "setuptools<82",
+    "setuptools<83",
     "myst-parser",
     "nbsphinx",
     "furo",
@@ -106,7 +106,7 @@ docs = [
     "sphinxcontrib-autoprogram",
     "sphinx-reredirects",
     "sphinxcontrib-bibtex",
-    "setuptools<82",
+    "setuptools<83",
     "myst-parser",
     "nbsphinx",
     "furo",
