### 2026-04-09 d4ae789
remove poetry lock from precommit CI (not from real precommit hook)
