+++ .github/workflows/precommits.yaml
@@ -30,7 +30,6 @@ jobs:
           pre-commit run check-merge-conflict --all-files
           pre-commit run detect-private-key --all-files
           pre-commit run poetry-check --all-files
-          pre-commit run poetry-lock --all-files
           pre-commit run notebook-clean-and-check --all-files
         continue-on-error: true
 
