### 2026-04-09 d423d70
increase permissions for dependabot
