### 2026-04-09 35561bb
remove autogenerated keras ops from coverage
