+++ pyproject.toml
@@ -238,3 +238,8 @@ docstring-code-format = true
 # This only has an effect when the `docstring-code-format` setting is
 # enabled.
 docstring-code-line-length = "dynamic"
+
+[tool.coverage.report]
+omit = [
+    "zea/ops/keras_ops.py",
+]
