### 2026-04-09 cd7dc7f
Bump the github-actions group with 8 updates
Bumps the github-actions group with 8 updates:

| Package | From | To |
| --- | --- | --- |
| [actions/checkout](https://github.com/actions/checkout) | `4` | `6` |
| [docker/setup-buildx-action](https://github.com/docker/setup-buildx-action) | `2` | `4` |
| [docker/login-action](https://github.com/docker/login-action) | `2` | `4` |
| [docker/build-push-action](https://github.com/docker/build-push-action) | `3` | `7` |
| [actions/setup-python](https://github.com/actions/setup-python) | `5` | `6` |
| [actions/upload-artifact](https://github.com/action
[…]
