### 2026-04-10 tue-bmd/zea#278 @tristan-deep
Removed the TF installation mention again in the notebook, since that requirement has been removed. We now use backend agnostic dataloading with Grain, which was added in #256.

@tomelse let me know if the issues you raised have been resolved, then I will close this.
