+++ tests/data/test_file.py
@@ -3,7 +3,8 @@
 import numpy as np
 import pytest
 
-from zea.data.file import File, dict_to_sorted_list, load_file
+from zea.data.file import File, GroupProxy, dict_to_sorted_list, load_file
+from zea.data.spec import FileSpec, Image, Segmentation
 from zea.probes import Probe
 from zea.scan import Scan
 
@@ -127,3 +128,580 @@ def test_dict_to_sorted_list():
     assert sorted_list == [1, 2, 3], "The sorted list should be [1, 2, 3]"
 
     assert dict_to_sorted_list({}) == [], "The sorted list of an empty dict should be []"
+
+
+def _scan_minimal(n_frames=3, n_tx=2, n_el=4):
+    return {
+        "probe_geometry": np.zeros((n_el, 3), dtype=np.float32),
+        "sampling_frequency": np.float32(30e6),
+        "center_frequency": np.float32(5e6),
+        "demodulation_frequency": np.float32(5e6),
+        "initial_times": np.zeros((n_tx,), dtype=np.float32),
+        "t0_delays": np.zeros((n_tx, n_el), dtype=np.float32),
+        "tx_apodizations": np.ones((n_tx, n_el), dtype=np.float32),
+        "focus_distances": np.zeros((n_tx,), dtype=np.float32),
+        "transmit_origins": np.zeros((n_tx, 3), dtype=np.float32),
+        "polar_angles": np.zeros((n_tx,), dtype=np.float32),
+        "azimuth_angles": np.zeros((n_tx,), dtype=np.float32),
+        "time_to_next_transmit": np.ones((n_frames, n_tx), dtype=np.float32),
+    }
+
+
+@pytest.fixture
+def spec_file(tmp_path):
+    """Create a spec-format HDF5 file via FileSpec.save()."""
+    n_frames, n_tx, n_el, n_ax, n_ch = 2, 3, 4, 8, 1
+    raw = np.random.randn(n_frames, n_tx, n_ax, n_el, n_ch).astype(np.float32)
+    env = np.random.randn(n_frames, 16, 12).astype(np.float32)
+
+    fspec = FileSpec(
+        data={"raw_data": raw, "envelope_data": env},
+        scan=_scan_minimal(n_frames=n_frames, n_tx=n_tx, n_el=n_el),
+        probe_name="test_probe",
+        description="spec format test file",
+    )
+    path = tmp_path / "spec_format.hdf5"
+    fspec.save(str(path))
+    return str(path), fspec, raw, env
+
+
+class TestGroupProxy:
+    def test_attribute_access_returns_dataset(self, spec_file):
+        path, _, raw, _ = spec_file
+        import h5py as _h5py
+
+        with File(path) as f:
+            ds = f.data.raw_data
+            assert isinstance(ds, _h5py.Dataset)
+            assert ds.shape == raw.shape
+
+    def test_slicing_loads_subset(self, spec_file):
+        path, _, raw, _ = spec_file
+
+        with File(path) as f:
+            loaded = f.data.raw_data[:, :2]
+            np.testing.assert_array_equal(loaded, raw[:, :2])
+
+    def test_nested_group_access(self, tmp_path):
+        """Groups within 'data' are also proxied."""
+        n_frames, n_tx, n_el, n_ax, n_ch = 2, 3, 4, 8, 1
+        fspec = FileSpec(
+            data={
+                "raw_data": np.zeros((n_frames, n_tx, n_ax, n_el, n_ch), dtype=np.float32),
+                "image": {
+                    "pixels": np.zeros((n_frames, 16, 12, 1), dtype=np.uint8),
[…]
