### 2026-04-10 5dedb8b
More integration dataspec * add zea version * clean up names, clearly add Spec * cleanup imports * more tests
