+++ tests/data/test_spec.py
@@ -5,7 +5,20 @@ import pytest
 
 from zea.data import spec as spec_module
 from zea.data.file import File
-from zea.data.spec import Data, FileSpec, Image, Map, Scan, Segmentation, SignalND, Spec
+from zea.data.spec import (
+    Data,
+    FileSpec,
+    Image,
+    Map,
+    MetricsSpec,
+    ScanSpec,
+    Segmentation,
+    Signal1D,
+    SignalND,
+    SosMap,
+    Spec,
+    Subject,
+)
 
 
 def test_segmentation_spec():
@@ -189,7 +202,7 @@ def test_scan_requires_required_fields():
     with pytest.raises(
         TypeError, match="missing 1 required positional argument: 'demodulation_frequency'"
     ):
-        Scan(**scan)
+        ScanSpec(**scan)
 
 
 def test_scan_dimension_count_consistency():
@@ -197,7 +210,7 @@ def test_scan_dimension_count_consistency():
     scan["initial_times"] = np.zeros((3,), dtype=np.float32)
 
     with pytest.raises(ValueError, match="Dimension 'n_tx' has inconsistent sizes"):
-        Scan(**scan)
+        ScanSpec(**scan)
 
 
 def test_signal_nd_accepts_variable_trailing_dimensions_with_ellipsis():
@@ -239,17 +252,19 @@ def test_scan_accepts_float_inputs_and_casts_to_float32():
     scan["initial_times"] = np.zeros((2,), dtype=np.float64)
     scan["t0_delays"] = np.zeros((2, 4), dtype=np.float64)
 
-    scan_spec = Scan(**scan)
+    scan_spec = ScanSpec(**scan)
 
     assert np.dtype(scan_spec.sampling_frequency.dtype) == np.dtype(
-        Scan.SCHEMA["sampling_frequency"]["dtype"]
+        ScanSpec.SCHEMA["sampling_frequency"]["dtype"]
     )
-    assert scan_spec.center_frequency.dtype == np.dtype(Scan.SCHEMA["center_frequency"]["dtype"])
+    assert scan_spec.center_frequency.dtype == np.dtype(
+        ScanSpec.SCHEMA["center_frequency"]["dtype"]
+    )
     assert np.dtype(scan_spec.demodulation_frequency.dtype) == np.dtype(
-        Scan.SCHEMA["demodulation_frequency"]["dtype"]
+        ScanSpec.SCHEMA["demodulation_frequency"]["dtype"]
     )
-    assert scan_spec.initial_times.dtype == np.dtype(Scan.SCHEMA["initial_times"]["dtype"])
-    assert scan_spec.t0_delays.dtype == np.dtype(Scan.SCHEMA["t0_delays"]["dtype"])
+    assert scan_spec.initial_times.dtype == np.dtype(ScanSpec.SCHEMA["initial_times"]["dtype"])
+    assert scan_spec.t0_delays.dtype == np.dtype(ScanSpec.SCHEMA["t0_delays"]["dtype"])
 
 
 def test_dataset_builder_accepts_float_raw_data_and_casts_to_float32():
@@ -432,26 +447,26 @@ class TestScanValidationErrors:
         scan = _scan_minimal()
         scan["probe_geometry"] = np.zeros((4, 3), dtype=np.int32)
         with pytest.raises(TypeError, match="probe_geometry"):
-            Scan(**scan)
+            ScanSpec(**scan)
 
     def test_probe_geometry_wrong_shape_raises(self):
         """probe_geometry must be (n_el, 3) — literal 3 is enforced."""
         scan = _scan_minimal(n_el=4)
         scan["probe_geometry"] = np.zeros((4, 2), dtype=np.float32)
[…]

[…]
