+++ docs/source/spec_doc.py
@@ -1,0 +1,444 @@
+"""Auto-generate the Data Spec RST section from spec.py source of truth.
+
+Run from the project root::
+
+    python docs/source/spec_doc.py
+
+This writes ``docs/source/_spec_ref.rst``, which is included verbatim by
+``data-acquisition.rst``.
+"""
+
+import sys
+from dataclasses import MISSING, fields
+from pathlib import Path
+
+sys.path.insert(0, str(Path(__file__).parent.parent.parent))
+
+import numpy as np
+
+from zea.data.spec import (
+    Annotations,
+    ColorDopplerMap,
+    Data,
+    Image,
+    MetadataSpec,
+    MetricsSpec,
+    ProbeOrientation,
+    ScanSpec,
+    Segmentation,
+    ShearWaveElastographyMap,
+    Signal1D,
+    SosMap,
+    StrainPercentageMap,
+    Subject,
+    TissueDopplerMap,
+)
+
+AUTOGEN_HEADER = """\
+.. THIS FILE WAS AUTOGENERATED USING docs/source/spec_doc.py. DO NOT EDIT MANUALLY.
+   Run ``python docs/source/spec_doc.py`` from the project root to regenerate.
+
+"""
+
+OUT_PATH = Path(__file__).parent / "_spec_ref.rst"
+
+_IND = "   "  # 3-space RST indent unit
+
+# ---------------------------------------------------------------------------
+# helpers
+# ---------------------------------------------------------------------------
+
+
+def _dtype_str(dtype) -> str:
+    """Return inline-code RST for a dtype — always in double backticks."""
+    if isinstance(dtype, tuple):
+        return " | ".join(_dtype_str(d) for d in dtype)
+    # Python str and numpy str_ both render as ``str``
+    if dtype is str or dtype is np.str_:
+        return "``str``"
+    # Map numpy bool_ to bool for readability
+    if dtype is np.bool_:
+        return "``bool``"
+    if hasattr(dtype, "__name__"):
+        return f"``{dtype.__name__}``"
+    return f"``{dtype}``"
+
+
+def _shape_str(shape) -> str:
+    if isinstance(shape, tuple) and shape and isinstance(shape[0], tuple):
+        return " or ".join(_shape_str(s) for s in shape)
+    if shape == ():
+        return "scalar"
+    return "(" + ", ".join(str(s) for s in shape) + ")"
+
+
+def _is_optional(cls, name: str) -> bool:
+    try:
+        for f in fields(cls):
+            if f.name == name:
+                return f.default is not MISSING or f.default_factory is not MISSING
+    except TypeError:
+        pass
+    return False
+
+
+def _first_sentence(docstring: str) -> str:
+    for line in (docstring or "").splitlines():
+        line = line.strip()
+        if line:
+            return line
+    return ""
+
+
+BADGES = """\
+.. |badge-req| replace:: :bdg-danger:`required`
+.. |badge-opt| replace:: :bdg-secondary:`optional`
+"""
+
+# ---------------------------------------------------------------------------
+# Table renderers
+# ---------------------------------------------------------------------------
+
+
+def rst_full_table(cls, base_indent: int = 0) -> str:
+    """Render a spec class as a list-table with all columns including description."""
+    schema = cls.SCHEMA
+    meta = getattr(cls, "FIELD_METADATA", {})
+    p = _IND * base_indent
+
+    rows = []
+    for fname, fspec in schema.items():
+        fm = meta.get(fname, {})
+        unit = fm.get("unit", "\u2013")
+        desc = fm.get("description", "")
+
+        if "spec" in fspec:
+            sub = fspec["spec"]
[…]
