+++ docs/source/config.rst
@@ -1,15 +1,16 @@
-.. THIS FILE WAS AUTOGENERATED USING docs/parameters_doc.py. DO NOT EDIT MANUALLY.
+.. THIS FILE WAS AUTOGENERATED USING docs/source/parameters_doc.py. DO NOT EDIT MANUALLY.
 
-.. _parameters:
+.. _config:
 
-Parameters
-===========
+Config
+======
 
-This page provides a comprehensive overview of all configuration parameters available in zea.
-These parameters are used in the YAML config files to control data loading, preprocessing, model settings, scan parameters, and more.
+This page documents the ``zea`` configuration system.  Configs are YAML files that
+control data loading, preprocessing, model settings, and scan parameters.
 
 .. note::
-  You can use these configs to, for instance, initialize :doc:`zea.Models <models>` or :doc:`zea.Pipeline <pipeline>`.
+  Configs are used to initialize :doc:`zea.Models <models>` and the :doc:`pipeline`.
+  For the data format and file I/O, see :doc:`data-acquisition`.
 
 Configs are written in YAML format and can be loaded, edited, and saved using the ``zea`` API.
 

[…]
