### 2026-04-13 ea86ccc
jit complex_to_channels
