### 2026-04-13 67cc16b
remove outdated yaml
