### 2026-04-13 9993c9d
refactor MachBeamform
