+++ zea/ops/ultrasound.py
@@ -1,4 +1,3 @@
-import inspect
 import uuid
 from typing import Tuple
 
@@ -37,21 +36,14 @@ from zea.simulator import simulate_rf
 from zea.utils import canonicalize_axis
 
 try:
-    from mach import experimental as mach_experimental
-except ImportError:
-    mach_experimental = None
-
-try:
     import cupy as cp
 except ImportError:
     cp = None
 
 try:
     import jax
-    import jax.dlpack as jdlpack
 except ImportError:
     jax = None
-    jdlpack = None
 
 
 @ops_registry("simulate_rf")
@@ -224,7 +216,19 @@ class TOFCorrection(Operation):
 
 @ops_registry("mach_beamform")
 class MachBeamform(Operation):
-    """Mach-based delay-and-sum beamforming for RF or IQ data."""
+    """Mach-based delay-and-sum beamforming for RF or IQ data.
+
+    Beamforming is split into three stages:
+
+    1. :meth:`_prepare_inputs` — translate Zea scan parameters and data to
+       the layout expected by ``mach.kernel.beamform``.  Uses only Keras/Zea
+       ops so it can be JIT-compiled.
+    2. :meth:`_run_cuda_kernel` — loop over transmits and accumulate the
+       CUDA beamform result.  Not JIT-compilable (GPU kernel invocation).
+    3. :meth:`_prepare_outputs` — convert the accumulated ``(n_pix, n_frames)``
+       result back to the Zea data layout.  Uses only Keras/Zea ops so it
+       can be JIT-compiled.
+    """
 
     STATIC_PARAMS = ["interp_type", "tukey_alpha"]
 
@@ -235,16 +239,328 @@ class MachBeamform(Operation):
             jittable=False,
             **kwargs,
         )
+
+        try:
+            import mach  # noqa: F401
+        except ImportError as exc:
+            raise ImportError(
+                "mach is not installed. Please run "
+                "`pip install mach-beamform` to use this operation."
+            ) from exc
+
+        if cp is None:
+            raise ImportError(
+                "cupy is not installed. Please run "
+                "`pip install cupy` to use this operation."
+            )
+
         self.interp_type = interp_type
         self.tukey_alpha = float(tukey_alpha)
+
+        # Resolve the interp_type enum once at construction time so it can be
+        # passed straight to the kernel without a string lookup each call.
+        from mach.kernel import InterpolationType
+
+        _interp_map = {
+            "nearest": InterpolationType.NearestNeighbor,
+            "linear": InterpolationType.Linear,
+            "quadratic": InterpolationType.Quadratic,
+        }
+        if isinstance(interp_type, str):
+            key = interp_type.lower()
+            if key not in _interp_map:
+                raise ValueError(
+                    f"Unsupported interp_type '{interp_type}'. "
+                    "Use 'nearest', 'linear', or 'quadratic'."
+                )
+            self._interp_type_enum = _interp_map[key]
+        else:
+            self._interp_type_enum = interp_type
+
         if keras.backend.backend() == "jax":
             self._calculate_delays = backend_jit(calculate_delays, static_argnums=(7, 8, 14))
-            # JIT compile channels_to_complex for jax backend to ensure all jax calls
-            # inside MachBeamform are jitted (especially conversion of IQ channels).
             self._channels_to_complex = backend_jit(channels_to_complex)
         else:
             self._calculate_delays = backend_jit(calculate_delays)
             self._channels_to_complex = channels_to_complex
+
+    # ------------------------------------------------------------------
+    # Stage 1: Zea → mach  (pure tensor ops, JIT-able)
+    # ------------------------------------------------------------------
+
[…]
