### 2026-04-13 e830954
add overlay_masks func
