### 2026-04-14 tue-bmd/zea#278 @tomelse
hi, sorry I'm still getting issues running the first example script. I get an error like this :

```
zea: Searching [/tmp/zea_cache_q5nkpu3q/huggingface/datasets/datasets--zeahub--camus-sample/snapshots/617cf91a1267b5ffbcfafe9bebf0813c7cee8493/val](https://file+.vscode-resource.vscode-cdn.net/tmp/zea_cache_q5nkpu3q/huggingface/datasets/datasets--zeahub--camus-sample/snapshots/617cf91a1267b5ffbcfafe9bebf0813c7cee8493/val) for ['.hdf5', '.h5'] files...
Checking dataset files on validity (zea format): 100%|██████████| 2/2 [00:00<00:00, 492.84it/s]
zea: Dataset validated. Check [/tmp/zea_cache_q5nkpu3q/huggingface/datasets/datasets--zeahub--camus-sample/snapshots/617cf91a1267b5ffbcfafe9bebf0813c7cee8493/val/validated.flag](https://file+.vscode-resource.vscode-cdn.net/tmp/zea_cache_q5nkpu3q/huggingface/datasets/datasets--zeahub--camus-sample/snapshots/617cf91a1267b5ffbcfafe9bebf0813c7cee8493/val/validated.flag) for details.
zea: Caching is globally disabled for _find_h5_file_shapes.

[/home/telse/.local/share/uv/python/cpython-3.13.2-linux-x86_64-gnu/lib/python3.13/multiprocessing/popen_fork.py:67](https://file+.vscode-resource.vscode-cdn.net/home/telse/.local/share/uv/python/cpython-3.13.2-linux-x86_64-gnu/lib/python3.13/multiprocessing/popen_fork.py:67): RuntimeWarning: os.fork() was called. os.fork() is incompatible with multithreaded code, and JAX is multithreaded, so this will likely lead to a deadlock.
  self.pid = os.fork()
Getting file shapes in each h5 file:   0%|          | 0/2 [00:00<?, ?it/s]

FileNotFoundError                         Traceback (most recent call last)
Cell In[7], [line 2](vscode-notebook-cell:?execution_count=7&line=2)
      1 dataset_path = "hf://zeahub/camus-sample/val"
----> [2](vscode-notebook-cell:?execution_count=7&line=2) dataloader = Dataloader(
      3     dataset_path,
      4     key="data/image_sc",
      5     batch_size=4,

File ~/Documents/Paper Review Reports/zea_review/.venv/lib/python3.13/site-packages/zea/data/dataloader.py:511, in Dataloader.__init__(self, file_paths, key, batch_size, n_frames, shuffle, return_filename, seed, limit_n_samples, limit_n_frames, drop_remainder, image_size, resize_type, resize_axes, resize_kwargs, image_range, normalization_range, clip_image_range, assert_image_range, dataset_repetitions, cache, additional_axes_iter, sort_files, overlapping_blocks, augmentation, initial_frame_axis, insert_frame_axis, frame_index_stride, frame_axis, validate, prefetch, shard_index, num_shards, num_threads, prefetch_buffer_size, **kwargs)
    508 self._rng = np.random.default_rng(seed)
    510 # ── Data source ───────────────────────────────────────────────
--> [511](https://file+.vscode-resource.vscode-cdn.net/home/telse/Documents/Paper%20Review%20Reports/zea_review/~/Documents/Paper%20Review%20Reports/zea_review/.venv/lib/python3.13/site-packages/zea/data/dataloader.py:511) self.source = H5DataSource(
    512     file_paths=file_paths,
    513     key=key,
    514     n_frames=n_frames,
    515     frame_index_stride=frame_index_stride,
    516     frame_axis=frame_axis,
    517     insert_frame_axis=insert_frame_axis,
    518     initial_frame_axis=initial_frame_axis,
    519     additional_axes_iter=additional_axes_iter,
    520     sort_files=sort_files,
    521     overlapping_blocks=overlapping_blocks,
    522     limit_n_samples=limit_n_samples,
    523     limit_n_frames=limit_n_frames,
    524     return_filename=return_filename,
    525     cache=cache,
    526     validate=validate,
    527     **kwargs,
    528 )
    530 # ── Store pipeline config for rebuilding per epoch ────────────
    531 self._pipeline_cfg = dict(
    532     num_shards=num_shards,
    533     shard_index=shard_index,
   (...)    541     resizer=None,
    542 )

File ~/Documents/Paper Review Reports/zea_review/.venv/lib/python3.13/site-packages/zea/data/dataloader.py:242, in H5DataSource.__init__(self, file_paths, key, n_frames, frame_index_stride, frame_axis, insert_frame_axis, initial_frame_axis, additional_axes_iter, sort_files, overlapping_blocks, limit_n_samples, limit_n_frames, return_filename, cache, validate, **kwargs)
    240 _dataset = Dataset(file_paths, validate=validate, **kwargs)
    241 self.file_paths = _dataset.file_paths
--> [242](https://file+.vscode-resource.vscode-cdn.net/home/telse/Documents/Paper%20Review%20Reports/zea_review/~/Documents/Paper%20Review%20Reports/zea_review/.venv/lib/python3.13/site-packages/zea/data/dataloader.py:242) self.file_shapes = _dataset.load_file_shapes(key)
    243 _dataset.close()
    245 num_dims = len(self.file_shapes[0])

File ~/Documents/Paper Review Reports/zea_review/.venv/lib/python3.13/site-packages/zea/data/datasets.py:477, in Dataset.load_file_shapes(self, key)
    475 def load_file_shapes(self, key: str):
    476     """Load the shapes of the datasets in each file."""
--> [477](https://file+.vscode-resource.vscode-cdn.net/home/telse/Documents/Paper%20Review%20Reports/zea_review/~/Documents/Paper%20Review%20Reports/zea_review/.venv/lib/python3.13/site-packages/zea/data/datasets.py:477)     return _find_h5_file_shapes(self.file_paths, key, _file_hash(self.file_paths))

File ~/Documents/Paper Review Reports/zea_review/.venv/lib/python3.13/site-packages/zea/internal/cache.py:160, in cache_output.<locals>.decorator.<locals>.wrapper(*args, **kwargs)
    158     if verbose:
    159         log.info(f"Caching is globally disabled for {func.__qualname__}.")
--> [160](https://file+.vscode-resource.vscode-cdn.net/home/telse/Documents/Paper%20Review%20Reports/zea_review/~/Documents/Paper%20Review%20Reports/zea_review/.venv/lib/python3.13/site-packages/zea/internal/cache.py:160)     return func(*args, **kwargs)
    161 try:
    162     cache_key = generate_cache_key(func, args, kwargs, arg_names)

File ~/Documents/Paper Review Reports/zea_review/.venv/lib/python3.13/site-packages/zea/data/datasets.py:156, in _find_h5_file_shapes(filepaths, key, _filepath_hash, verbose)
    150 if os.environ.get("ZEA_FIND_H5_SHAPES_PARALLEL", "1") in ("1", "true", "yes"):
    151     # using multiprocessing to speed up reading hdf5 files
    152     # make sure to call find_h5_file_shapes from within a function
    153     # or use if __name__ == "__main__" to avoid freezing the main process
    155     with multiprocessing.Pool() as pool:
--> [156](https://file+.vscode-resource.vscode-cdn.net/home/telse/Documents/Paper%20Review%20Reports/zea_review/~/Documents/Paper%20Review%20Reports/zea_review/.venv/lib/python3.13/site-packages/zea/data/datasets.py:156)         file_shapes = list(
    157             tqdm.tqdm(
    158                 pool.imap(get_shape, filepaths),
    159                 total=len(filepaths),
    160                 desc="Getting file shapes in each h5 file",
    161                 disable=not verbose,
    162             )
    163         )
    164 else:
    165     file_shapes = []

File ~/Documents/Paper Review Reports/zea_review/.venv/lib/python3.13/site-packages/tqdm/std.py:[1181](https://file+.vscode-resource.vscode-cdn.net/home/telse/Documents/Paper%20Review%20Reports/zea_review/~/Documents/Paper%20Review%20Reports/zea_review/.venv/lib/python3.13/site-packages/tqdm/std.py:1181), in tqdm.__iter__(self)
   1178 time = self._time
   1180 try:
-> 1181     for obj in iterable:
   1182         yield obj
   1183         # Update and possibly print the progressbar.
   1184         # Note: does not call self.update(1) for speed optimisation.

File ~/.local/share/uv/python/cpython-3.13.2-linux-x86_64-gnu/lib/python3.13/multiprocessing/pool.py:[873](https://file+.vscode-resource.vscode-cdn.net/home/telse/Documents/Paper%20Review%20Reports/zea_review/~/.local/share/uv/python/cpython-3.13.2-linux-x86_64-gnu/lib/python3.13/multiprocessing/pool.py:873), in IMapIterator.next(self, timeout)
    871 if success:
    872     return value
--> 873 raise value
FileNotFoundError: [Errno 2] Unable to synchronously open file (unable to open file: name = '/tmp/zea_cache_q5nkpu3q/huggingface/datasets/datasets--zeahub--camus-sample/snapshots/617cf91a1267b5ffbcfafe9bebf0813c7cee8493/val/patient0401/patient0401_4CH_half_sequence.hdf5', errno = 2, error message = 'No such file or directory', flags = 0, o_flags = 0)
```

after running this section:

```python
dataset_path = "hf://zeahub/camus-sample/val"
dataloader = Dataloader(
    dataset_path,
    key="data/image_sc",
    batch_size=4,
    shuffle=True,
    clip_image_range=[-60, 0],
    image_range=[-60, 0],
    normalization_range=[0, 1],
    image_size=(256, 256),
    resize_type="resize",  # or "center_crop or "random_crop"
    seed=4,
)

for batch in dataloader:
    print("Batch shape:", batch.shape)
    break  # Just show the first batch

fig, _ = zea.visualize.plot_image_grid(batch)

```

I kept /tmp open while I ran the script, and it seems like the /tmp/zea_cache_q5nkpu3q is being deleted between at some point while running this section. Any ideas what this could be?
