### 2026-04-14 tue-bmd/zea#314 @tomelse
Example 2: plot_rf function
https://github.com/openjournals/joss-reviews/issues/9881
Note that the plots don't show up with the example scripts as listed here: [zea.readthedocs.io/en/latest/notebooks/data/zea_simulation_example.html](zea.readthedocs.io/en/latest/notebooks/data/zea_simulation_example.html).

I think plot_rf should be changed to have `plt.show()` at the end instead of `plt.close()`

```python
def plot_rf(rf_data, title="RF Data", cmap="gray"):
    """Plot the first transmit and first channel of the RF data."""
    plt.figure(figsize=(8, 4))
    plt.imshow(
        rf_data[0, :, :, 0].T,
        aspect="auto",
        cmap=cmap,
        extent=[0, rf_data.shape[1], 0, rf_data.shape[2]],
    )
    plt.xlabel("Sample (axial)")
    plt.ylabel("Element (lateral)")
    plt.title(title)
    plt.colorbar(label="Amplitude")
    plt.tight_layout()
    plt.savefig("simulation_plot_rf.png")
    plt.show()

```
