### 2026-04-14 tue-bmd/zea#314 @tomelse
Same goes for the fish example
