### 2026-04-14 tue-bmd/zea#315 @tomelse
CAMUS echonet examples don't work under JAX
https://github.com/openjournals/joss-reviews/issues/9881

The example here: https://zea.readthedocs.io/en/latest/notebooks/models/left_ventricle_segmentation_example.html doesn't seem to work under JAX. I tried installing the recommended version of tf2jax, but that didn't work. Instead, I installed tf2jax as usual without specifying a version, but I get a JAX error:

`AttributeError: module 'jax.core' has no attribute 'ClosedJaxpr'`
