### 2026-04-14 tue-bmd/zea#315 @tomelse
Here are further details of the error:

```
Cell In[3], [line 7](vscode-notebook-cell:?execution_count=3&line=7)
      3 # Load model
      4 model_echonet = EchoNetDynamic.from_preset("echonet-dynamic")
      5
      6 # Inference (expects RGB input)
----> [7](vscode-notebook-cell:?execution_count=3&line=7) masks_echonet = model_echonet(rgb_batch)
      8 masks_echonet = ops.squeeze(masks_echonet, axis=-1)
      9 masks_echonet = ops.convert_to_numpy(masks_echonet)
     10

File ~/Documents/Paper Review Reports/zea_review/.venv/lib/python3.13/site-packages/keras/src/utils/traceback_utils.py:122, in filter_traceback.<locals>.error_handler(*args, **kwargs)
    119     filtered_tb = _process_traceback_frames(e.__traceback__)
    120     # To get the full stack trace, call:
    121     # `keras.config.disable_traceback_filtering()`
--> [122](https://file+.vscode-resource.vscode-cdn.net/home/telse/Documents/Paper%20Review%20Reports/zea_review/~/Documents/Paper%20Review%20Reports/zea_review/.venv/lib/python3.13/site-packages/keras/src/utils/traceback_utils.py:122)     raise e.with_traceback(filtered_tb) from None
    123 finally:
    124     del filtered_tb

File ~/Documents/Paper Review Reports/zea_review/.venv/lib/python3.13/site-packages/zea/models/echonet.py:81, in EchoNetDynamic.build(self, input_shape)
     79 def build(self, input_shape):
     80     """Builds the network."""
---> [81](https://file+.vscode-resource.vscode-cdn.net/home/telse/Documents/Paper%20Review%20Reports/zea_review/~/Documents/Paper%20Review%20Reports/zea_review/.venv/lib/python3.13/site-packages/zea/models/echonet.py:81)     self.maybe_convert_to_jax(input_shape)

File ~/Documents/Paper Review Reports/zea_review/.venv/lib/python3.13/site-packages/zea/models/echonet.py:87, in EchoNetDynamic.maybe_convert_to_jax(self, input_shape)
     85 if backend.backend() == "jax":
     86     inputs = ops.zeros(input_shape)
---> [87](https://file+.vscode-resource.vscode-cdn.net/home/telse/Documents/Paper%20Review%20Reports/zea_review/~/Documents/Paper%20Review%20Reports/zea_review/.venv/lib/python3.13/site-packages/zea/models/echonet.py:87)     from zea.backend import tf2jax
     89     jax_func, jax_params = tf2jax.convert(tf.function(self.network), inputs)
     91     def call_fn(params, state, rng, inputs, training):

File ~/Documents/Paper Review Reports/zea_review/.venv/lib/python3.13/site-packages/zea/backend/tf2jax.py:4
      1 """Wrapper module to load tf2jax if available."""
      3 try:
----> [4](https://file+.vscode-resource.vscode-cdn.net/home/telse/Documents/Paper%20Review%20Reports/zea_review/~/Documents/Paper%20Review%20Reports/zea_review/.venv/lib/python3.13/site-packages/zea/backend/tf2jax.py:4)     from tf2jax import *  # noqa: F401, F403
      5 except ImportError as exc:
      6     raise ImportError(
      7         "tf2jax is not installed. "
      8         "Suggested installation: `pip install tf2jax==0.3.6 && pip install keras -U`"
      9         "Note that this may install a newer version of Keras than the one you have installed! "
     10     ) from exc

File ~/Documents/Paper Review Reports/zea_review/.venv/lib/python3.13/site-packages/tf2jax/__init__.py:21
     18 from tf2jax._src.config import override_config
     19 from tf2jax._src.config import update_config
---> [21](https://file+.vscode-resource.vscode-cdn.net/home/telse/Documents/Paper%20Review%20Reports/zea_review/~/Documents/Paper%20Review%20Reports/zea_review/.venv/lib/python3.13/site-packages/tf2jax/__init__.py:21) from tf2jax._src.tf2jax import AnnotatedFunction
     23 from tf2jax._src.tf2jax import convert
     24 from tf2jax._src.tf2jax import convert_from_restored

File ~/Documents/Paper Review Reports/zea_review/.venv/lib/python3.13/site-packages/tf2jax/_src/tf2jax.py:29
     27 import tensorflow as tf
     28 from tf2jax._src import config
---> [29](https://file+.vscode-resource.vscode-cdn.net/home/telse/Documents/Paper%20Review%20Reports/zea_review/~/Documents/Paper%20Review%20Reports/zea_review/.venv/lib/python3.13/site-packages/tf2jax/_src/tf2jax.py:29) from tf2jax._src import ops
     30 from tf2jax._src import utils
     31 import tree

File ~/Documents/Paper Review Reports/zea_review/.venv/lib/python3.13/site-packages/tf2jax/_src/ops.py:2179
   2173     return jax.lax.pad(operand, padding_value, padding_config)
   2175   return _func
   2178 def _maybe_get_jaxpreqn(
-> [2179](https://file+.vscode-resource.vscode-cdn.net/home/telse/Documents/Paper%20Review%20Reports/zea_review/~/Documents/Paper%20Review%20Reports/zea_review/.venv/lib/python3.13/site-packages/tf2jax/_src/ops.py:2179)     jaxpr: jax.core.ClosedJaxpr) -> Optional[jax.core.JaxprEqn]:
   2180   def is_all_vars(vs):
   2181     return all([isinstance(v, jax.core.Var) for v in vs])

File ~/Documents/Paper Review Reports/zea_review/.venv/lib/python3.13/site-packages/jax/_src/deprecations.py:57, in deprecation_getattr.<locals>.getattr(name)
     55   warnings.warn(message, DeprecationWarning, stacklevel=2)
     56   return fn
---> [57](https://file+.vscode-resource.vscode-cdn.net/home/telse/Documents/Paper%20Review%20Reports/zea_review/~/Documents/Paper%20Review%20Reports/zea_review/.venv/lib/python3.13/site-packages/jax/_src/deprecations.py:57) raise AttributeError(f"module {module!r} has no attribute {name!r}")

AttributeError: module 'jax.core' has no attribute 'ClosedJaxpr'
```
