### 2026-04-14 tue-bmd/zea#278 @tristan-deep
Hi @tomelse thanks for posting this. I can't reproduce it for now (with the snippet or notebook) but I'm investigating, since clearly there is some brittle logic causing this (maybe related to the caching).
