### 2026-04-14 #9881 @tomelse
I've just run through all the examples. A couple of small issues came up (see above and zea examples). Will finish off once the issues are addressed! Amazing work, though, truly an impressive feat!!
