### 2026-04-14 tue-bmd/zea#314 @tristan-deep
@tomelse, what we decided to do at some point is, instead of directly showing the fig with matplotlib, saving it and then subsequently showing it in the next markdown cell. The main reason was to not have these large images (especially when gif) embedded in the notebook, cluttering our git tree. Essentially every time you would update the notebook it would show a giant diff. Related issue https://github.com/tue-bmd/zea/issues/88

Does the image not show in the markdown cell after for you?
- For the RF data, see [this markdown cell](https://github.com/tue-bmd/zea/blob/b4ae6bb0b5799eafc303664bf7431fa1d6e1224a/docs/source/notebooks/data/zea_simulation_example.ipynb#L316)
- For the fish, see [this markdown cell](https://github.com/tue-bmd/zea/blob/b4ae6bb0b5799eafc303664bf7431fa1d6e1224a/docs/source/notebooks/data/zea_simulation_example.ipynb#L379).

I do understand this kind of "hides" the plotting in the next markdown cell, which could confuse users. But for easier maintaining of the notebooks we made that trade-off. Adding `plt.show()` would embed the image into the notebook itself. What do you think?
