### 2026-04-14 4f8b7e1
potential fix for _disable_cache
