### 2026-04-14 tue-bmd/zea#278 @tristan-deep
Just to check @tomelse, are you on linux and using python=3.13 (based on your logs).

Also did you turn on `os.environ["ZEA_DISABLE_CACHE"] = "1"`? Does it still happen without it?

I think the [`_disable_cache()`](https://github.com/tue-bmd/zea/blob/88fccdf24c7fd0a7e8d4ad3483347c4c8ad29653/zea/internal/cache.py#L39) func might be at fault. I've pushed a potential fix to #283, see [4f8b7e1](https://github.com/tue-bmd/zea/pull/283/commits/4f8b7e115ea1bcf68425a305bed6b7d5274e0a86). Let me know if that resolves it.
