### 2026-04-14 20f45bc
effort to replace deprecated load_data with h5py indexing
