### 2026-04-14 8beaf37
more typo and also add generation to pre-commit
