### 2026-04-15 e13edf3
revert bandwidth_percent default to 200%
