### 2026-04-15 4e34878
pass down compression arg
