### 2026-04-15 053a009
add test with ZEA_DISABLE_CACHE
