### 2026-04-15 7d11501
init_device respect CUDA_VISIBLE_DEVICES env variable
