+++ tests/test_device.py
@@ -6,23 +6,42 @@ import numpy as np
 import pytest
 
 from zea.backend import func_on_device
-from zea.internal.device import init_device
+from zea.internal.device import (
+    _cuda_visible_devices_disables_gpus,
+    get_gpu_memory,
+    init_device,
+)
 
-from . import backend_equality_check, DEFAULT_TEST_SEED
+from . import DEFAULT_TEST_SEED, backend_equality_check
+
+# NOTE: tests/__init__.py sets CUDA_VISIBLE_DEVICES="" so the full test suite runs
+# in CPU-only mode in CI (no GPU required).  Tests marked with @pytest.mark.gpu test
+# GPU-selection logic that falls back to CPU automatically in CI; they still contribute
+# to coverage but are most meaningful when run locally with a real GPU.
 
 devices = ["cpu", "gpu:0", "cuda:0", "auto:-1", "auto:1"]
 backends = ["tensorflow", "torch", "jax", "auto", "numpy"]
 
 
+@pytest.mark.gpu
 @pytest.mark.parametrize("device, backend", list(product(devices, backends)))
 def test_init_device(device, backend):
-    """Test device initialization with combinations of device and backend"""
+    """Test device initialization with combinations of device and backend.
+
+    In CI (CUDA_VISIBLE_DEVICES="") all GPU device strings fall back to CPU;
+    run locally with a GPU to exercise the full selection path.
+    """
     init_device(device=device, backend=backend, verbose=False)
 
 
+@pytest.mark.gpu
 @pytest.mark.parametrize("backend", backends)
 def test_default_init_device(backend):
-    """Test gpu usage setting script with defaults"""
+    """Test gpu usage setting script with defaults.
+
+    In CI (CUDA_VISIBLE_DEVICES="") this exercises the CPU fallback path;
+    run locally with a GPU to test real GPU selection.
+    """
     init_device(backend=backend, verbose=False)
 
 
@@ -38,3 +57,55 @@ def test_func_on_device():
         return x + y
 
     return func_on_device(f, "cpu", x, y)
+
+
+class TestCudaVisibleDevicesDisablesGpus:
+    """Tests for _cuda_visible_devices_disables_gpus helper."""
+
+    def test_unset(self, monkeypatch):
+        monkeypatch.delenv("CUDA_VISIBLE_DEVICES", raising=False)
+        assert _cuda_visible_devices_disables_gpus() is False
+
+    def test_empty_string(self, monkeypatch):
+        monkeypatch.setenv("CUDA_VISIBLE_DEVICES", "")
+        assert _cuda_visible_devices_disables_gpus() is True
+
+    def test_minus_one(self, monkeypatch):
+        monkeypatch.setenv("CUDA_VISIBLE_DEVICES", "-1")
+        assert _cuda_visible_devices_disables_gpus() is True
+
+    def test_negative_ids(self, monkeypatch):
+        monkeypatch.setenv("CUDA_VISIBLE_DEVICES", "-1,-2")
+        assert _cuda_visible_devices_disables_gpus() is True
+
+    def test_valid_id(self, monkeypatch):
+        monkeypatch.setenv("CUDA_VISIBLE_DEVICES", "0")
+        assert _cuda_visible_devices_disables_gpus() is False
+
+    def test_mixed_ids(self, monkeypatch):
+        monkeypatch.setenv("CUDA_VISIBLE_DEVICES", "0,-1")
+        assert _cuda_visible_devices_disables_gpus() is False
+
+    def test_whitespace(self, monkeypatch):
[…]
