+++ zea/internal/device.py
@@ -58,6 +58,29 @@ def print_gpu_memory_table(memory_free_values):
         print(f"{idx:<6}{mem:>7}")
 
 
+def _cuda_visible_devices_disables_gpus():
+    """Check if CUDA_VISIBLE_DEVICES is set to a value that disables all GPUs.
+
+    Returns ``True`` when the environment variable is set to an empty string
+    or contains only negative device IDs (the common convention being "-1").
+    Returns ``False`` when the variable is unset or contains at least one
+    non-negative device ID.
+    """
+    cuda_visible = os.environ.get("CUDA_VISIBLE_DEVICES")
+    if cuda_visible is None:
+        return False  # Not set – all GPUs visible
+    cuda_visible = cuda_visible.strip()
+    if cuda_visible == "":
+        return True  # Empty means no GPUs
+    try:
+        device_ids = [int(x.strip()) for x in cuda_visible.split(",")]
+        if all(d < 0 for d in device_ids):
+            return True
+    except ValueError:
+        pass  # Non-numeric values (e.g. GPU UUIDs) – let nvidia-smi decide
+    return False
+
+
 def get_gpu_memory(verbose=True):
     """Retrieve memory allocation information of all gpus.
 
@@ -68,6 +91,16 @@ def get_gpu_memory(verbose=True):
         memory_free_values: list of available memory for each gpu in MiB.
         Returns empty list if nvidia-smi is not available.
     """
+    # Respect CUDA_VISIBLE_DEVICES *before* calling nvidia-smi, which
+    # always reports all physical GPUs regardless of this variable.
+    if _cuda_visible_devices_disables_gpus():
+        if verbose:
+            log.info(
+                f"CUDA_VISIBLE_DEVICES={os.environ.get('CUDA_VISIBLE_DEVICES')!r} "
+                "disables all GPUs. Falling back to CPU."
+            )
+        return []
+
     if not check_nvidia_smi():
         log.warning(
             "nvidia-smi is not available. Please install nvidia-utils. "
@@ -107,7 +140,9 @@ def get_gpu_memory(verbose=True):
     # only show enabled devices
     if os.environ.get("CUDA_VISIBLE_DEVICES", "") != "":
         gpus = os.environ["CUDA_VISIBLE_DEVICES"]
-        gpus = [int(gpu) for gpu in gpus.split(",")][: len(memory_free_values)]
+        gpus = [int(gpu) for gpu in gpus.split(",")]
+        # Filter out negative and out-of-range IDs
+        gpus = [g for g in gpus if 0 <= g < len(memory_free_values)]
         if verbose:
             # Report the number of disabled GPUs out of the total
             num_gpus = len(memory_free_values)
