+++ pytest.ini
@@ -6,6 +6,7 @@ markers =
     performance: marks tests as performance (deselect with '-m "not performance"')
     heavy: marks tests as heavy (deselect with '-m "not heavy"')
     notebook: mark a test as a notebook test (deselect with '-m "not notebook"')
+    gpu: marks tests that exercise real GPU hardware (most meaningful offline; in CI CUDA_VISIBLE_DEVICES="" forces CPU)
 filterwarnings =
     ignore::DeprecationWarning:jupyter_client.*
     ignore::DeprecationWarning:papermill.*
