### 2026-04-15 ff5caff
migrate pytest.init to pyproject.toml
