### 2026-04-15 tue-bmd/zea#315 @tristan-deep
@tomelse, yes that is tricky. It boils down to only the `EchoNetDynamic` model in the end (CAMUS dataset loading, and the second model  `AugmentedCamusSeg` should be fine)

```python
model_echonet = EchoNetDynamic.from_preset("echonet-dynamic")
```

This model is originally a TF model, and that is why it works great with "tensorflow" backend. We did also find a way to make it work with JAX, but you need to install JAX, TensorFlow and tf2jax package together. As you can imagine, this is one big dependency nightmare. If you use our [docker image](https://hub.docker.com/r/zeahub/all), this is solved for you. However, if you do a custom installation, it might not be so clear how to manage the dependency versions to make it work.

Maybe the best is to swap the models order in the notebook, that way the example notebook will run for most people almost all the way through. Then we can make it clear what the options there would be if you would like to run `EchoNetDynamic` model?
