### 2026-04-15 fc4417a
more device tests
