### 2026-04-15 3d1c6e7
improve docs on EchoNetDynamic installation
