+++ zea/models/echonet.py
@@ -20,6 +20,14 @@ To try this model, simply load one of the available presets:
     A tutorial notebook where this model is used:
     :doc:`../notebooks/models/left_ventricle_segmentation_example`.
 
+.. note::
+    This model is only currently supported with the TensorFlow or JAX :ref:`backend-installation`
+    When using TensorFlow as backend, the model will work out of the box. When using JAX as backend,
+    the model is built using TensorFlow and then converted to JAX. This requires both
+    TensorFlow and JAX to be installed, which can be tricky regarding compatible CUDA versions.
+    One option is to run in our :ref:`docker-information` container, which has been tested
+    to work with both backends.
+
 """
 
 from pathlib import Path
