### 2026-04-15 tue-bmd/zea#315 @tristan-deep
I improved the notebook, and added more notes in various places (notebook + docs) to inform users about running the `EchoNetDynamic` with various backends. See #283.

EDIT: fixed a small issue with `EchoNetDynamic` running with JAX backend in #309. The notebook now runs with both TensorFlow and JAX backends using the `zeahub/all:latest` [docker image](https://zea.readthedocs.io/en/latest/installation.html#docker).
