### 2026-04-15 86f7565
improve overlay_masks test
