### 2026-04-15 8b21dcf
more pragma: no cover
