### 2026-04-15 1488ea6
fix torch metrics vmap
