### 2026-04-15 3d87341
AugmentedCamusSeg instead of EchoNetDynamic
