### 2026-04-15 5b847b0
Handle non-canonical CUDA_VISIBLE_DEVICES values without crashing
