### 2026-04-15 129531e
fix: apply CodeRabbit auto-fixes
Fixed 2 file(s) based on 3 unresolved review comments.

Co-authored-by: CodeRabbit <noreply@coderabbit.ai>
