### 2026-04-15 3e23083
fix convert_echonetlvh and make scan optional but required with raw_data provided
