### 2026-04-16 d0bdff2
echonet allow batch with jax backend
