### 2026-04-16 999c3bf
remove time step conditioning
