+++ zea/models/unet.py
@@ -120,7 +120,6 @@ class UNetTimeConditional(BaseModel):
         embedding_min_frequency=1.0,
         embedding_max_frequency=1000.0,
         embedding_dims=32,
-        use_frame_time_conditioning=False,
         name="unet_time_conditional",
         **kwargs,
     ):
@@ -132,7 +131,6 @@ class UNetTimeConditional(BaseModel):
         self.embedding_min_frequency = embedding_min_frequency
         self.embedding_max_frequency = embedding_max_frequency
         self.embedding_dims = embedding_dims
-        self.use_frame_time_conditioning = use_frame_time_conditioning
         self.network = get_time_conditional_unetwork(
             self.image_shape,
             self.widths,
@@ -399,7 +397,6 @@ class UNetTemporalTimeConditional(BaseModel):
         embedding_max_frequency=1000.0,
         embedding_dims=32,
         embedding_conditioning="add",  # | "concat"
-        use_frame_time_conditioning=False,
         name="unet_temporal_time_conditional",
         **kwargs,
     ):
@@ -421,7 +418,6 @@ class UNetTemporalTimeConditional(BaseModel):
         self.embedding_max_frequency = embedding_max_frequency
         self.embedding_dims = embedding_dims
         self.embedding_conditioning = embedding_conditioning
-        self.use_frame_time_conditioning = use_frame_time_conditioning
 
         self.network = get_temporal_time_conditional_unetwork(
             image_shape=self.image_shape,
@@ -436,7 +432,6 @@ class UNetTemporalTimeConditional(BaseModel):
             embedding_max_frequency=self.embedding_max_frequency,
             embedding_dims=self.embedding_dims,
             embedding_conditioning=self.embedding_conditioning,
-            use_frame_time_conditioning=self.use_frame_time_conditioning,
         )
 
     def get_config(self):
@@ -455,7 +450,6 @@ class UNetTemporalTimeConditional(BaseModel):
                 "embedding_min_frequency": self.embedding_min_frequency,
                 "embedding_max_frequency": self.embedding_max_frequency,
                 "embedding_dims": self.embedding_dims,
-                "use_frame_time_conditioning": self.use_frame_time_conditioning,
             }
         )
         return config
@@ -487,7 +481,6 @@ def get_temporal_time_conditional_unetwork(
     embedding_max_frequency=1000.0,
     embedding_dims=32,
     embedding_conditioning="add",  # | "concat"
-    use_frame_time_conditioning=False,
 ):
     """Build a temporal UNet with time-conditional sinusoidal embeddings.
 
@@ -542,10 +535,6 @@ def get_temporal_time_conditional_unetwork(
 
     noisy_images = keras.Input(shape=(image_height, image_width, n_channels))
     noise_variances = keras.Input(shape=(1, 1, 1))
-    if use_frame_time_conditioning:
-        frame_times = keras.Input(shape=(1, 1, n_channels))  # (B, T)
-    else:
-        frame_times = None
 
     # ---- Time embedding ----
     @keras.saving.register_keras_serializable()
@@ -553,12 +542,6 @@ def get_temporal_time_conditional_unetwork(
         return sinusoidal_embedding(
             x, embedding_min_frequency, embedding_max_frequency, embedding_dims
         )
-
-    # add frame-wise time embeddings to input noisy images
-    if frame_times is not None:
-        ft_emb = layers.Lambda(_sinusoidal_embedding, output_shape=(1, 1, n_channels))(frame_times)
-        ft_bias = layers.Dense(n_channels, name="frame_time_proj")(ft_emb)  # (B, T, n_frames)
[…]
