+++ zea/models/diffusion.py
@@ -66,10 +66,6 @@ class DiffusionModel(DeepGenerativeModel):
         max_t=1.0,
         noise_correlation_across_channels=None,
         noise_correlation_alpha=1.0,
-        use_frame_time_conditioning=False,
-        frame_time_embedding_min_frequency=1.0,
-        frame_time_embedding_max_frequency=None,
-        frame_time_embedding_dims=None,
         **kwargs,
     ):
         """Initialize a diffusion model.
@@ -109,12 +105,6 @@ class DiffusionModel(DeepGenerativeModel):
         self.ema_val = ema_val
         self.noise_correlation_across_channels = noise_correlation_across_channels
         self.noise_correlation_alpha = noise_correlation_alpha
-        self.use_frame_time_conditioning = use_frame_time_conditioning
-
-        if self.use_frame_time_conditioning:
-            # Ensure the underlying UNet accepts the extra frame_times input
-            self.network_kwargs = dict(self.network_kwargs)
-            self.network_kwargs.setdefault("use_frame_time_conditioning", True)
 
         # Frame-time conditioning is now handled inside the UNet when enabled.
         # We keep the flag to control dataset plumbing but do not instantiate
@@ -162,7 +152,6 @@ class DiffusionModel(DeepGenerativeModel):
                 "network_name": self.network_name,
                 "network_kwargs": self.network_kwargs,
                 "ema_val": self.ema_val,
-                "use_frame_time_conditioning": self.use_frame_time_conditioning,
                 "noise_correlation_across_channels": self.noise_correlation_across_channels,
                 "noise_correlation_alpha": self.noise_correlation_alpha,
             }
@@ -227,7 +216,7 @@ class DiffusionModel(DeepGenerativeModel):
 
         return network(inputs, training=training, **kwargs)
 
-    def sample(self, n_samples=1, n_steps=20, seed=None, frame_times=None, **kwargs):
+    def sample(self, n_samples=1, n_steps=20, seed=None, **kwargs):
         """Sample from the model.
 
         Args:
@@ -251,7 +240,6 @@ class DiffusionModel(DeepGenerativeModel):
             initial_noise=initial_noise,
             diffusion_steps=n_steps,
             seed=seed,
-            frame_times=frame_times,
             **kwargs,
         )
 
@@ -263,7 +251,6 @@ class DiffusionModel(DeepGenerativeModel):
         initial_step=0,
         initial_samples=None,
         seed=None,
-        frame_times=None,
         initial_noise=None,
         **kwargs,
     ):
@@ -308,8 +295,6 @@ class DiffusionModel(DeepGenerativeModel):
             initial_samples = ops.reshape(initial_samples, (-1, *self.input_shape))
         if "mask" in kwargs:
             kwargs["mask"] = _tile_with_sample_dim(kwargs["mask"])
-        if frame_times is not None:
-            frame_times = _tile_with_sample_dim(frame_times)
 
         seed1, seed2 = split_seed(seed, 2)
 
@@ -325,7 +310,6 @@ class DiffusionModel(DeepGenerativeModel):
             initial_samples=initial_samples,
             initial_step=initial_step,
             seed=seed2,
-            frame_times=frame_times,
             **kwargs,
         )  # ( batch_size * n_samples, *self.input_shape)
 
@@ -350,21 +334,10 @@ class DiffusionModel(DeepGenerativeModel):
         return [*self.noise_loss_tracker, *self.image_loss_tracker]
 
     def _split_data_and_frame_times(self, data):
-        """Accept either `images` or `(images, frame_times)` batches."""
+        """Accept either `images` or `(images, ...)` batches."""
         if isinstance(data, (tuple, list)):
-            images = data[0]
-            frame_times = data[1] if len(data) > 1 else None
[…]
