### 2026-04-16 0e07acc
support legacy unet layers
