+++ zea/models/unet.py
@@ -21,11 +21,19 @@ from zea import log
 
 # ---------------------------------------------------------------------------
 # Registry for network architecture builders (not full Model classes).
-# This maps string names → callable(image_shape, **kwargs) → keras.Model
+# This maps string names → callable(input_shape, **kwargs) → keras.Model
 # ---------------------------------------------------------------------------
 from zea.internal.registry import model_registry
 from zea.models.base import BaseModel
-from zea.models.layers import DownBlock, ResidualBlock, UpBlock, sinusoidal_embedding
+from zea.models.layers import (
+    DownBlock,
+    ResidualBlock,
+    UpBlock,
+    _legacy_down_block,
+    _legacy_residual_block,
+    _legacy_up_block,
+    sinusoidal_embedding,
+)
 from zea.models.preset_utils import register_presets
 from zea.models.presets import unet_presets
 
@@ -71,36 +79,45 @@ class UNet(BaseModel):
 
 
 def get_unetwork(
-    image_shape,
+    input_shape,
     widths,
     block_depth,
+    _legacy_layers=False,
 ):
     """Get a basic UNet architecture
 
     Args:
-        image_shape: tuple, (height, width, channels)
+        input_shape: tuple, (height, width, channels)
         widths: list, number of filters in each layer
         block_depth: int, number of residual blocks in each down/up block
+        _legacy_layers: bool, if True use legacy closure-based layers for
+            backward-compatible weight loading.
 
     Returns:
         keras.Model
     """
-    assert len(image_shape) == 3, "image_shape must be a tuple of (height, width, channels)"
+    assert len(input_shape) == 3, "input_shape must be a tuple of (height, width, channels)"
 
-    image_height, image_width, n_channels = image_shape
+    image_height, image_width, n_channels = input_shape
     noisy_images = keras.Input(shape=(image_height, image_width, n_channels))
 
     x = layers.Conv2D(widths[0], kernel_size=1)(noisy_images)
 
     skips = []
-    for width in widths[:-1]:
-        x = DownBlock(width, block_depth)([x, skips])
-
-    for _ in range(block_depth):
-        x = ResidualBlock(widths[-1])(x)
-
-    for width in reversed(widths[:-1]):
-        x = UpBlock(width, block_depth)([x, skips])
+    if _legacy_layers:
+        for width in widths[:-1]:
+            x = _legacy_down_block(width, block_depth)([x, skips])
+        for _ in range(block_depth):
+            x = _legacy_residual_block(widths[-1])(x)
+        for width in reversed(widths[:-1]):
+            x = _legacy_up_block(width, block_depth)([x, skips])
+    else:
+        for width in widths[:-1]:
+            x = DownBlock(width, block_depth)([x, skips])
+        for _ in range(block_depth):
+            x = ResidualBlock(widths[-1])(x)
+        for width in reversed(widths[:-1]):
+            x = UpBlock(width, block_depth)([x, skips])
 
     x = layers.Conv2D(n_channels, kernel_size=1, kernel_initializer="zeros")(x)
 
@@ -109,48 +126,99 @@ def get_unetwork(
 
 @model_registry(name="unet_time_conditional")
 class UNetTimeConditional(BaseModel):
-    """UNet model with time-conditional sinusoidal embedding"""
+    """UNet model with time-conditional sinusoidal embedding.
+
+    Optionally includes temporal self-attention at the bottleneck for video
+    diffusion, where frames are packed as channels
+    (``input_shape = (H, W, T*C)``).
+
+    Args:
+        input_shape: ``(H, W, C)`` or ``(H, W, T*C)`` for video.
+        widths: Filter counts per resolution level.
+        block_depth: Number of residual blocks per down/up stage.
[…]
