### 2026-04-16 bf292cf
update camus conversion script
