+++ tests/data/test_conversion_scripts.py
@@ -680,7 +680,6 @@ def verify_converted_camus_test_data(dst):
         # Load the hdf5 file and check for expected datasets
         for h5_file in h5_files:
             with File(h5_file, "r") as f:
-                assert "scan" in f, f"Missing 'scan' in {h5_file}"
                 f.validate()
 
 
