### 2026-04-16 53382f6
remove generate_zea_dataset as much as possible
