+++ tests/data/__init__.py
@@ -1,7 +1,6 @@
-import h5py
 import numpy as np
 
-from zea.data.data_format import DatasetElement, generate_zea_dataset
+from zea.data.file import File
 
 
 def generate_example_dataset(
@@ -32,90 +31,52 @@ def generate_example_dataset(
     """
 
     # creating some fake raw and image data
-    raw_data = np.ones((n_frames, n_tx, n_ax, n_el, n_ch))
+    raw_data = np.ones((n_frames, n_tx, n_ax, n_el, n_ch), dtype=np.float32)
     # image data is in dB
-    image = np.ones((n_frames, grid_size_z, grid_size_x)) * -40
+    image_sc = np.ones((n_frames, grid_size_z, grid_size_x), dtype=np.float32) * -40
 
     # creating some fake scan parameters
     t0_delays = np.zeros((n_tx, n_el), dtype=np.float32)
-    tx_apodizations = np.zeros((n_tx, n_el), dtype=np.float32)
     probe_geometry = np.zeros((n_el, 3), dtype=np.float32)
     probe_geometry[:, 0] = np.linspace(-0.02, 0.02, n_el)
     initial_times = np.zeros((n_tx,), dtype=np.float32)
 
+    scan = {
+        "probe_geometry": probe_geometry,
+        "sampling_frequency": np.float32(sampling_frequency),
+        "center_frequency": np.float32(center_frequency),
+        "demodulation_frequency": np.float32(center_frequency),
+        "initial_times": initial_times,
+        "t0_delays": t0_delays,
+        "sound_speed": np.float32(sound_speed),
+        "tx_apodizations": np.ones((n_tx, n_el), dtype=np.float32),
+        "focus_distances": np.full(n_tx, np.inf, dtype=np.float32),
+        "transmit_origins": np.zeros((n_tx, 3), dtype=np.float32),
+        "polar_angles": np.zeros((n_tx,), dtype=np.float32),
+        "azimuth_angles": np.zeros((n_tx,), dtype=np.float32),
+    }
+
     if add_optional_fields:
-        focus_distances = np.ones((n_tx,), dtype=np.float32) * np.inf
-        tx_apodizations = np.zeros((n_tx, n_el), dtype=np.float32)
-        polar_angles = np.zeros((n_tx,), dtype=np.float32)
-        azimuth_angles = np.zeros((n_tx,), dtype=np.float32)
-    else:
-        focus_distances = None
-        tx_apodizations = None
-        polar_angles = None
-        azimuth_angles = None
+        pass  # all required fields already provided above
+
+    data = {
+        "raw_data": raw_data,
+        "image_sc": image_sc,
+    }
 
     if add_optional_dtypes:
-        aligned_data = np.ones((n_frames, n_tx, n_ax, n_el, n_ch))
-        envelope_data = np.ones((n_frames, grid_size_z, grid_size_x))
-        beamformed_data = np.ones((n_frames, grid_size_z, grid_size_x, n_ch))
-        image_sc = np.ones_like(image)
-    else:
-        aligned_data = None
-        envelope_data = None
-        beamformed_data = None
-        image_sc = None
+        data["aligned_data"] = np.ones((n_frames, n_tx, n_ax, n_el, n_ch), dtype=np.float32)
+        data["envelope_data"] = np.ones((n_frames, grid_size_z, grid_size_x), dtype=np.float32)
+        data["beamformed_data"] = np.ones(
+            (n_frames, grid_size_z, grid_size_x, n_ch), dtype=np.float32
+        )
 
-    generate_zea_dataset(
+    f = File.create(
         path,
-        raw_data=raw_data,
-        aligned_data=aligned_data,
-        envelope_data=envelope_data,
[…]

[…]
