### 2026-04-16 cbe9308
update echonet lvh conversion script
