### 2026-04-16 e2b58cb
update cetus conversion script
