+++ zea/data/convert/verasonics.py
@@ -109,7 +109,9 @@ from keras import ops
 from schema import And, Optional, Or, Regex, Schema
 
 from zea import log
-from zea.data.data_format import DatasetElement, generate_zea_dataset
+from zea.data.data_format import DatasetElement
+from zea.data.file import File
+from zea.data.spec import DataSpec, ScanSpec
 from zea.func import log_compress, normalize
 from zea.internal.device import init_device
 from zea.utils import strtobool
@@ -1133,46 +1135,31 @@ class VerasonicsFile(h5py.File):
 
         additional_elements = []
 
-        # Add Verasonics lens correction to additional elements
-        el_lens_correction = DatasetElement(
-            group_name=group_name,
-            dataset_name="lens_correction",
-            data=self.lens_correction,
-            description=(
-                "The lens correction value used by Verasonics. This value is the "
-                "additional path length in wavelength that the lens introduces. "
-                "(This disregards refraction.)"
-            ),
-            unit="wavelengths",
-        )
-        additional_elements.append(el_lens_correction)
-
-        # Add Verasonics ImgDataP buffer to additional elements
-        try:
-            verasonics_image_buffer = self.read_image_data_p(event, frames=frames)
-            verasonics_image_buffer = DatasetElement(
-                dataset_name="verasonics_image_buffer",
-                data=verasonics_image_buffer,
-                description=(
-                    "The Verasonics ImgDataP buffer. "
-                    "WARNING: This buffer may skip frames compared to the raw data! "
-                    "Use only for reference."
-                ),
-                unit="unitless",
-            )
-            additional_elements.append(verasonics_image_buffer)
-        except Exception as e:
-            log.error(f"Could not read Verasonics ImgDataP buffer: {e}, skipping.")
-
         # Add additional elements from user-defined functions
         for additional_function in additional_functions:
             additional_elements.append(additional_function(self))
 
-        return {
+        result = {
             "raw_data": raw_data,
             **scan_dict,
+            "lens_correction": self.lens_correction,
             "additional_elements": additional_elements,
         }
+
+        # Add Verasonics ImgDataP buffer as image_sc if frame count matches
+        try:
+            image_sc = self.read_image_data_p(event, frames=frames)
+            if image_sc.shape[0] == raw_data.shape[0]:
+                result["image_sc"] = image_sc
+            else:
+                log.warning(
+                    f"Verasonics ImgDataP buffer has {image_sc.shape[0]} frames "
+                    f"but raw data has {raw_data.shape[0]} frames. Skipping image_sc."
+                )
+        except Exception as e:
+            log.error(f"Could not read Verasonics ImgDataP buffer: {e}, skipping.")
+
+        return result
 
     def _parse_frames_argument(self, frames, n_frames):
         value_error = ValueError(
@@ -1273,44 +1260,157 @@ class VerasonicsFile(h5py.File):
             num_events = num_events.pop()
 
             # loop over TX_Agent entries and overwrite TX each time
-            data = {}
+            all_event_data = {}
             for event in range(num_events):
-                data[event] = self.read_verasonics_file(
+                all_event_data[event] = self.read_verasonics_file(
                     event=event,
                     additional_functions=additional_functions,
                     allow_accumulate=allow_accumulate,
                 )
 
-            # convert dict of events to dict of lists
-            data = {key: [data[event][key] for event in data] for key in data[0]}
-            description = ["Verasonics data with multiple events"] * num_events
+            compression = "gzip" if enable_compression else None
 
-            # Generate the zea dataset
-            generate_zea_dataset(
-                path=output_path,
-                **data,
-                event_structure=True,
-                description=description,
-                enable_compression=enable_compression,
-            )
[…]
