### 2026-04-16 7229c9b
torch vmap fix
