+++ zea/metrics.py
@@ -396,11 +396,15 @@ class Metrics:
         metric_fn = fun
         for _ in range(num_batch_axes):
             # recursively vmap the leading axis
+            # disable_jit only when on the torch backend and the metric is not
+            # torch-vmappable (e.g. LPIPS), so that other backends (JAX, etc.)
+            # still benefit from their native vmap / vectorized_map paths.
+            disable_jit = not _use_torch_vmap and keras.backend.backend() == "torch"
             metric_fn = tensor.vmap(
                 metric_fn,
                 in_axes=0,
                 _use_torch_vmap=_use_torch_vmap,
-                disable_jit=not _use_torch_vmap,
+                disable_jit=disable_jit,
                 batch_size=mapped_batch_size,
             )
 
