### 2026-04-16 c324a0f
fix registry issues
