### 2026-04-17 d80fc4a
replace nan with minimun dynamic range
