### 2026-04-17 ff2fb1e
Include beamformed_data, envelope_data, image_sc as spatial maps
