+++ zea/data/convert/camus.py
@@ -172,10 +172,16 @@ def process_camus(source_path, output_path, overwrite=False):
         [0.0, n_x * 1e-4, 0.0, 1e-4, 0.0, n_z * 1e-4], dtype=np.float32
     )
 
+    # Build image_sc extent from sequence dimensions
+    n_x_sc, n_z_sc = image_seq.shape[1], image_seq.shape[2]
+    image_sc_extent = np.array(
+        [0.0, n_x_sc * 1e-4, 0.0, 1e-4, 0.0, n_z_sc * 1e-4], dtype=np.float32
+    )
+
     File.create(
         path=output_path,
         data={
-            "image_sc": image_seq,
+            "image_sc": {"pixels": image_seq, "extent": image_sc_extent},
             "image": {"pixels": image_seq_polar, "extent": extent},
         },
         probe_name="generic",
