+++ zea/data/file_operations.py
@@ -44,10 +44,10 @@ def save_file(
     probe: Probe,
     raw_data: np.ndarray = None,
     aligned_data: np.ndarray = None,
-    beamformed_data: np.ndarray = None,
-    envelope_data: np.ndarray = None,
-    image: np.ndarray = None,
-    image_sc: np.ndarray = None,
+    beamformed_data: dict = None,
+    envelope_data: dict = None,
+    image: dict = None,
+    image_sc: dict = None,
     description="",
     custom_maps: dict | None = None,
     metadata: dict | None = None,
@@ -61,6 +61,14 @@ def save_file(
         scan (Scan): The scan object containing the parameters of the acquisition.
         probe (Probe): The probe object containing the parameters of the probe.
         description (str): A description for the dataset.
+        beamformed_data (dict, optional): Beamformed data as a dict with ``"pixels"`` and
+            ``"extent"`` keys (validated as :class:`~zea.data.spec.BeamformedData`).
+        envelope_data (dict, optional): Envelope-detected data as a dict with ``"pixels"``
+            and ``"extent"`` keys (validated as :class:`~zea.data.spec.EnvelopeData`).
+        image_sc (dict, optional): Scan-converted image data as a dict with ``"pixels"``
+            and ``"extent"`` keys (validated as :class:`~zea.data.spec.ImageSc`).
+        image (dict, optional): Reconstructed (log-compressed) image data as a dict with
+            ``"pixels"`` and ``"extent"`` keys (validated as :class:`~zea.data.spec.Image`).
         custom_maps (dict, optional): Custom spatial map entries to include in the ``data`` group.
             Each key maps to a dict with ``"pixels"`` (np.ndarray, uint8) and ``"extent"``
             (np.ndarray, float32, shape ``(6,)``) fields, plus optional ``"labels"``,
@@ -91,12 +99,11 @@ def save_file(
         ("aligned_data", aligned_data),
         ("beamformed_data", beamformed_data),
         ("envelope_data", envelope_data),
+        ("image", image),
         ("image_sc", image_sc),
     ]:
         if arr is not None:
             data[key] = arr
-    if image is not None:
-        data["image_sc"] = image
 
     if custom_maps:
         for key, map_dict in custom_maps.items():
@@ -171,24 +178,37 @@ def sum_data(input_paths: list[Path], output_path: Path, overwrite=False):
 
         if data_dict["beamformed_data"] is not None:
             _assert_shapes_equal(
-                data_dict["beamformed_data"], new_data["beamformed_data"], "beamformed_data"
+                data_dict["beamformed_data"]["pixels"],
+                new_data["beamformed_data"]["pixels"],
+                "beamformed_data",
             )
-            data_dict["beamformed_data"] += new_data["beamformed_data"]
+            data_dict["beamformed_data"]["pixels"] += new_data["beamformed_data"]["pixels"]
 
         if data_dict["envelope_data"] is not None:
             _assert_shapes_equal(
-                data_dict["envelope_data"], new_data["envelope_data"], "envelope_data"
+                data_dict["envelope_data"]["pixels"],
+                new_data["envelope_data"]["pixels"],
+                "envelope_data",
             )
-            data_dict["envelope_data"] += new_data["envelope_data"]
+            data_dict["envelope_data"]["pixels"] += new_data["envelope_data"]["pixels"]
 
         if data_dict["image"] is not None:
-            _assert_shapes_equal(data_dict["image"], new_data["image"], "image")
-            data_dict["image"] = np.log(np.exp(new_data["image"]) + np.exp(data_dict["image"]))
+            _assert_shapes_equal(
[…]
