+++ zea/data/spec.py
@@ -590,6 +590,97 @@ class Image(UnsignedIntMap):
 
 
 @dataclass
+class BeamformedData(FloatMap):
+    """Beamformed (beamsummed) data and spatial extent metadata.
+
+    Args:
+        pixels: The beamformed data of shape (n_frames, x, z, n_ch) or
+            (n_frames, x, z, y, n_ch) and type float32.
+            n_ch is 1 for RF data or 2 for IQ data.
+        extent: Spatial extent in meters of shape (n_frames, 6) or (6,).
+            A shape of (6,) is broadcast to all frames. Values are ordered as
+            (xmin, xmax, ymin, ymax, zmax, zmin) and stored as float32.
+        labels: The labels for the channel dimension, e.g. ["RF"] or ["I", "Q"].
+            Auto-generated from n_ch if not provided.
+    """
+
+    SCHEMA = {
+        **FloatMap.SCHEMA,
+        "pixels": {
+            "dtype": np.float32,
+            "shape": (
+                ("n_frames", "x", "z", "y", "n_ch"),
+                ("n_frames", "x", "z", "n_ch"),
+            ),
+        },
+        "labels": {"dtype": np.str_, "shape": ("n_ch",)},
+    }
+
+    def __post_init__(self):
+        n_ch = self.pixels.shape[-1]
+        if n_ch not in (1, 2):
+            raise ValueError(
+                f"Beamformed data must have n_ch ∈ {{1, 2}} (RF or IQ), "
+                f"got n_ch={n_ch} (shape {self.pixels.shape})."
+            )
+        if self.labels is None:
+            self.labels = (
+                np.array(["RF"], dtype=np.str_)
+                if n_ch == 1
+                else np.array(["I", "Q"], dtype=np.str_)
+            )
+        super().__post_init__()
+
+
+@dataclass
+class EnvelopeData(FloatMap):
+    """Envelope-detected data and spatial extent metadata.
+
+    Args:
+        pixels: The envelope data of shape (n_frames, x, z) or
+            (n_frames, x, z, y) and type float32.
+        extent: Spatial extent in meters of shape (n_frames, 6) or (6,).
+            A shape of (6,) is broadcast to all frames. Values are ordered as
+            (xmin, xmax, ymin, ymax, zmax, zmin) and stored as float32.
+    """
+
+    SCHEMA = {
+        **FloatMap.SCHEMA,
+        "pixels": {
+            "dtype": np.float32,
+            "shape": (
+                ("n_frames", "x", "z", "y"),
+                ("n_frames", "x", "z"),
+            ),
+        },
+    }
+
+
+@dataclass
+class ImageSc(FloatMap):
+    """Scan-converted image data and spatial extent metadata.
+
+    Args:
+        pixels: The scan-converted image of shape (n_frames, x, z) or
+            (n_frames, x, z, y) and type float32.
+        extent: Spatial extent in meters of shape (n_frames, 6) or (6,).
+            A shape of (6,) is broadcast to all frames. Values are ordered as
+            (xmin, xmax, ymin, ymax, zmax, zmin) and stored as float32.
+    """
+
+    SCHEMA = {
+        **FloatMap.SCHEMA,
+        "pixels": {
+            "dtype": np.float32,
+            "shape": (
+                ("n_frames", "x", "z", "y"),
+                ("n_frames", "x", "z"),
+            ),
+        },
+    }
+
+
+@dataclass
 class SosMap(FloatMap):
     """Speed-of-sound map data and spatial extent metadata.
 
@@ -691,15 +782,11 @@ class DataSpec(Spec):
             and type float32 or int16.
         aligned_data: Time-of-flight corrected data of shape
             (n_frames, n_tx, n_ax, n_el, n_ch) and type float32 or int16.
[…]
