### 2026-04-17 4dc063e
add announcement banner OpenH-RF
