+++ zea/ops/ultrasound.py
@@ -277,12 +277,25 @@ class MachBeamform(Operation):
         else:
             self._interp_type_enum = interp_type
 
-        if keras.backend.backend() == "jax":
-            self._calculate_delays = backend_jit(calculate_delays, static_argnums=(7, 8, 14))
-            self._channels_to_complex = backend_jit(channels_to_complex)
-        else:
-            self._calculate_delays = backend_jit(calculate_delays)
-            self._channels_to_complex = channels_to_complex
+        # if keras.backend.backend() == "jax":
+        #     #self._calculate_delays = backend_jit(calculate_delays, static_argnums=(7, 8, 14))
+        #     #self._channels_to_complex = backend_jit(channels_to_complex)
+        #     # apply_lens_correction (arg 13) is a Python bool that must be concrete
+        #     # for the inner _calculate_delays static arg to work under JAX JIT.
+        #     # is_iq (arg 1 of _prepare_outputs) drives Python control flow.
+        self._prepare_inputs = backend_jit(
+            self._prepare_inputs, static_argnums=(13,)
+        )
+        self._prepare_outputs = backend_jit(
+            self._prepare_outputs, static_argnums=(1,)
+        )
+        # else:
+        #     # TODO: check if this is needed, passing the static_argnums is probably
+        #     # always ok (even if it doesn't do anything outside JAX)
+        #     #self._calculate_delays = backend_jit(calculate_delays)
+        #     #self._channels_to_complex = channels_to_complex
+        #     self._prepare_inputs = backend_jit(self._prepare_inputs)
+        #     self._prepare_outputs = backend_jit(self._prepare_outputs)
 
     # ------------------------------------------------------------------
     # Stage 1: Zea → mach  (pure tensor ops, JIT-able)
@@ -295,8 +308,7 @@ class MachBeamform(Operation):
         probe_geometry,
         sampling_frequency,
         sound_speed,
-        demodulation_frequency,
-        initial_times,
+        initial_times_for_tx,
         t0_delays,
         tx_apodizations,
         focus_distances,
@@ -307,71 +319,45 @@ class MachBeamform(Operation):
         apply_lens_correction,
         lens_thickness,
         lens_sound_speed,
-        rx_start_s,
     ):
-        """Translate Zea parameters and data into mach kernel inputs.
+        """Translate pre-validated Zea parameters and data into mach kernel inputs.
+
+        This method contains only pure tensor operations and is JIT-compiled
+        in ``__init__`` via ``backend_jit``.  All Python-level validation,
+        None-resolution, and logging must be performed by the caller before
+        invoking this method.
+
+        Args:
+            data: Input tensor with the channel dimension already removed and
+                IQ data already converted to complex dtype.  Shape is
+                ``([n_tx,] n_ax, n_el)`` (no batch) or
+                ``(n_frames, [n_tx,] n_ax, n_el)`` (batch).
+            flatgrid: ``(n_pix, 3)`` grid coordinates in metres.
+            probe_geometry: ``(n_el, 3)`` element positions in metres.
+            sampling_frequency: Sampling frequency in Hz.
+            sound_speed: Speed of sound in m/s.
+            initial_times_for_tx: Per-transmit time offsets ``(n_tx,)``
+                (already zeroed-out when ``rx_start_s`` was externally set).
+            t0_delays: Transmit delays ``(n_tx, n_el)`` in seconds.
+            tx_apodizations: Transmit apodizations ``(n_tx, n_el)``.
+            focus_distances: Focus distances ``(n_tx,)``.
+            polar_angles: Polar angles ``(n_tx,)``.
[…]
