### 2026-04-17 9ff50b4
cast pipeline inputs to float
