### 2026-04-17 8a4561f
fix grid docs
