+++ zea/ops/keras_ops.py
@@ -10,7 +10,7 @@ They can be used in zea pipelines like any other :class:`zea.Operation`, for exa
     >>> op = Squeeze(axis=1)
 
 This file is generated automatically. Do not edit manually.
-Generated with Keras 3.14.0
+Generated with Keras 3.12.0
 """
 
 import keras
@@ -207,16 +207,6 @@ class Array(Lambda):
             super().__init__(func=keras.ops.array, **kwargs)
         except AttributeError as e:
             raise MissingKerasOps("Array", "keras.ops.array") from e
-
-@ops_registry("keras.ops.array_split")
-class ArraySplit(Lambda):
-    """Operation wrapping keras.ops.array_split."""
-
-    def __init__(self, **kwargs):
-        try:
-            super().__init__(func=keras.ops.array_split, **kwargs)
-        except AttributeError as e:
-            raise MissingKerasOps("ArraySplit", "keras.ops.array_split") from e
 
 @ops_registry("keras.ops.average")
 class Average(Lambda):
@@ -538,16 +528,6 @@ class Deg2rad(Lambda):
         except AttributeError as e:
             raise MissingKerasOps("Deg2rad", "keras.ops.deg2rad") from e
 
-@ops_registry("keras.ops.depth_to_space")
-class DepthToSpace(Lambda):
-    """Operation wrapping keras.ops.depth_to_space."""
-
-    def __init__(self, **kwargs):
-        try:
-            super().__init__(func=keras.ops.depth_to_space, **kwargs)
-        except AttributeError as e:
-            raise MissingKerasOps("DepthToSpace", "keras.ops.depth_to_space") from e
-
 @ops_registry("keras.ops.det")
 class Det(Lambda):
     """Operation wrapping keras.ops.det."""
@@ -637,16 +617,6 @@ class Elu(Lambda):
             super().__init__(func=keras.ops.elu, **kwargs)
         except AttributeError as e:
             raise MissingKerasOps("Elu", "keras.ops.elu") from e
-
-@ops_registry("keras.ops.empty_like")
-class EmptyLike(Lambda):
-    """Operation wrapping keras.ops.empty_like."""
-
-    def __init__(self, **kwargs):
-        try:
-            super().__init__(func=keras.ops.empty_like, **kwargs)
-        except AttributeError as e:
-            raise MissingKerasOps("EmptyLike", "keras.ops.empty_like") from e
 
 @ops_registry("keras.ops.erf")
 class Erf(Lambda):
@@ -748,26 +718,6 @@ class Flip(Lambda):
         except AttributeError as e:
             raise MissingKerasOps("Flip", "keras.ops.flip") from e
 
-@ops_registry("keras.ops.fliplr")
-class Fliplr(Lambda):
-    """Operation wrapping keras.ops.fliplr."""
-
-    def __init__(self, **kwargs):
-        try:
-            super().__init__(func=keras.ops.fliplr, **kwargs)
-        except AttributeError as e:
-            raise MissingKerasOps("Fliplr", "keras.ops.fliplr") from e
-
-@ops_registry("keras.ops.flipud")
-class Flipud(Lambda):
-    """Operation wrapping keras.ops.flipud."""
-
-    def __init__(self, **kwargs):
-        try:
-            super().__init__(func=keras.ops.flipud, **kwargs)
-        except AttributeError as e:
-            raise MissingKerasOps("Flipud", "keras.ops.flipud") from e
-
 @ops_registry("keras.ops.floor")
 class Floor(Lambda):
     """Operation wrapping keras.ops.floor."""
@@ -777,16 +727,6 @@ class Floor(Lambda):
[…]
