+++ zea/data/convert/camus.py
@@ -168,9 +168,7 @@ def process_camus(source_path, output_path, overwrite=False):
     image_seq_polar = np.expand_dims(image_seq_polar, axis=-1)
 
     n_x, n_z = image_seq_polar.shape[1], image_seq_polar.shape[2]
-    extent = np.array(
-        [0.0, n_x * 1e-4, 0.0, 1e-4, 0.0, n_z * 1e-4], dtype=np.float32
-    )
+    extent = np.array([0.0, n_x * 1e-4, 0.0, 1e-4, 0.0, n_z * 1e-4], dtype=np.float32)
 
     File.create(
         path=output_path,
@@ -181,6 +179,7 @@ def process_camus(source_path, output_path, overwrite=False):
         probe_name="generic",
         description="camus dataset converted to zea format",
     )
+
 
 splits = {"train": [1, 401], "val": [401, 451], "test": [451, 501]}
 
