### 2026-04-18 170c0cc
add is_3d docstring
