### 2026-04-20 c3c56d5
rabbit fixes
