### 2026-04-20 73947bb
Prevent keras_ops.py downgrade when developer uses older Keras version
Agent-Logs-Url: https://github.com/tue-bmd/zea/sessions/9c493019-cf04-4ab5-990e-4f713adbb1a3

Co-authored-by: tristan-deep <58224357+tristan-deep@users.noreply.github.com>
