### 2026-04-20 1002684
Remove casting from pipeline and add to default pipeline
