### 2026-04-20 043f9d7
Fix tests (images should be uint8 for now)
