+++ tests/data/__init__.py
@@ -32,8 +32,8 @@ def generate_example_dataset(
 
     # creating some fake raw and image data
     raw_data = np.ones((n_frames, n_tx, n_ax, n_el, n_ch), dtype=np.float32)
-    # image data is in dB
-    image_sc_pixels = np.ones((n_frames, grid_size_z, grid_size_x), dtype=np.float32) * -40
+    # image data
+    pixels = np.ones((n_frames, grid_size_z, grid_size_x), dtype=np.uint8)
     # dummy extent in meters: (xmin, xmax, ymin, ymax, zmax, zmin)
     _extent = np.array([-0.02, 0.02, 0, 0, -0.03, 0], dtype=np.float32)
 
@@ -63,7 +63,8 @@ def generate_example_dataset(
 
     data = {
         "raw_data": raw_data,
-        "image_sc": {"pixels": image_sc_pixels, "extent": _extent},
+        "image": {"pixels": pixels, "extent": _extent},
+        "image_sc": {"pixels": pixels, "extent": _extent},
     }
 
     if add_optional_dtypes:
