+++ tests/data/test_dataset.py
@@ -20,7 +20,7 @@ _ALL_DATA_TYPES = _IMAGE_DATA_TYPES + _NON_IMAGE_DATA_TYPES
     [
         (
             0,
-            "all",
+            slice(None),
             (DUMMY_DATASET_N_FRAMES, DUMMY_DATASET_GRID_SIZE_Z, DUMMY_DATASET_GRID_SIZE_X),
         ),
         (
@@ -67,7 +67,7 @@ def test_dataset_indexing(file_idx, idx, expected_shape, dummy_dataset_path):
     dataset = Dataset.from_config(**config.data)
 
     file = dataset[file_idx]
-    data = file[file.format_key(config.data.dtype)][idx]
+    data = file[file.format_key(config.data.dtype)]["pixels"][idx]
 
     assert data.shape == expected_shape, (
         f"Data shape {data.shape} does not match expected shape {expected_shape}"
