### 2026-04-20 #9881 @tristan-deep
Thanks @tomelse and @rafaelorozco for the kind words and the time to take a look at our repository!

We have now resolved the issues related to the example notebook that @tomelse has brought up.

@adamltyson what will be the remaining tasks after the review checklist have been completed?
