### 2026-04-20 tue-bmd/zea#314 @tomelse
Ah okay I'm with you. Anyone who has used matplotlib before will know what to do. I think the only issue is if someone copies and pastes your code into their own notebook, rather than downloading your file, they will come across the same issue. Happy to close the issue.
