### 2026-04-20 #9881 @tomelse
One small final thing to quickly check. Some of the references are listed as "and others" rather than et al, is that the correct format for the journal? Otherwise, happy for the paper to be accepted!
