### 2026-04-20 16650ae
Refactor version guard into _check_version_and_generate; add monkeypatch test for older Keras exit
Agent-Logs-Url: https://github.com/tue-bmd/zea/sessions/74abf74c-f983-4f67-b6cb-4532c8d9f876

Co-authored-by: tristan-deep <58224357+tristan-deep@users.noreply.github.com>
