### 2026-04-20 #9881 @tristan-deep
> One small final thing to quickly check. Some of the references are listed as "and others" rather than et al, is that the correct format for the journal? Otherwise, happy for the paper to be accepted!

Thanks @tomelse, seems like we are using the correct citations style! Also confirmed by editor @adamltyson in https://github.com/tue-bmd/zea/issues/260#issuecomment-3927481759.

I also checked a few other [published JOSS articles](https://joss.theoj.org/papers/published), and they have the same formatted citation look.
