### 2026-04-21 #9881 @tristan-deep
> I don't think it's an issue. One of the very last things we do after review, and before publishing is to set the version to the latest one (and this is the one that gets archived).

@adamltyson, we just released a new version! [(v0.0.12)](https://github.com/tue-bmd/zea/releases/tag/v0.0.12)
