### 2026-04-21 2b9b1bb
remove unecessary if name main
