### 2026-04-21 4b5bbc7
Update image dtype
