### 2026-04-21 6c55c0f
change spec extent coords to match scan grid
