+++ zea/data/spec.py
@@ -476,7 +476,7 @@ class Map(Spec):
     The most flexible map spec, which can be used for any spatially aligned data product.
 
     Args:
-        pixels: The map pixels of shape (n_frames, x, z, y, n_ch) or (n_frames, x, z, y)
+        pixels: The map pixels of shape (n_frames, z, x, y, n_ch) or (n_frames, z, x, y)
             and type uint8 or float32 or int16.
         extent: The map extent in meters of shape (n_frames, 6) or (6,).
             A shape of (6,) is broadcast to all frames. Values are ordered as
@@ -499,9 +499,9 @@ class Map(Spec):
         "pixels": {
             "dtype": (np.uint8, np.float32, np.int16, np.complex64),
             "shape": (
-                ("n_frames", "x", "z", "y", "n_spatial_ch"),
-                ("n_frames", "x", "z", "y"),
-                ("n_frames", "x", "z"),
+                ("n_frames", "z", "x", "y", "n_spatial_ch"),
+                ("n_frames", "z", "x", "y"),
+                ("n_frames", "z", "x"),
             ),
         },
         "extent": {"dtype": np.float32, "shape": (("n_frames", 6), (6,))},
@@ -578,7 +578,7 @@ class Segmentation(BooleanMap):
     """Segmentation data and spatial extent metadata.
 
     Args:
-        pixels: The segmentation pixels of shape (n_frames, x, z, y, n_labels) and type bool.
+        pixels: The segmentation pixels of shape (n_frames, z, x, y, n_labels) and type bool.
         extent: The segmentation extent in meters of shape (n_frames, 6) or (6,).
             A shape of (6,) is broadcast to all frames. Values are ordered as
             (xmin, xmax, ymin, ymax, zmax, zmin) and stored as float32.
@@ -588,7 +588,7 @@ class Segmentation(BooleanMap):
 
     def __post_init__(self):
         assert self.pixels.ndim == 5, (
-            "Segmentation pixels must have 5 dimensions (n_frames, x, z, y, n_labels)"
+            "Segmentation pixels must have 5 dimensions (n_frames, z, x, y, n_labels)"
         )
         super().__post_init__()
 
@@ -598,7 +598,7 @@ class Image(UnsignedIntMap):
     """Reconstructed (log-compressed) image data and spatial extent metadata.
 
     Args:
-        pixels: The image pixels of shape (n_frames, x, z, y) and type uint8.
+        pixels: The image pixels of shape (n_frames, z, x, y) and type uint8.
         extent: The image extent in meters of shape (n_frames, 6) or (6,).
             A shape of (6,) is broadcast to all frames. Values are ordered as
             (radius_min, radius_max, theta_min, theta_max, phi_min, phi_max) and stored as float32.
@@ -610,8 +610,8 @@ class BeamformedData(FloatMap):
     """Beamformed (beamsummed) data and spatial extent metadata.
 
     Args:
-        pixels: The beamformed data of shape (n_frames, x, z, n_ch) or
-            (n_frames, x, z, y, n_ch) and type float32.
+        pixels: The beamformed data of shape (n_frames, z, x, n_ch) or
+            (n_frames, z, x, y, n_ch) and type float32.
             n_ch is 1 for RF data or 2 for IQ data.
         extent: Spatial extent in meters of shape (n_frames, 6) or (6,).
             A shape of (6,) is broadcast to all frames. Values are ordered as
@@ -625,8 +625,8 @@ class BeamformedData(FloatMap):
         "pixels": {
             "dtype": np.float32,
             "shape": (
-                ("n_frames", "x", "z", "y", "n_ch"),
-                ("n_frames", "x", "z", "n_ch"),
[…]
