### 2026-04-21 dffbea3
autogen'd docs
