+++ docs/source/_spec_ref.rst
@@ -191,7 +191,7 @@ Fields marked :bdg-secondary:`optional` may be absent; all others are
 
       .. dropdown:: ``beamformed_data``
 
-         Beamformed (beamsummed) data. Pixels are ``float32`` in (n_frames, x, z, n_ch) or (n_frames, x, z, y, n_ch); ``labels`` names each channel (RF or I/Q).
+         Beamformed (beamsummed) data. Pixels are ``float32`` in (n_frames, z, x, n_ch) or (n_frames, z, x, y, n_ch); ``labels`` names each channel (RF or I/Q).
 
          .. list-table::
             :header-rows: 1
@@ -204,7 +204,7 @@ Fields marked :bdg-secondary:`optional` may be absent; all others are
               - 
             * - ``pixels``
               - ``float32``
-              - (n_frames, x, z, y, n_ch) or (n_frames, x, z, n_ch)
+              - (n_frames, z, x, y, n_ch) or (n_frames, z, x, n_ch)
               - –
               - |badge-req|
             * - ``extent``
@@ -230,7 +230,7 @@ Fields marked :bdg-secondary:`optional` may be absent; all others are
 
       .. dropdown:: ``envelope_data``
 
-         Envelope-detected data. Pixels are ``float32`` in (n_frames, x, z) or (n_frames, x, z, y).
+         Envelope-detected data. Pixels are ``float32`` in (n_frames, z, x) or (n_frames, z, x, y).
 
          .. list-table::
             :header-rows: 1
@@ -243,7 +243,7 @@ Fields marked :bdg-secondary:`optional` may be absent; all others are
               - 
             * - ``pixels``
               - ``float32``
-              - (n_frames, x, z, y) or (n_frames, x, z)
+              - (n_frames, z, x, y) or (n_frames, z, x)
               - –
               - |badge-req|
             * - ``extent``
@@ -269,7 +269,7 @@ Fields marked :bdg-secondary:`optional` may be absent; all others are
 
       .. dropdown:: ``image_sc``
 
-         Scan-converted image. Pixels are ``float32`` in (n_frames, x, z) or (n_frames, x, z, y).
+         Scan-converted image. Pixels are ``float32`` in (n_frames, z, x) or (n_frames, z, x, y).
 
          .. list-table::
             :header-rows: 1
@@ -282,7 +282,7 @@ Fields marked :bdg-secondary:`optional` may be absent; all others are
               - 
             * - ``pixels``
               - ``uint8``
-              - (n_frames, x, z, y, n_spatial_ch) or (n_frames, x, z, y) or (n_frames, x, z)
+              - (n_frames, z, x, y, n_spatial_ch) or (n_frames, z, x, y) or (n_frames, z, x)
               - –
               - |badge-req|
             * - ``extent``
@@ -308,7 +308,7 @@ Fields marked :bdg-secondary:`optional` may be absent; all others are
 
       .. dropdown:: ``image``
 
-         Reconstructed (log-compressed) image. Pixels are ``uint8`` in (n_frames, x, z) or (n_frames, x, z, y).
+         Reconstructed (log-compressed) image. Pixels are ``uint8`` in (n_frames, z, x) or (n_frames, z, x, y).
 
          .. list-table::
             :header-rows: 1
@@ -321,7 +321,7 @@ Fields marked :bdg-secondary:`optional` may be absent; all others are
               - 
             * - ``pixels``
               - ``uint8``
-              - (n_frames, x, z, y, n_spatial_ch) or (n_frames, x, z, y) or (n_frames, x, z)
+              - (n_frames, z, x, y, n_spatial_ch) or (n_frames, z, x, y) or (n_frames, z, x)
               - –
               - |badge-req|
             * - ``extent``
[…]
