### 2026-04-21 1eedf8d
zmin zmax fix
