### 2026-04-21 5dc863f
spec + scan extent update
