+++ zea/scan.py
@@ -413,18 +413,46 @@ class Scan(Parameters):
         return zlims
 
     @cache_with_dependencies("grid", "grid_type", "distance_to_apex")
-    def extent(self):
-        """The extent of the beamforming grid in the format (xmin, xmax, zmax, zmin).
-        Can be directly used with `plt.imshow(x, extent=scan.extent)` for visualization.
+    def extent(self, imshow=False):
+        """The extent of the beamforming grid in the format: (xmin, xmax, ymin, ymax, zmin, zmax).
+
+        An optional bool parameter "imshow" can be passed to return an extent
+        (xmin, xmax, zmax, zmin) compatible with `plt.imshow` for 2D grids.
+
+        Args:
+            imshow (bool): Whether to return extent in the format compatible with
+            `plt.imshow` for 2D grids, allowing direct use via
+            `plt.imshow(x, extent=scan.extent)`. In this case, the extent
+            is returned as (xmin, xmax, zmax, zmin).
+
+        Returns:
+            np.ndarray: The extent of the beamforming grid in the format
+            (xmin, xmax, ymin, ymax, zmin, zmax) or (xmin, xmax, zmax, zmin)
+            if imshow=True.
         """
         xlims = (self.grid[:, :, 0].min(), self.grid[:, :, 0].max())
+        ylims = (self.grid[:, :, 1].min(), self.grid[:, :, 1].max())
         zlims = (self.grid[:, :, 2].min(), self.grid[:, :, 2].max())
 
         # For polar grids, adjust zlims to account for distance to apex
         if self.grid_type == "polar":
             zlims = (zlims[0] + self.distance_to_apex, zlims[1])
 
-        return np.array([xlims[0], xlims[1], zlims[1], zlims[0]])
+        if imshow:
+            if ylims[0] != ylims[1]:
+                log.warning("Are you sure you want to use 2D imshow extent for a 3D grid?")
+            return np.array([xlims[0], xlims[1], zlims[1], zlims[0]])
+
+        return np.array(
+            [
+                xlims[0],
+                xlims[1],
+                ylims[0],
+                ylims[1],
+                zlims[0],
+                zlims[1],
+            ]
+        )
 
     @cache_with_dependencies("grid")
     def flatgrid(self):
