### 2026-04-21 22f88cb
update existing usage of extent to have imshow=True
