### 2026-04-22 03938de
Add min/max to Map
