### 2026-04-22 8fcf362
make map spec docs more complete
