### 2026-04-22 833e566
update scan to have two extents, one general one and one imshow-specific
