### 2026-04-22 0fc9641
change pixels to values
