+++ tests/data/test_spec.py
@@ -23,18 +23,18 @@ from zea.data.spec import (
 
 def test_segmentation_spec():
     # Correct usage
-    pixels = np.zeros((10, 256, 256, 1, 4), dtype=np.bool_)
+    values = np.zeros((10, 256, 256, 1, 4), dtype=np.bool_)
     labels = np.array(["background", "label1", "label2", "label3"], dtype=np.str_)
     extent = np.array([0.0, 1.0, 0.0, 1.0, -1.0, 0.0], dtype=np.float32)
-    segmentation = Segmentation(pixels=pixels, labels=labels, extent=extent)
-    assert segmentation.pixels.shape == (10, 256, 256, 1, 4)
+    segmentation = Segmentation(values=values, labels=labels, extent=extent)
+    assert segmentation.values.shape == (10, 256, 256, 1, 4)
     assert segmentation.labels.shape == (4,)
     assert segmentation.extent.shape == (6,)
 
     # Incorrect usage: labels shape mismatch
     with pytest.raises(ValueError):
         Segmentation(
-            pixels=pixels, labels=np.array(["background", "label1"], dtype=np.str_), extent=extent
+            values=values, labels=np.array(["background", "label1"], dtype=np.str_), extent=extent
         )
 
 
@@ -93,28 +93,28 @@ def _example_data(n_frames, n_tx, n_el, n_ax, n_ch):
     return {
         "raw_data": np.zeros((n_frames, n_tx, n_ax, n_el, n_ch), dtype=np.float32),
         "image": {
-            "pixels": np.zeros((n_frames, 16, 12, 1), dtype=np.uint8),
+            "values": np.zeros((n_frames, 16, 12, 1), dtype=np.uint8),
             "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
         },
         "segmentation": {
-            "pixels": np.zeros((n_frames, 16, 12, 1, 2), dtype=np.bool_),
+            "values": np.zeros((n_frames, 16, 12, 1, 2), dtype=np.bool_),
             "labels": np.array(["background", "tissue"], dtype=np.str_),
             "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
         },
         "sos_map": {
-            "pixels": np.full((n_frames, 16, 12, 1), 1540.0, dtype=np.float32),
+            "values": np.full((n_frames, 16, 12, 1), 1540.0, dtype=np.float32),
             "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
         },
         "strain": {
-            "pixels": np.zeros((n_frames, 16, 12, 1), dtype=np.float32),
+            "values": np.zeros((n_frames, 16, 12, 1), dtype=np.float32),
             "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
         },
         "swe": {
-            "pixels": np.zeros((n_frames, 16, 12, 1), dtype=np.float32),
+            "values": np.zeros((n_frames, 16, 12, 1), dtype=np.float32),
             "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
         },
         "tissue_doppler": {
-            "pixels": np.zeros((n_frames, 16, 12, 1), dtype=np.float32),
+            "values": np.zeros((n_frames, 16, 12, 1), dtype=np.float32),
             "extent": np.array([0.0, 0.05, 0.0, 0.04, -0.04, -0.01], dtype=np.float32),
         },
     }
[…]
