### 2026-04-22 936faf7
remove duplicate test that ran on import time
