### 2026-04-22 eb33fe0
remove old event structure from dataformat
