### 2026-04-22 363ba47
Allow uint8 and float32 images
