### 2026-04-22 5e83481
Spec coordinates & extent (#330)
Co-authored-by: Wessel van Nierop <40076995+wesselvannierop@users.noreply.github.com>
