+++ zea/data/spec.py
@@ -476,11 +476,11 @@ class Map(Spec):
     The most flexible map spec, which can be used for any spatially aligned data product.
 
     Args:
-        values: The map values of shape (n_frames, x, z, y, n_ch) or (n_frames, x, z, y)
-            and type uint8 or float32 or int16.
+        values: The map values of shape (n_frames, z, x, y, n_ch) or (n_frames, z, x, y)
+            or (n_frames, z, x, n_ch) or (n_frames, z, x) and type uint8 or float32 or int16.
         extent: The map extent in meters of shape (n_frames, 6) or (6,).
             A shape of (6,) is broadcast to all frames. Values are ordered as
-            (xmin, xmax, ymin, ymax, zmax, zmin) and stored as float32.
+            (xmin, xmax, ymin, ymax, zmin, zmax) and stored as float32.
         labels: The labels corresponding to the `n_ch` channels in the values.
             This is required when values have an n_ch dimension, and should be None otherwise.
             For IQ data, this would typically be ["I", "Q"].
@@ -503,9 +503,9 @@ class Map(Spec):
         "values": {
             "dtype": (np.uint8, np.float32, np.int16, np.complex64),
             "shape": (
-                ("n_frames", "x", "z", "y", "n_spatial_ch"),
-                ("n_frames", "x", "z", "y"),
-                ("n_frames", "x", "z"),
+                ("n_frames", "z", "x", "y", "n_spatial_ch"),
+                ("n_frames", "z", "x", "y"),
+                ("n_frames", "z", "x"),
             ),
         },
         "extent": {"dtype": np.float32, "shape": (("n_frames", 6), (6,))},
@@ -531,7 +531,7 @@ class Map(Spec):
             if np.any(self.extent[..., 2] > self.extent[..., 3]):
                 raise ValueError("Map extent ylims must have ymin <= ymax")
             if np.any(self.extent[..., 4] > self.extent[..., 5]):
-                raise ValueError("Map extent zlims must have zmax <= zmin")
+                raise ValueError("Map extent zlims must have zmin <= zmax")
 
             # Ultrasound specific warning: if extent values are unusually large, log a warning
             if np.any(self.extent >= 1.0) or np.any(self.extent <= -1.0):
@@ -590,17 +590,17 @@ class Segmentation(BooleanMap):
     """Segmentation data and spatial extent metadata.
 
     Args:
-        values: The segmentation values of shape (n_frames, x, z, y, n_labels) and type bool.
+        values: The segmentation values of shape (n_frames, z, x, y, n_labels) and type bool.
         extent: The segmentation extent in meters of shape (n_frames, 6) or (6,).
             A shape of (6,) is broadcast to all frames. Values are ordered as
-            (xmin, xmax, ymin, ymax, zmax, zmin) and stored as float32.
+            (xmin, xmax, ymin, ymax, zmin, zmax) and stored as float32.
         labels: The labels corresponding to the segmentation values, where each unique value
             in the values corresponds to a label in this list of shape (n_labels,) and type str.
     """
 
     def __post_init__(self):
         assert self.values.ndim == 5, (
-            "Segmentation values must have 5 dimensions (n_frames, x, z, y, n_labels)"
+            "Segmentation values must have 5 dimensions (n_frames, z, x, y, n_labels)"
         )
         super().__post_init__()
 
@@ -610,7 +610,7 @@ class Image(Map):
     """Reconstructed (log-compressed) image data and spatial extent metadata.
 
     Args:
-        values: The image values of shape (n_frames, x, z, y) or (n_frames, x, z)
[…]
