+++ zea/data/convert/echonetlvh/__init__.py
@@ -292,22 +292,21 @@ class LVHProcessor(H5Processor):
 
         angle = cone_params["opening_angle"] / 2  # angular field spans (-angle, +angle)
         polar_im_set = self.cart2pol_batched(sequence_processed, angle)
-        sequence_processed = translate(sequence_processed, self._process_range, self.range_from)
-        sequence_processed_f32 = jnp.asarray(sequence_processed, dtype=jnp.float32)
+        sequence_processed_db = jnp.asarray(self._translate(sequence_processed), dtype=jnp.float32)
         del sequence_processed
 
         polar_im_set = translate(polar_im_set, self._process_range, (0, 255))
         polar_im_set_uint8 = jnp.asarray(jnp.floor(polar_im_set + 0.5), dtype=jnp.uint8)
         del polar_im_set
 
-        if jnp.all(sequence_processed_f32 == 0):
+        if jnp.all(sequence_processed_db == 0):
             raise ValueError(f"Processed sequence is all zeros for file {avi_file}")
 
         if jnp.all(polar_im_set_uint8 == 0):
             raise ValueError(f"Polar sequence is all zeros for file {avi_file}")
 
         # Convert JAX arrays to numpy for File.create / spec validation
-        image_sc_np = np.asarray(sequence_processed_f32)
+        image_sc_np = np.asarray(sequence_processed_db)
         polar_np = np.asarray(polar_im_set_uint8)
 
         # Image spec requires (n_frames, x, z, y) — add y=1 dimension
