### 2026-04-22 2bf24b5
remove obsolete docstring about events and DataElement
