### 2026-04-23 c4c33c1
Add TF_CPP_MIN_LOG_LEVEL=3 to Dockerfile to suppress CUDA dump messages
Agent-Logs-Url: https://github.com/tue-bmd/zea/sessions/ded5ea00-ad6c-4bd0-ae5b-a884fef59c84

Co-authored-by: tristan-deep <58224357+tristan-deep@users.noreply.github.com>
