### 2026-04-23 8bb2c77
Replace TF_CPP_MIN_LOG_LEVEL with sitecustomize.py stderr filter for CUDA warnings
TF_CPP_MIN_LOG_LEVEL is read after ABSL initializes, but the XLA CUDA factory
registration messages are emitted during shared-library load at import time,
before ABSL's own initializer runs. No environment variable can suppress them.

sitecustomize.py is Python's designed mechanism for site-wide installation
customization: it runs before any user code, so the stderr wrapper is in place
when the first 'import tensorflow'/'import jax' triggers the CUDA .so load.

Agent-Logs-Url: https://github.com/tue-bmd/zea/sess
[…]
