### 2026-04-23 04b4f5d
Revert all suppression attempts: neither TF_CPP_MIN_LOG_LEVEL nor sitecustomize.py can intercept C-level fd 2 writes
Agent-Logs-Url: https://github.com/tue-bmd/zea/sessions/e9eabe17-6a49-40cf-b9f2-fddb3a3d82e0

Co-authored-by: tristan-deep <58224357+tristan-deep@users.noreply.github.com>
