### 2026-04-23 08cdb7b
update Dockerfile docstring to reflect real docker image names used on Docker Hub
Co-authored-by: Copilot <copilot@github.com>
