### 2026-04-23 1974c81
Change Dockerfile INSTALL_* defaults from cpu to false; add clearer usage comments
Agent-Logs-Url: https://github.com/tue-bmd/zea/sessions/fe8efef9-478f-4c87-8cb7-a81083fc2202

Co-authored-by: tristan-deep <58224357+tristan-deep@users.noreply.github.com>
