### 2026-04-23 3a9bda7
nuclear diffusion method
Co-authored-by: Copilot <copilot@github.com>
