+++ zea/internal/operators.py
@@ -196,3 +196,90 @@ class FourierBlurOperator(Operator):
 
     def __str__(self):
         return f"y = F^(-1)(M * F(x)) filter at {self.cutoff_freq}"
+
+
+@operator_registry(name="haze")
+class HazeOperator(Operator):
+    """Haze operator for ultrasound image degradation modeling.
+
+    The haze operator models the forward process of haze corruption in ultrasound imaging,
+    where a measurement is formed by blending clean tissue images with a haze component:
+
+    .. math::
+
+        \\mathbf{y} = (1 - \\alpha) \\mathbf{x} + \\alpha \\mathbf{h}
+
+    where:
+
+    - :math:`\\mathbf{x}` is the clean tissue/signal image
+    - :math:`\\mathbf{h}` is the haze/background component
+    - :math:`\\alpha \\in [0, 1]` is the haze level controlling the mixing ratio
+    - :math:`\\mathbf{y}` is the observed (hazy) measurement
+
+    This operator is particularly used in the Nuclear Diffusion framework for cardiac
+    ultrasound dehazing, where structured background artifacts obscure dynamic content.
+
+    Note:
+        The haze component must be provided as an additional argument to both
+        :meth:`forward` and :meth:`transpose` methods.
+
+    See Also:
+        - :class:`~zea.models.diffusion.NuclearDiffusion`: Uses this operator for posterior sampling
+        - :doc:`../notebooks/models/nuclear_dehazing_example`: Example notebook
+
+    Example:
+        .. code-block:: python
+
+            from zea.internal.operators import HazeOperator
+            import numpy as np
+
+            operator = HazeOperator()
+            tissue = np.random.randn(64, 64, 1)
+            haze = np.random.randn(64, 64, 1)
+
+            # Create hazy measurement
+            measurement = operator.forward(tissue, haze, haze_level=0.5)
+    """
+
+    def corrupt(self, data, haze, haze_level: float = 0.5):
+        """Apply haze corruption to clean data.
+
+        Args:
+            data: Clean tissue/signal images.
+            haze: Haze/background component.
+            haze_level: Mixing coefficient :math:`\\alpha \\in [0, 1]`. Higher values
+                mean more haze. Default is 0.5.
+
+        Returns:
+            Corrupted (hazy) measurement.
+        """
+        out = (1 - haze_level) * data + haze_level * haze
+        return out
+
+    def forward(self, data, haze, haze_level: float = 0.5):
+        """Forward operator: apply haze corruption.
+
+        Args:
+            data: Clean tissue/signal images.
+            haze: Haze/background component.
+            haze_level: Mixing coefficient :math:`\\alpha \\in [0, 1]`. Default is 0.5.
+
+        Returns:
+            Corrupted (hazy) measurement.
+        """
+        return self.corrupt(data, haze, haze_level)
+
+    def transpose(self, data, haze):
+        """Transpose operator (identity for haze).
+
+        Args:
+            data: Input data.
+            haze: Haze component (unused).
+
+        Returns:
+            Input data unchanged.
+        """
+        return data
+
+    def __str__(self):
+        return "y = (1-α)x + αh"
