### 2026-04-23 7c6c82f
notebook improvements and fix test
Co-authored-by: Copilot <copilot@github.com>
