### 2026-04-24 4ab67cf
fix docstrings
