+++ zea/models/diffusion.py
@@ -1087,7 +1087,7 @@ class DDS(DiffusionGuidance):
 
 @diffusion_guidance_registry(name="nuclear-dps")
 class NuclearDiffusion(DPS):
-    """Nuclear Diffusion posterior sampling guidance.
+    r"""Nuclear Diffusion posterior sampling guidance.
 
     A hybrid framework that combines diffusion posterior sampling (DPS) with low-rank
     temporal modeling for video restoration. This method replaces the sparsity assumption
@@ -1095,41 +1095,43 @@ class NuclearDiffusion(DPS):
     maintaining a nuclear norm penalty on the background component to encourage low-rank
     temporal structure.
 
+
+    .. seealso::
+
+        - :func:`~zea.func.dehaze_nuclear_diffusion`: The dehazing application of this method
+        - :doc:`../notebooks/models/nuclear_dehazing_example`: Example notebook demonstrating
+          the method on cardiac ultrasound dehazing
+        - :class:`DPS`: Base diffusion posterior sampling guidance
+
     **Mathematical Formulation:**
 
-    Given observations :math:`\\mathbf{Y} \\in \\mathbb{R}^{n \\times p}` (video frames),
-    Nuclear Diffusion jointly samples the signal :math:`\\mathbf{X}` and low-rank background
-    :math:`\\mathbf{L}` from the posterior:
+    Given observations :math:`\mathbf{Y} \in \mathbb{R}^{n \times p}` (video frames),
+    Nuclear Diffusion jointly samples the signal :math:`\mathbf{X}` and low-rank background
+    :math:`\mathbf{L}` from the posterior:
 
     .. math::
 
-        \\mathbf{X}, \\mathbf{L} \\sim p_\\theta(\\mathbf{X}, \\mathbf{L} \\mid \\mathbf{Y})
+        \mathbf{X}, \mathbf{L} \sim p_\theta(\mathbf{X}, \mathbf{L} \mid \mathbf{Y})
 
     The posterior is factorized as:
 
     .. math::
 
-        p(\\mathbf{Y}, \\mathbf{L}, \\mathbf{X}) = p(\\mathbf{Y} \\mid \\mathbf{L}, \\mathbf{X}) \\, p(\\mathbf{L}) \\, p_\\theta(\\mathbf{X})
+        p(\mathbf{Y}, \mathbf{L}, \mathbf{X}) = p(\mathbf{Y} \mid \mathbf{L}, \mathbf{X}) \, p(\mathbf{L}) \, p_\theta(\mathbf{X})
 
     where:
 
-    - :math:`p(\\mathbf{Y} \\mid \\mathbf{L}, \\mathbf{X}) = \\mathcal{N}(\\mathbf{Y}; \\mathbf{L}+\\mathbf{X}, \\mu^{-1} \\mathbf{I})`
+    - :math:`p(\mathbf{Y} \mid \mathbf{L}, \mathbf{X}) = \mathcal{N}(\mathbf{Y}; \mathbf{L}+\mathbf{X}, \mu^{-1} \mathbf{I})`
       is the likelihood (measurement model)
-    - :math:`p(\\mathbf{L}) \\propto \\exp(-\\gamma \\|\\mathbf{L}\\|_*)` enforces low-rank structure
-      via the nuclear norm :math:`\\|\\mathbf{L}\\|_* = \\sum_i \\sigma_i(\\mathbf{L})`
-    - :math:`p_\\theta(\\mathbf{X})` is a learned diffusion prior capturing complex signal structure
+    - :math:`p(\mathbf{L}) \propto \exp(-\gamma \|\mathbf{L}\|_*)` enforces low-rank structure
[…]
