### 2026-04-24 #9881 @rafaelorozco
Happy to accept!

On Fri, Apr 24, 2026 at 10:50 AM Adam Tyson ***@***.***>
wrote:

> *adamltyson* left a comment (openjournals/joss-reviews#9881)
> <https://github.com/openjournals/joss-reviews/issues/9881#issuecomment-4311867196>
>
> Hi @tristan-deep <https://github.com/tristan-deep>, this all looks good.
> If @tomelse <https://github.com/tomelse>, @benfrey
> <https://github.com/benfrey>, @rafaelorozco
> <https://github.com/rafaelorozco> could all confirm they are happy to
> accept this submission, we can get started with the final tasks before
> acceptance.
>
