### 2026-04-24 #9881 @adamltyson
Hi @tristan-deep, could you take a look at the references? The "SKIP DOIs" are ok if there really are no DOIs associated, but the missing DOI looks like it should be addressed.

Also I noticed a noticed one senteance that reads strangely. In the summary, it states "This approach requires a redesign of current common ultrasound imaging pipeline", this seems like it's missing an article, or maybe pluralising "pipeline"?

Once these two points are addressed, please take a look at the "Additional author tasks..." below.
