### 2026-04-24 #9881 @tristan-deep
@adamltyson!
- Regarding DOIs, indeed the skip DOIs really don't seem to have a DOI. The one missing DOI ([`federici2026informative`](https://github.com/tue-bmd/zea/blob/854e86cb5b149a467e940cf35e3bea07fa76f3da/paper/paper.bib#L425)) is a reference that only appears in our documentation (paper is yet to be published, but already accepted, hence no DOI). It does not appear in the JOSS paper.
- Nicely spotted, indeed a typo! Have fixed that now [here](https://github.com/tue-bmd/zea/commit/854e86cb5b149a467e940cf35e3bea07fa76f3da).
- We released [v0.0.12](https://github.com/tue-bmd/zea/releases/tag/v0.0.12) a few days ago. Is also archived on [Zenodo](https://zenodo.org/records/19676300). License on both archive and software are listed as _Apache License 2.0_. Also author lists are matching.
