### 2026-04-24 abeff1e
refactor nuclear posterior diffusion step
Co-authored-by: Copilot <copilot@github.com>
