### 2026-04-24 #9881 @adamltyson
@tristan-deep could you remove the reference from your bib file? I think that's why the reference checker is complaining.
