### 2026-04-24 #9881 @tristan-deep
Let me have a look, I could try to make two bib files, one for the paper and one for the [docs](https://zea.readthedocs.io/en/latest/about.html#papers).
