### 2026-04-24 #9881 @adamltyson
@tristan-deep I think it's fine. If it's a pain to change, I don't think we have any specific policy about making the reference checker happy, as long as all the references in the paper are as complete as they can be.
