### 2026-04-24 4b8054f
improved docstring and resolve small issues
Co-authored-by: Copilot <copilot@github.com>
