### 2026-04-24 7eae414
rename operator to LinearInterpOperator
Co-authored-by: Copilot <copilot@github.com>
