### 2026-04-24 bed5222
make NuclearDiffusion more general
Co-authored-by: Copilot <copilot@github.com>
