+++ zea/models/diffusion.py
@@ -1147,7 +1147,7 @@ class NuclearDiffusion(DPS):
     """  # noqa: E501
 
     @staticmethod
-    def nuclear_norm_penalty(haze_images):
+    def nuclear_norm_penalty(background_images):
         r"""Compute nuclear norm penalty for low-rank enforcement.
 
         The nuclear norm (sum of singular values) encourages low-rank structure in
@@ -1161,7 +1161,7 @@ class NuclearDiffusion(DPS):
         where :math:`\sigma_i` are the singular values and :math:`r` is the rank.
 
         Args:
-            haze_images: Background/haze images of shape
+            background_images: Background images of shape
                 ``(batch, frames, height, width, channels)``.
                 Each sequence is reshaped to a matrix of shape ``(frames, height x width x channels)``
                 before computing the nuclear norm.
@@ -1173,20 +1173,20 @@ class NuclearDiffusion(DPS):
             The input is reshaped from ``(batch, frames, H, W, C)`` to ``(batch, frames, HxWxC)``
             before computing the singular values.
         """  # noqa: E501
-        n_batch, n_frames, height, width, channels = ops.shape(haze_images)
-        haze_images_flattened = ops.reshape(
-            haze_images, (n_batch, n_frames, height * width * channels)
+        n_batch, n_frames, height, width, channels = ops.shape(background_images)
+        background_images_flattened = ops.reshape(
+            background_images, (n_batch, n_frames, height * width * channels)
         )
-        haze_nuclear_penalty = ops.norm(haze_images_flattened, axis=(1, 2), ord="nuc")
+        background_nuclear_penalty = ops.norm(background_images_flattened, axis=(1, 2), ord="nuc")
 
         # normalize nuclear penalty
-        haze_nuclear_penalty /= n_frames
+        background_nuclear_penalty /= n_frames
 
         # sum across batches
-        return ops.sum(haze_nuclear_penalty)
+        return ops.sum(background_nuclear_penalty)
 
     @staticmethod
-    def weighted_nuclear_norm_penalty(haze_images, weight_factor: float = 2.0):
+    def weighted_nuclear_norm_penalty(background_images, weight_factor: float = 2.0):
         r"""Compute weighted nuclear norm penalty with enhanced rank control.
 
         This variant penalizes larger singular values more heavily than the standard
@@ -1200,7 +1200,7 @@ class NuclearDiffusion(DPS):
         index :math:`i`, and :math:`\alpha` is the ``weight_factor``.
 
         Args:
-            haze_images: Background/haze images of shape ``(batch, frames, height, width, channels)``.
+            background_images: Background images of shape ``(batch, frames, height, width, channels)``.
             weight_factor: Scaling factor :math:`\alpha` controlling how much more to penalize
                 higher-indexed singular values. Default is 2.0.
 
@@ -1211,8 +1211,10 @@ class NuclearDiffusion(DPS):
             This is a drop-in replacement for :meth:`nuclear_norm_penalty` that provides
             better rank control by more aggressively penalizing large singular values.
         """  # noqa: E501
-        n_batch, n_frames, height, width, _ = ops.shape(haze_images)
-        haze_images_flattened = ops.reshape(haze_images, (n_batch, n_frames, height * width))
+        n_batch, n_frames, height, width, channels = ops.shape(background_images)
+        background_images_flattened = ops.reshape(
+            background_images, (n_batch, n_frames, height * width * channels)
+        )
 
         def weighted_svd_penalty(matrix):
             """Compute weighted SVD penalty for a matrix"""
@@ -1224,7 +1226,7 @@ class NuclearDiffusion(DPS):
[…]
