### 2026-04-24 dbad8e3
add assertion
Co-authored-by: Copilot <copilot@github.com>
