### 2026-04-24 3a5a324
add more tests for split windows
Co-authored-by: Copilot <copilot@github.com>
