### 2026-04-24 ba7432e
add tests for guidance
Co-authored-by: Copilot <copilot@github.com>
