### 2026-04-24 c682ab4
fix guidance tests
Co-authored-by: Copilot <copilot@github.com>
