### 2026-04-24 01dafbc
Add HDF5 save, async saving, subsample transmits on read
Co-authored-by: Copilot <copilot@github.com>
