+++ zea/data/process.py
@@ -1,4 +1,5 @@
 import argparse
+from concurrent.futures import ThreadPoolExecutor
 from pathlib import Path
 
 import keras
@@ -6,6 +7,9 @@ import numpy as np
 from keras import ops
 
 import zea
+from zea.backend import jit
+from zea.data.data_format import generate_zea_dataset
+from zea.internal.core import reduce_to_signature
 from zea.utils import FunctionTimer
 
 
@@ -62,8 +66,13 @@ def get_parser() -> argparse.ArgumentParser:
     parser.add_argument(
         "--num_threads",
         type=int,
-        default=8,
+        default=16,
         help="Number of threads to use for dataloader. Default is 8.",
+    )
+    parser.add_argument(
+        "--overwrite",
+        action="store_true",
+        help="Whether to overwrite existing files in `save_dir` on conflict. Default is False.",
     )
     return parser
 
@@ -78,19 +87,14 @@ def run_processing(
     keep_keys=("maxval",),
     timings=False,
     num_threads=8,
+    overwrite=False,
 ) -> None:
-    def to_8bit(data, scan):
-        data = ops.convert_to_numpy(data)
-        data = np.nan_to_num(data, copy=False, nan=scan.dynamic_range[0])
-        data = zea.display.to_8bit(data, scan.dynamic_range)
+    @jit
+    def to_8bit(data, dynamic_range):
+        # data = ops.convert_to_numpy(data)
+        data = ops.nan_to_num(data, nan=dynamic_range[0])
+        data = zea.display.to_8bit(data, dynamic_range, to_numpy=False)
         return data
-
-    def save_video_frames(filestem: str, frames: list[np.ndarray], fps: float) -> None:
-        video = np.stack(frames, axis=0)
-        save_path = save_dir / f"{filestem}.{save_as}"
-        print("\n")
-        zea.io_lib.save_video(video, save_path, fps=fps)
-        print("\n")
 
     # Load config and pipeline
     config = zea.Config.from_path(config_path)
@@ -99,7 +103,25 @@ def run_processing(
     # Create save directory if it doesn't exist
     save_dir.mkdir(parents=True, exist_ok=True)
 
-    # Set up dataloader
+    # Peek at the first file to determine which transmits the scan will select.
+    # Passing these to the dataloader lets h5py read only those transmits from disk,
+    # avoiding a large CPU-side slice on every frame.
+    dataset_files = zea.Dataset(dataset_path, validate=False)
+    first_file_path = dataset_files.file_paths[0]
+    dataset_files.close()
+    with zea.File(first_file_path) as first_file:
+        first_scan = first_file.scan(**config.scan.as_dict())
+    selected_transmits = [int(i) for i in first_scan.selected_transmits]
+    # h5py requires strictly increasing indices. The scan also uses this list to index
+    # its own per-transmit parameters, so re-ordering here would desync frame vs params.
+    if any(b <= a for a, b in zip(selected_transmits, selected_transmits[1:])):
+        raise ValueError(
+            "selected_transmits must be strictly increasing to pre-filter on disk; "
+            f"got {selected_transmits}"
+        )
+
+    # Set up dataloader. axis_selections pre-filters the transmits axis (axis 1 of
+    # raw_data: frames, transmits, ...) at read time.
     dataloader = zea.Dataloader(
         dataset_path,
         key=data_type,
@@ -111,6 +133,7 @@ def run_processing(
         num_threads=num_threads,
         insert_frame_axis=False,
         sort_files=True,
+        axis_selections={1: selected_transmits},
     )
     iterator = iter(dataloader)
     total_batches = len(dataloader)
[…]
