### 2026-04-24 cd7a16c
Addition of REFoCUS Pipeline operation
Co-authored-by: Copilot <copilot@github.com>
