+++ zea/ops/pipeline.py
@@ -9,7 +9,7 @@ from keras import ops
 from zea import log
 from zea.backend import jit
 from zea.config import Config
-from zea.func.tensor import vmap
+from zea.func.tensor import vmap, hinv_adjoint, hinv_tikhonov, hinv_rsvd, hinv_tsvd
 from zea.func.ultrasound import channels_to_complex, complex_to_channels
 from zea.internal.core import DataTypes, ZEADecoderJSON, ZEAEncoderJSON, dict_to_tensor
 from zea.internal.core import Object as ZEAObject
@@ -1399,6 +1399,195 @@ class GeneralizedCoherenceFactor(Operation):
         """
         data = kwargs[self.key]
         return {self.output_key: self.process_image(data, m_zero=m_zero)}
+
+
+@ops_registry("refocus")
+class Refocus(Operation):
+    r"""REFoCUS (Retrospective Encoding For Conventional Ultrasound Sequences).
+
+    Decodes plane-wave or focused transmit data into synthetic aperture
+    (multistatic / full-matrix capture) data by inverting the transmit
+    encoding model in the frequency domain.
+
+    The transmit encoding is modelled as a matrix :math:`H` whose entry
+    :math:`H_{t,e}` describes the complex phase shift applied to element
+    :math:`e` during transmit event :math:`t`:
+
+    .. math::
+
+        H_{t,e}(f) = a_{t,e} \exp(-j 2\pi f \tau_{t,e})
+
+    where :math:`\tau_{t,e}` is the transmit delay in samples and
+    :math:`a_{t,e}` is the apodization.
+
+    At each temporal frequency the received RF spectrum is decoded by
+    multiplying with the pseudo-inverse :math:`H^{-1}`:
+
+    .. math::
+
+        \hat{S}(f) = H^{-1}(f) \, S(f)
+
+    producing a synthetic aperture dataset where each decoded channel
+    corresponds to a virtual single-element transmission.
+
+    The **input** data has shape ``(n_tx, n_ax, n_el, n_ch)`` and the
+    **output** has shape ``(n_el, n_ax, n_el, n_ch)``, where the new first
+    axis indexes the decoded virtual transmit elements.
+
+    .. admonition:: References
+
+        Bottenus, N. (2018).
+        "Recovery of the complete data set from focused transmit beams."
+        *IEEE Transactions on Ultrasonics, Ferroelectrics, and Frequency
+        Control*, 65(1), 30–38.
+
+        Ali, R., Dahl, J., & Bottenus, N. (2019).
+        "Extending Retrospective Encoding for Robust Recovery of the Multistatic Dataset."
+        *IEEE Transactions on Ultrasonics, Ferroelectrics, and Frequency
+        Control*, 67(5), 943–956.
+
+        https://github.com/nbottenus/REFoCUS
+
+    Args:
+        method (str): Inversion method. One of:
+
+            - ``'adjoint'``: Adjoint (matched-filter) pseudo-inverse with
+              an optional ramp filter in frequency. Fast and parameter-free.
+              Default.
+            - ``'tikhonov'``: Tikhonov-regularized inverse.
+            - ``'rsvd'``: Regularized SVD-based inverse.
+            - ``'tsvd'``: Truncated SVD-based inverse.
+
+        param (float or None): Regularization / filter parameter.
+
+            - ``'adjoint'``: ``None`` applies a ramp filter (multiply by
+              :math:`f`). Set to ``0`` to disable the ramp filter.
+            - ``'tikhonov'``, ``'rsvd'``, ``'tsvd'``: Relative regularization
[…]
