### 2026-04-24 3183f7e
docstring improvements
