### 2026-04-24 00da537
seed issue fix (per window)
Co-authored-by: Copilot <copilot@github.com>
