### 2026-04-24 702d253
remove jitting such that codecov catches executed lines
Co-authored-by: Copilot <copilot@github.com>
