### 2026-04-24 521b473
Fix ambiguous initial_samples docstrings in diffusion.py
Agent-Logs-Url: https://github.com/tue-bmd/zea/sessions/c721de10-d5c2-4c94-8c73-aa3b9b5b01e5

Co-authored-by: tristan-deep <58224357+tristan-deep@users.noreply.github.com>
