### 2026-04-28 835527b
exception -> warning and do not do `import zea` internally
