+++ zea/func/tensor.py
@@ -1732,12 +1732,12 @@ def h_matrix(delays, f_vec, apod):
     """Compute the H model matrix for a vector of normalised frequencies.
 
     Args:
-        delays: ``(n_tx, n_el)`` transmit delays in samples.
-        f_vec: ``(n_freq,)`` normalised frequencies.
-        apod: ``(n_tx, n_el)`` apodization weights.
+        delays: (n_tx, n_el) transmit delays in samples.
+        f_vec: (n_freq,) normalised frequencies.
+        apod: (n_tx, n_el) apodization weights.
 
     Returns:
-        H: ``(n_freq, n_tx, n_el)`` complex model matrix.
+        H: (n_freq, n_tx, n_el) complex model matrix.
     """
     f_c = ops.cast(f_vec[:, None, None], "complex64")  # (n_freq, 1, 1)
     d_c = ops.cast(delays[None], "complex64")  # (1, n_tx, n_el)
@@ -1749,39 +1749,36 @@ def hinv_adjoint(delays, f_vec, apod, param=None):
     """Adjoint (matched-filter) pseudo-inverse with optional ramp filter.
 
     Args:
-        delays: ``(n_tx, n_el)`` delays in samples.
-        f_vec: ``(n_freq,)`` normalised frequencies.
-        apod: ``(n_tx, n_el)`` apodization.
-        param: Ramp scale (``None`` → multiply by *f*; ``0`` → no ramp).
+        delays: (n_tx, n_el) delays in samples.
+        f_vec: (n_freq,) normalised frequencies.
+        apod: (n_tx, n_el) apodization.
+        param: Ramp control.
+            None (default) applies a ramp filter (multiply by f).
+            0 disables the ramp filter (multiply by 1, i.e. plain
+            adjoint / matched filter).
 
     Returns:
-        Hinv: ``(n_freq, n_el, n_tx)`` inverse model matrices.
+        Hinv: (n_freq, n_el, n_tx) inverse model matrices.
     """
     H = h_matrix(delays, f_vec, apod)
     Hinv = ops.conj(ops.transpose(H, (0, 2, 1)))
-    ramp_vals = f_vec if param is None else ops.ones_like(f_vec) * param
+    ramp_vals = f_vec if param is None else ops.ones_like(f_vec)
     ramp = ops.cast(ramp_vals, "complex64")[:, None, None]  # (n_freq, 1, 1)
     return ramp * Hinv
 
 
 def hinv_tikhonov(delays, f_vec, apod, param=None):
-    """Tikhonov-regularized inverse via SVD.
-
-    Computes :math:`V \\cdot \\text{diag}(s_i / (s_i^2 + \\lambda^2)) \\cdot U^H`
-    where :math:`\\lambda = \\text{param} \\cdot s_\\max`.  This is the closed-form
-    analytical solution to the normal-equations system
-    :math:`(H^H H + \\lambda^2 I)^{-1} H^H`, but avoids squaring the condition
-    number by working directly from the SVD.
+    """Tikhonov-regularized inverse.
 
     Args:
-        delays: ``(n_tx, n_el)`` delays in samples.
-        f_vec: ``(n_freq,)`` normalised frequencies.
-        apod: ``(n_tx, n_el)`` apodization.
+        delays: (n_tx, n_el) delays in samples.
+        f_vec: (n_freq,) normalised frequencies.
+        apod: (n_tx, n_el) apodization.
         param: Regularization parameter (relative to max singular value).
-            Defaults to ``1e-2``.
+            Defaults to 1e-2.
 
     Returns:
-        Hinv: ``(n_freq, n_el, n_tx)`` inverse model matrices.
+        Hinv: (n_freq, n_el, n_tx) inverse model matrices.
[…]
