+++ zea/ops/pipeline.py
@@ -9,7 +9,7 @@ from keras import ops
 from zea import log
 from zea.backend import jit
 from zea.config import Config
-from zea.func.tensor import vmap, hinv_adjoint, hinv_tikhonov, hinv_rsvd, hinv_tsvd
+from zea.func.tensor import vmap
 from zea.func.ultrasound import channels_to_complex, complex_to_channels
 from zea.internal.core import DataTypes, ZEADecoderJSON, ZEAEncoderJSON, dict_to_tensor
 from zea.internal.core import Object as ZEAObject
@@ -1452,8 +1452,7 @@ class Refocus(Operation):
         method (str): Inversion method. One of:
 
             - ``'adjoint'``: Adjoint (matched-filter) pseudo-inverse with
-              an optional ramp filter in frequency. Fast and parameter-free.
-              Default.
+              an optional ramp filter in frequency. Default.
             - ``'tikhonov'``: Tikhonov-regularized inverse.
             - ``'rsvd'``: Regularized SVD-based inverse.
             - ``'tsvd'``: Truncated SVD-based inverse.
@@ -1474,6 +1473,7 @@ class Refocus(Operation):
     def __init__(self, method="adjoint", param=None, **kwargs):
         if method not in self._VALID_METHODS:
             raise ValueError(f"method must be one of {self._VALID_METHODS}, got '{method}'")
+        kwargs.setdefault("jittable", False)
         super().__init__(
             input_data_type=DataTypes.RAW_DATA,
             output_data_type=DataTypes.RAW_DATA,
@@ -1482,96 +1482,140 @@ class Refocus(Operation):
         self.method = method
         self.param = param
 
-    def _get_hinv(self, delays, f_vec, apod):
-        """Return inverse model matrices ``(n_freq, n_el, n_tx)`` for all frequencies."""
-        _method_map = {
-            "adjoint": hinv_adjoint,
-            "tikhonov": hinv_tikhonov,
-            "rsvd": hinv_rsvd,
-            "tsvd": hinv_tsvd,
-        }
-        return _method_map[self.method](delays, f_vec, apod, param=self.param)
+    def _hinv_numpy(self, delays, f, apod):
+        """Compute Hinv at a single normalised frequency *f* using numpy.
+
+        Closely follows the reference implementation by N. Bottenus
+        (https://github.com/nbottenus/REFoCUS).
+
+        Args:
+            delays: ``(n_tx, n_el)`` numpy array of delays in samples.
+            f: scalar normalised frequency (cycles/sample).
+            apod: ``(n_tx, n_el)`` numpy apodization array.
+
+        Returns:
+            Hinv: ``(n_el, n_tx)`` complex numpy matrix.
+        """
+        H = np.matrix(apod * np.exp(-1j * 2 * np.pi * f * delays))
+
+        if self.method == "adjoint":
+            # param=None  → ramp filter (multiply by f)
+            # param=0     → no ramp (multiply by 1, plain adjoint)
+            ramp = f if self.param is None else 1.0
+            return ramp * H.H
+
+        # SVD-based methods
+        U, s, VH = np.linalg.svd(H)
+        lam = self.param if self.param is not None else 1e-2
+
+        if self.method == "tikhonov" or self.method == "rsvd":
+            sinv = s / (s**2 + (lam * s[0]) ** 2)
+        elif self.method == "tsvd":
+            sinv = 1.0 / s
+            sinv[s < lam * s[0]] = 0.0
+
[…]
