### 2026-04-28 66f7164
Add shape attr and shuffle on pipeline basis
