### 2026-04-29 ff4e1c7
Refactor ProbeOrientation to ProbePose and offset description
