### 2026-04-29 17f04b5
Add option to store as float
