### 2026-04-29 9ccdfef
Bump the python-packages group with 2 updates
Updates the requirements on [protobuf](https://github.com/protocolbuffers/protobuf) and [ruff](https://github.com/astral-sh/ruff) to permit the latest version.

Updates `protobuf` to 7.34.1
- [Release notes](https://github.com/protocolbuffers/protobuf/releases)
- [Commits](https://github.com/protocolbuffers/protobuf/commits)

Updates `ruff` from 0.15.11 to 0.15.12
- [Release notes](https://github.com/astral-sh/ruff/releases)
- [Changelog](https://github.com/astral-sh/ruff/blob/main/CHANGELOG.md)
- [Commits](https://github.com/astral-sh/ruff/compare
[…]
