### 2026-04-30 fa40cf3
Split into find_transmits and set_transmits
