### 2026-04-30 6910cff
Move check for aliasing
