### 2026-04-30 2cbb38b
Move len(idx) == 0 check to set_transmits
