### 2026-04-30 983e8ee
move refocus from np to keras ops
Co-authored-by: Copilot <copilot@github.com>
