+++ zea/ops/pipeline.py
@@ -1452,7 +1452,8 @@ class Refocus(Operation):
         method (str): Inversion method. One of:
 
             - ``'adjoint'``: Adjoint (matched-filter) pseudo-inverse with
-              an optional ramp filter in frequency. Default.
+              an optional ramp filter in frequency. Fast and parameter-free.
+              Default.
             - ``'tikhonov'``: Tikhonov-regularized inverse.
             - ``'rsvd'``: Regularized SVD-based inverse.
             - ``'tsvd'``: Truncated SVD-based inverse.
@@ -1473,7 +1474,6 @@ class Refocus(Operation):
     def __init__(self, method="adjoint", param=None, **kwargs):
         if method not in self._VALID_METHODS:
             raise ValueError(f"method must be one of {self._VALID_METHODS}, got '{method}'")
-        kwargs.setdefault("jittable", False)
         super().__init__(
             input_data_type=DataTypes.RAW_DATA,
             output_data_type=DataTypes.RAW_DATA,
@@ -1482,96 +1482,104 @@ class Refocus(Operation):
         self.method = method
         self.param = param
 
-    def _hinv_numpy(self, delays, f, apod):
-        """Compute Hinv at a single normalised frequency *f* using numpy.
-
-        Closely follows the reference implementation by N. Bottenus
-        (https://github.com/nbottenus/REFoCUS).
+    def _get_hinv(self, delays, f_vec, apod):
+        """Compute batched Hinv for all normalised frequencies at once.
 
         Args:
-            delays: ``(n_tx, n_el)`` numpy array of delays in samples.
-            f: scalar normalised frequency (cycles/sample).
-            apod: ``(n_tx, n_el)`` numpy apodization array.
+            delays: ``(n_tx, n_el)`` delays in samples.
+            f_vec: ``(n_freq,)`` normalised frequencies (cycles/sample).
+            apod: ``(n_tx, n_el)`` apodization.
 
         Returns:
-            Hinv: ``(n_el, n_tx)`` complex numpy matrix.
+            Hinv: ``(n_freq, n_el, n_tx)`` complex64 tensor.
         """
-        H = np.matrix(apod * np.exp(-1j * 2 * np.pi * f * delays))
+        # H: (n_freq, n_tx, n_el)
+        f_c = ops.cast(f_vec[:, None, None], "complex64")
+        d_c = ops.cast(delays[None], "complex64")
+        a_c = ops.cast(apod[None], "complex64")
+        H = a_c * ops.exp(ops.cast(-1j * 2 * np.pi, "complex64") * f_c * d_c)
 
         if self.method == "adjoint":
             # param=None  → ramp filter (multiply by f)
             # param=0     → no ramp (multiply by 1, plain adjoint)
-            ramp = f if self.param is None else 1.0
-            return ramp * H.H
+            Hinv = ops.conj(ops.transpose(H, (0, 2, 1)))
+            ramp_vals = f_vec if self.param is None else ops.ones_like(f_vec)
+            ramp = ops.cast(ramp_vals, "complex64")[:, None, None]
+            return ramp * Hinv
 
         # SVD-based methods
-        U, s, VH = np.linalg.svd(H)
+        U, s, VH = ops.svd(H, full_matrices=False)
         lam = self.param if self.param is not None else 1e-2
 
[…]
