### 2026-04-30 47600b9
add tests, finish example notebook
Co-authored-by: Copilot <copilot@github.com>
