+++ tests/test_refocus.py
@@ -1,0 +1,326 @@
+"""Tests for the Refocus operation (REFoCUS pipeline operation)."""
+
+import numpy as np
+import pytest
+
+from . import DEFAULT_TEST_SEED, backend_equality_check
+
+N_EL = 8  # number of transducer elements
+N_TX = 5  # number of transmit events
+N_AX = 64  # number of axial samples
+SAMPLING_FREQ = np.float32(40e6)  # Hz
+SOUND_SPEED = 1540.0  # m/s
+
+
+@pytest.fixture
+def probe_geometry():
+    """Linear array with N_EL elements spanning ±10 mm in x."""
+    xs = np.linspace(-10e-3, 10e-3, N_EL)
+    return np.stack([xs, np.zeros(N_EL), np.zeros(N_EL)], axis=-1).astype(np.float32)
+
+
+@pytest.fixture
+def plane_wave_delays(probe_geometry):
+    """Plane-wave transmit delays (n_tx, n_el) at a few steering angles."""
+    from zea.beamform.delays import compute_t0_delays_planewave
+
+    polar_angles = np.linspace(-0.2, 0.2, N_TX).astype(np.float32)
+    return compute_t0_delays_planewave(
+        probe_geometry, polar_angles, sound_speed=SOUND_SPEED
+    ).astype(np.float32)
+
+
+@pytest.fixture
+def rf_data():
+    """Random RF data: (n_tx, n_ax, n_el, 1)."""
+    rng = np.random.default_rng(DEFAULT_TEST_SEED)
+    return rng.standard_normal((N_TX, N_AX, N_EL, 1)).astype(np.float32)
+
+
+@pytest.fixture
+def iq_data():
+    """Random IQ data: (n_tx, n_ax, n_el, 2)."""
+    rng = np.random.default_rng(DEFAULT_TEST_SEED)
+    return rng.standard_normal((N_TX, N_AX, N_EL, 2)).astype(np.float32)
+
+
+def _call_refocus(op, data_np, probe_geometry_np, plane_wave_delays_np):
+    """Helper to call Refocus with standard numpy inputs (no batch dim)."""
+    import keras
+
+    return op(
+        data=keras.ops.convert_to_tensor(data_np),
+        t0_delays=keras.ops.convert_to_tensor(plane_wave_delays_np),
+        sampling_frequency=SAMPLING_FREQ,
+        probe_geometry=keras.ops.convert_to_tensor(probe_geometry_np),
+        initial_times=np.zeros(N_TX, dtype=np.float32),
+    )
+
+
+# ---------------------------------------------------------------------------
+# Construction / validation
+# ---------------------------------------------------------------------------
+
+
+def test_invalid_method_raises():
+    """Constructing Refocus with an unknown method must raise ValueError."""
+    from zea.ops import Refocus
+
+    with pytest.raises(ValueError, match="method must be one of"):
+        Refocus(method="unknown_method")
+
+
+def test_valid_methods_construct():
+    """All documented methods should construct without error."""
+    from zea.ops import Refocus
+
+    for method in ("adjoint", "tikhonov", "rsvd", "tsvd"):
+        op = Refocus(method=method)
+        assert op.method == method
+
+
+# ---------------------------------------------------------------------------
+# Output shape
+# ---------------------------------------------------------------------------
+
+
+@pytest.mark.parametrize("method", ["adjoint", "tikhonov", "rsvd", "tsvd"])
+def test_output_shape_rf(method, probe_geometry, plane_wave_delays, rf_data):
+    """Decoded RF output must have shape (n_el, n_ax, n_el, 1)."""
+    import keras
+    from zea.ops import Refocus
+
[…]
