### 2026-05-01 2ce3ba8
notebook and test fix
