### 2026-05-01 b04f349
Refactor ProbeOrientation to ProbePose and offset description (#340)
Co-authored-by: simon <s.w.penninga@tue.nl>
