+++ zea/data/spec.py
@@ -1218,19 +1218,35 @@ class Subject(Spec):
 
 @dataclass
 class Signal(Spec):
-    """Base class for additional signals with timing offset and sampling frequency metadata.
+    """Base class for additional signals with timing and sampling-frequency metadata.
 
     Args:
-        offset: Time offset in seconds relative to frame timing.
+        start_time_offset: Time offset in seconds between the first transmit event
+            of the ultrasound acquisition and sample 0 of this data. Negative
+            means this data starts before the first transmit event; positive
+            means it starts after.
         sampling_frequency: Sampling frequency in Hz for the additional signal samples.
     """
 
-    offset: np.ndarray | float
+    start_time_offset: np.ndarray | float
     sampling_frequency: np.ndarray | float
 
     SCHEMA = {
-        "offset": {"dtype": np.float32, "shape": ()},
+        "start_time_offset": {"dtype": np.float32, "shape": ()},
         "sampling_frequency": {"dtype": np.float32, "shape": ()},
+    }
+
+    FIELD_METADATA = {
+        "start_time_offset": {
+            "unit": "s",
+            "description": (
+                "Time offset between the first transmit event of the ultrasound "
+                "acquisition and sample 0 of this data. Negative means this data "
+                "starts before the first transmit event; positive means it starts "
+                "after."
+            ),
+        },
+        "sampling_frequency": {"unit": "Hz", "description": "Sampling frequency."},
     }
 
     def __post_init__(self):
@@ -1241,21 +1257,87 @@ class Signal(Spec):
 
 
 @dataclass
-class ProbeOrientation(Signal):
-    """Probe pose and timing metadata.
+class ProbePose(Signal):
+    """Sampled probe pose metadata at the tip of the transducer.
+
+    The pose uses the coordinate convention x = lateral along the transducer,
+    y = elevation (out of plane), and z = axial (depth).
 
     Args:
-        pose: Probe pose in meters of shape (T, 6), ordered as (x, y, z, az, el, roll).
-        offset: Time offset in seconds relative to frame timing.
-        sampling_frequency: Sampling frequency in Hz for probe orientation samples.
+        translation: Position of the transducer tip in meters of shape (T, 3),
+            ordered as (x, y, z).
+        rotation: Orientation of the transducer tip of shape (T, 3) or (T, 4),
+            interpreted according to ``rotation_representation``.
+        rotation_representation: Rotation parameterization. Supported values are
+            ``"euler_xyz"``, ``"quaternion_wxyz"``, and ``"quaternion_xyzw"``.
+        start_time_offset: Time offset in seconds between the first transmit event
+            of the ultrasound acquisition and sample 0 of this data.
+        sampling_frequency: Sampling frequency in Hz for probe pose samples.
     """
 
-    pose: np.ndarray
+    translation: np.ndarray
+    rotation: np.ndarray
+    rotation_representation: str
 
     SCHEMA = {
-        "pose": {"dtype": np.float32, "shape": ("T", 6)},
+        "translation": {"dtype": np.float32, "shape": ("T", 3)},
+        "rotation": {"dtype": np.float32, "shape": (("T", 3), ("T", 4))},
+        "rotation_representation": {"dtype": str, "shape": ()},
         **Signal.SCHEMA,
     }
+
+    FIELD_METADATA = {
+        "translation": {
+            "unit": "m",
+            "description": (
+                "Position of the transducer tip, ordered as (x, y, z), where x is "
+                "lateral along the transducer, y is elevation (out of plane), and "
+                "z is axial (depth)."
+            ),
+        },
+        "rotation": {
+            "unit": "-",
[…]
