### 2026-05-02 #9881 @arfon
@tristan-deep – could you please fix this typo:

- publically → publicly (Statement of need, first sentence)
