### 2026-05-02 #9881 @tristan-deep
> [@tristan-deep](https://github.com/tristan-deep) – could you please fix this typo:
>
> * publically → publicly (Statement of need, first sentence)

Done!
